### 2024-09-21 #7258 @editorialbot
[PRE REVIEW]: J-UBIK: The JAX-accelerated Universal Bayesian Imaging Kit
**Submitting author:** @wmarg (<a href="http://orcid.org/0000-0001-7218-8282">Margret Westerkamp</a>)
**Repository:** https://github.com/NIFTy-PPL/J-UBIK
**Branch with paper.md** (empty if default branch): main
**Version:** 0.2
