+++ demos/paper/visualize_mock_sky.py
@@ -28,13 +28,14 @@ if __name__ == "__main__":
     sky = sky_model.create_sky_model()
     sky_dict = sky_model.sky_model_to_dict()
 
-    sat_max = {'sky': [2e-6, 2e-7, 5e-8], 'diffuse': [2e-6, 2e-7, 5e-8], 'points': [5e-8, 5e-8, 5e-8]}
     for key, op in sky_dict.items():
         real_pos = op(pos)
         bbox_info = [(7, 4), 7, 24,  'black']
-        plot_rgb(real_pos, sat_min=[1e-9, 1e-9, 1e-9],
-                 sat_max=sat_max[key],
-                 sigma=None, log=True,
+        plot_rgb(real_pos,
+                 sat_min=[1e-10, 1e-10, 1e-10],
+                 sat_max=[5e-8, 5e-8, 5e-8],
+                 # log=True,
+                 sigma=None,
                  title='simulated sky', fs=18, pixel_measure=28,
                  output_file=join(output_dir, f'simulated_{key}_rgb.png'),
                  alpha=0.5,
