+++ demos/paper/visualize_mock_data.py
@@ -18,15 +18,40 @@ from plot_eROSITA_image import plot, plot_rgb
 if __name__ == "__main__":
     results_path = "results/Mock-17092024-001M"
     config_name = "eROSITA_demo.yaml"
+    path_to_caldb = '../data/'
     output_dir = ju.create_output_directory(join(results_path, 'paper'))
     config_path = join(results_path, config_name)
     config_dict = ju.get_config(config_path)
     grid_info = config_dict["grid"]
-    tm_ids = config_dict["telescope"]["tm_ids"]
+    file_info = config_dict["files"]
+    tel_info = config_dict["telescope"]
+    tm_ids = tel_info["tm_ids"]
     e_min = grid_info['energy_bin']['e_min']
     e_max = grid_info['energy_bin']['e_max']
     data_path = join(results_path, 'data.pkl')
+    processed_data_path = join(file_info['obs_path'], 'processed')
 
+    epix = grid_info['edim']
+    spix = grid_info['sdim']
+    pixel_area = (config_dict['telescope']['fov'] / config_dict['grid']['sdim']) **2 # density to flux
+
+    exposures = []
+    for it, tm_id in enumerate(tm_ids):
+        exposure_filenames = f'tm{tm_id}_' + file_info['exposure']
+        exposure_filenames = [join(processed_data_path,
+                                   f"{exposure_filenames.split('.')[0]}_emin{e}_emax{E}.fits")
+                              for e, E in zip(e_min, e_max)]
+        exposures.append([])
+        for e, output_filename in enumerate(exposure_filenames):
+            with fits.open(exposure_filenames[e]) as hdul:
+                exposures[it].append(hdul[0].data)
+    exposures = np.array(exposures, dtype=float)
+    exposures[exposures<=tel_info['exp_cut']] = 0 # FIXME FROM CONFIG Instroduce Exposure cut
+    correct_exposures_for_effective_area = True
+    if correct_exposures_for_effective_area:
+        # from src.library.response import calculate_erosita_effective_area
+        ea = ju.calculate_erosita_effective_area(path_to_caldb, tm_ids, e_min, e_max)
+        exposures *= ea[:, :, np.newaxis, np.newaxis]
 
     data = ju.load_masked_data_from_config(config_path)
     plottable_vector = jft.Vector({key: val.astype(float) for key, val
@@ -43,11 +68,12 @@ if __name__ == "__main__":
     # Plotting the data
     unmasked_data = mask_adj_func(plottable_vector)
     plotting_kwargs = {'vmin':1e0, 'vmax':4e2}
-    bbox_info = [(7, 4), 7,  20, 'black']
+    bbox_info = [(14, 8), 14,  40, 'black']
     for i in range(unmasked_data.shape[0]):
         plot(unmasked_data[i],
+             pixel_factor=2,
              pixel_measure=28,
-             fs=8,
+             fs=12,
                         title=['0.2-1.0 keV',
                                '1.0-2.0 keV',
                                '2.0-4.5 keV'],
@@ -60,10 +86,10 @@ if __name__ == "__main__":
                         f'tm{list(tms)[i]}_data.png'),
                         **plotting_kwargs)
     summed_data = np.sum(unmasked_data, axis=0)
-    bbox_info = [(7, 4), 7,  20, 'black']
-    plot(unmasked_data[i],
+    plot(summed_data,
+         pixel_factor=2,
          pixel_measure=28,
-         fs=8,
+         fs=12,
          title=['0.2-1.0 keV',
                 '1.0-2.0 keV',
                 '2.0-4.5 keV'],
@@ -75,13 +101,70 @@ if __name__ == "__main__":
                           f'summed_data.png'),
          bbox_info=bbox_info,
[…]
