### 2024-09-26 62ee58a
Perform correct cut
