+++ demos/paper/plot_gt_vs_rec_2d_hist.py
@@ -45,9 +45,11 @@ if __name__ == "__main__":
 
     masked_data = jax.tree_map(lambda x: np.array(x, dtype=np.float64),
                             masked_data)
-    plotting_kwargs =  {'bins': 600, 'x_label': '$s_{gt}$', 'y_label': 'a', 'dpi': 400, 'title': 'standardised error',
+    plotting_kwargs =  {'bins': 600, 'x_label': r'$s_{gt, p}$',
+                        'y_label': r'$r_\text{rel}$', 'dpi': 400,
+                        'title': 'standardised error',
                         'x_lim': (1e-11,1e-6), 'y_lim': (8e-2, 1e2)}
     plot_2d_gt_vs_rec_histogram(samples=samples, operator_dict=sky_dict, diagnostics_path=output_dir,
                                 response_dict=response_dict, type='sampled',relative=True, response=False,
                                 reference_dict=gt_dict, plot_kwargs=plotting_kwargs, base_filename='gtvsrec',
-                                fs=18, alpha=0.5, max_counts=1e2)
+                                fs=14, alpha=0.5, max_counts=1e2)
