### 2024-10-16 47a62fd
more_plotting_options: add time option for plotting, add option not to plot_rgb_samples.
