### 2024-10-22 1151f74
75-standardize-config-file-loading: no config_file loading for sky_models.py
