+++ src/sky_models.py
@@ -12,7 +12,7 @@ import nifty8.re as jft
 import numpy as np
 from ducc0.fft import good_size as good_fft_size
 
-from .utils import get_config, add_functions, add_models
+from .utils import add_functions, add_models
 
 
 class SkyModel:
@@ -23,30 +23,25 @@ class SkyModel:
 
     Parameters:
     ----------
-    config_file : python-dictionary, containing information about the grid,
-                   (the telescope (for the pixelization)), # TODO IMHO this should be part of the grid
-                   the priors on the correlated field and the uncorrelated inverse gamma
-                    component.
+    config : python-dictionary, containing information about the grid,
+    (the telescope (for the pixelization)), # TODO IMHO this should be part of the grid
+    the priors on the correlated field and the uncorrelated inverse gamma
+    component.
     """
 
-    def __init__(self, config_file_path=None):
+    def __init__(self, config: dict = None):
 
-        """Gets the parameters needed for building the sky model from the config file
-        given the corresponding path.
+        """Gets the parameters needed for building the sky model from the
+        provided config dictionary.
 
         Parameters
         ----------
-        config_file_path : string
-            Path to the config file
+        config : dict
+            YAML configuration dictionary.
         """
-        if config_file_path is not None:
-            if not isinstance(config_file_path, str):
-                raise TypeError("The path to the config file needs to be a string")
-            if not config_file_path.endswith('.yaml') and not config_file_path.endswith('.yml'):
-                raise ValueError("The sky model parameters need to be saved in a .yaml or .yml "
-                                 "file.")
-
-            self.config = get_config(config_file_path)
+        if config is not None:
+            if not isinstance(config, dict):
+                raise TypeError("The config must be a python-dictionary.")
         else:
             self.config = {}
         self.s_distances = None
