### 2024-10-22 f792711
75-standardize-config-file-loading: no config_file loading for eROSITA.
