### 2024-10-22 f612b49
75-standardize-config-file-loading: fix sky_models.py
