### 2024-11-06 7d6b8b1
mini: minimization processes (tmp)
