+++ src/__init__.py
@@ -45,5 +45,5 @@ from .likelihood import get_n_constrained_dof
 from .instruments.erosita.erosita_likelihood import generate_erosita_likelihood_from_config
 from .diagnostics import calculate_nwr, calculate_uwr
 from .convolve import linpatch_convolve, convolve, slice_patches
-from .minimization_parser import MinimizationParser
+from .minimization.minimization_parser import MinimizationParser
 from .hashcollector import save_local_packages_hashes_to_txt
