### 2024-11-07 3609bcc
minimization: point source minimization
