+++ src/minimization/point_source_minimization.py
@@ -1,0 +1,248 @@
+import pickle
+from os.path import join
+from pprint import pprint
+
+import matplotlib.pyplot as plt
+import nifty8.re as jft
+import jubik0 as ju
+import numpy as np
+from jax import random
+
+from demos.xi_experiments import build_peaks_filter_from_config
+from src.mf_utils import get_new_ps_locations_from_peaks
+from src.multi_component_sky import build_multi_component_sky_from_config
+from src.plot import simple_eval_plots, plot_xi_analysis, plot_xi_relaxation, \
+    plot_new_position
+from src.select_xis import get_relaxed_excitations, \
+    get_multi_component_model_new_position
+from src.sugar import save_all_minimization_outputs, \
+    clear_jax_compilation_cache, get_detection_maps_from_mc_xi_operators, \
+    get_point_source_locations, update_state_with_minimization_paser, \
+    save_all_minisanities, get_filtered_xi_maps, get_xi_maps_from_samples, \
+    get_detection_maps_from_xi_maps
+from src.xi_filters import build_radial_filter
+
+if __name__ == "__main__":
+    minimize = True
+    if not minimize:
+        resume_it = 3
+
+    # Reconstruction config
+    config_path = '../configs/'
+    config_filename = "erosita_mc_sf.yml"
+    config = ju.get_config(join(config_path, config_filename))
+    output_directory = ju.create_output_directory(config["files"]["res_dir"])
+    key = random.PRNGKey(config['seed'])
+
+    # Copy config file to results directory
+    ju.copy_config(config_filename, config_path, output_directory)
+
+    # Load response
+    response_dict = ju.build_erosita_response_from_config(config)
+
+    # Generate eROSITA data (if it does not already exist)
+    ju.generate_erosita_data_from_config(config)
+    ju.create_erosita_data_from_config(config, response_dict)
+
+    # Load masked_data
+    masked_data = ju.load_masked_data_from_config(config)
+
+    # Creates eROSITA data
+    ju.mask_erosita_data_from_disk(
+        config['files'],
+        config['telescope'],
+        config['grid'],
+        response_dict['mask'],
+    )
+
+    # Load reconstruction model
+    ps_locations = np.empty((0, 2), dtype=np.float32)
+    ps_loc_dict = dict(background=ps_locations, )
+    sky_model = build_multi_component_sky_from_config(
+        config, point_source_positions=ps_loc_dict)
+
+    # Load plotting dict
+    plot_dict = sky_model.get_plottable_sky_dict()
+
+    # Load xi operators
+    xi_ops = sky_model.get_xi_operators()
+
+    # Plot callback
+    def plot_callback(s, x):
+        return simple_eval_plots(s, x,
+                                 output_directory=output_directory,
+                                 plotting_cfg=config['plotting'],
+                                 sky_dict=plot_dict,
+                                 xi_operators=xi_ops
+                                 )
+
+    # Generate response
+    def response(x):
+        return response_dict['exposure'](
+            response_dict['psf'](x * response_dict['pix_area'],
+            response_dict['kernel']))
+
+
+    # Generate loglikelihood (Building masked (mock) data and response)
+    log_likelihood = ju.generate_erosita_likelihood(response_dict,
+                                                    masked_data,
+                                                    sky_model,
+                                                    )
+
+    # Minimization
+    key, subkey = random.split(key)
+    detection_config = config['anomaly_detection']
+    pos_init = sky_model.init(subkey) * detection_config['relaxation_factor']
+    minimization_config = config['minimization']
+
+    def callback(s, x):
+        save_all_minimization_outputs(
+            x,
+            output_directory=output_directory,
+        )
[…]
