### 2024-11-07 eb1d0ea
mini: restructure lenscharm to minimization_from_initial_samples
