### 2024-11-12 06a02b0
75-standardize-config-file-loading: YAML -> dictionary renaming in docstrings.
