### 2024-11-12 db26e91
75-standardize-config-file-loading: better SkyModel docs.
