### 2024-11-22 3aecdb1
minimization: bugfix return samples and state
