### 2024-12-11 d0e8eba
minimization: some renaming and small fix.
