### 2024-12-12 3fefee1
77-erosita-pointing: introduce pointing_center as a requirement for eROSITA imaging.
