### 2024-12-16 cd0e08b
Plotting Chandra data
