### 2024-12-16 0248cd6
Adjust custom energy ranges in exposure
