### 2024-12-16 cd4d3ad
77-erosita-pointing: remove unnecessary check.
