### 2024-12-16 6fa552a
79-erosita-pointing: re-center the erosita_psf pointings around the observation pointing.
