### 2024-12-17 ab37ddb
79-erosita-pointing: address review.
