### 2024-12-24 cf7c111
81-fix-minimizationparser: fix minimization parser delta switches logic.
