### 2024-12-24 43a8be6
81-fix-minimizationparser: fix indexing.
