+++ src/instruments/jwst/plotting/jwst_plotting.py
@@ -1,0 +1,902 @@
+from os.path import join
+from os import makedirs
+from typing import Union, Optional
+from dataclasses import dataclass, field
+
+import numpy as np
+from numpy.typing import ArrayLike
+import matplotlib.pyplot as plt
+from matplotlib.colors import Normalize, LogNorm
+
+
+import nifty8.re as jft
+
+from ..mock_data.mock_evaluation import redchi2
+from ..mock_data.mock_plotting import display_text
+from ..rotation_and_shift.coordinates_correction import CoordinatesWithCorrection
+from ..filter_projector import FilterProjector
+from ..grid import Grid
+
+from charm_lensing.lens_system import LensSystem
+from charm_lensing.physical_models.multifrequency_models.colormix_model import (
+    ColorMix, NonSymmetricColorMatrix)
+
+
+def find_closest_factors(number):
+    """
+    Finds two integers whose multiplication is larger or equal to the input number,
+    with the two output numbers being as close together as possible.
+
+    Args:
+        number: The input integer number.
+
+    Returns:
+        A tuple containing two integers (x, y) such that x * y >= number and the
+        difference between x and y is minimized. If no such factors exist, returns
+        None.
+    """
+
+    # Start with the square root of the number.
+    ii = int(np.ceil(number**0.5))
+
+    jj, kk = ii, ii
+
+    jminus = kminus = 0
+    while ((jj-jminus)*(kk-kminus) >= number):
+        if kminus == jminus:
+            jminus += 1
+        else:
+            kminus += 1
+
+    if jminus == kminus:
+        return jj-jminus, kk-kminus+1
+    return jj-jminus+1, kk-kminus
+
+
+def get_position_or_samples_of_model(
+    position_or_samples: Union[dict, jft.Samples],
+    model: jft.Model,
+) -> tuple[ArrayLike, ArrayLike]:
+    '''Returns (mean, std) of model'''
+
+    if isinstance(position_or_samples, jft.Samples):
+        mean, std = jft.mean_and_std(
+            [model(si) for si in position_or_samples])
+    else:
+        mean = model(position_or_samples)
+        if isinstance(mean, dict):
+            std = {k: np.full(v.shape, 0) for k, v in mean.items()}
+        else:
+            std = np.full(mean.shape, 0.)
+
+    return mean, std
+
+
+def get_shift_rotation_correction(
+    position_or_samples: Union[dict, jft.Samples],
+    correction_model: Optional[CoordinatesWithCorrection],
+):
+    if not isinstance(correction_model, CoordinatesWithCorrection):
+        return (0, 0), (0, 0), 0, 0
+
+    shift_mean, shift_std = get_position_or_samples_of_model(
+        position_or_samples, correction_model.shift_prior)
+    rotation_mean, rotation_std = get_position_or_samples_of_model(
+        position_or_samples, correction_model.rotation_prior)
+    shift_mean, shift_std = shift_mean.reshape(2), shift_std.reshape(2)
+    rotation_mean, rotation_std = rotation_mean[0], rotation_std[0]
+
+    return (shift_mean, shift_std), (rotation_mean, rotation_std)
+
+
+def _get_data_model_and_chi2(
+    position_or_samples: Union[dict, jft.Samples],
+    sky_or_skies: ArrayLike,
+    data_model: jft.Model,
+    data: ArrayLike,
+    mask: ArrayLike,
+    std: ArrayLike,
+):
+    if isinstance(std, float):
+        std = np.full_like(data, std)
+
+    if isinstance(position_or_samples, jft.Samples):
+        model_data = []
+        for ii, si in enumerate(position_or_samples):
+            tmp = np.zeros_like(data)
+            while isinstance(si, jft.Vector):
[…]
