+++ src/instruments/jwst/wcs/wcs_astropy.py
@@ -4,28 +4,114 @@
 # Copyright(C) 2024 Max-Planck-Society
 
 # %%
+from .wcs_base import WcsBase
+from ..parse.wcs.coordinate_system import (
+    CoordinateSystemModel, CoordinateSystems)
+from ..parse.wcs.wcs_astropy import WcsModel
 
 import numpy as np
 
 from numpy.typing import ArrayLike
-from typing import List, Union, Tuple
+from typing import List, Union, Optional
 
 from astropy.coordinates import SkyCoord
 from astropy.wcs import WCS
-from astropy.units import Unit
-from astropy import units
-
-from .wcs_base import WcsBase
+from astropy import units as u
 
 
-class WcsAstropy(WcsBase):
+class WcsAstropy(WCS, WcsBase):
     """
-    A class for converting between world coordinates and pixel coordinates
-    in JWST data.
+    A wrapper around the astropy wcs, in order to define a common interface
+    with the gwcs.
     """
-    def __init__(self, wcs):
-        super().__init__(wcs)
 
+    def __init__(
+        self,
+        center: SkyCoord,
+        shape: tuple[int, int],
+        fov: tuple[u.Quantity, u.Quantity],
+        rotation: u.Quantity = 0.0 * u.deg,
+        coordinate_system: Optional[CoordinateSystemModel] = CoordinateSystems.icrs.value
+    ):
+        """
+        Create FITS header, use it to instantiate an WcsAstropy.
+
+        Parameters
+        ----------
+        center : SkyCoord
+            The value of the center of the coordinate system (crval).
+        shape : tuple
+            The shape of the grid.
+        fov : tuple
+            The field of view of the grid. Typically given in degrees.
+        rotation : u.Quantity
+            The rotation of the grid WCS, in degrees.
+        coordinate_system : CoordinateSystemConfig
+            Coordinate system to use ('icrs', 'fk5', 'fk4', 'galactic')
+        equinox : float, optional
+            Equinox for FK4/FK5 systems (e.g., 2000.0 for J2000)
+        """
+
+        self.shape = shape
+        self.fov = fov
+        self.distances = [
+            f.to(u.deg)/s for f, s in zip(fov, shape)]
+        self.center = center
+
+        # Calculate rotation matrix
+        rotation_value = rotation.to(u.rad).value
+        pc11 = np.cos(rotation_value)
+        pc12 = -np.sin(rotation_value)
+        pc21 = np.sin(rotation_value)
+        pc22 = np.cos(rotation_value)
+
+        # Transform center coordinates if necessary
+        if coordinate_system.radesys == CoordinateSystems.galactic.value.radesys:
+            lon = center.galactic.l.deg
+            lat = center.galactic.b.deg
+        else:
+            lon = center.ra.deg
+            lat = center.dec.deg
+
+        # Build the header dictionary
+        header = {
+            'WCSAXES': 2,
+            'CTYPE1': coordinate_system.ctypes[0],
+            'CTYPE2': coordinate_system.ctypes[1],
+            'CRPIX1': shape[0] / 2 + 0.5,
+            'CRPIX2': shape[1] / 2 + 0.5,
+            'CRVAL1': lon,
+            'CRVAL2': lat,
+            'CDELT1': -fov[0].to(u.deg).value / shape[0],
+            'CDELT2': fov[1].to(u.deg).value / shape[1],
+            'PC1_1': pc11,
+            'PC1_2': pc12,
+            'PC2_1': pc21,
+            'PC2_2': pc22,
+            'RADESYS': coordinate_system.radesys,
+            'CUNIT1': 'deg',
+            'CUNIT2': 'deg',
+        }
+
+        # Set equinox if needed for FK4/FK5
+        if coordinate_system.radesys in [
[…]
