+++ src/instruments/jwst/zero_flux_model.py
@@ -5,47 +5,45 @@
 
 # %
 
+from .parametric_model import build_parametric_prior_from_prior_config
+from .parse.parametric_model.parametric_prior import ProbabilityConfig
+
 import nifty8.re as jft
 
-from .parametric_model import build_parametric_prior
+from typing import Optional
 
 ZERO_FLUX_KEY = 'zero_flux'
+DEFAULT_KEY = 'default'
+SHAPE = (1,)
 
 
 def build_zero_flux_model(
     prefix: str,
-    likelihood_config: dict,
-) -> jft.Model:
+    prior_config: Optional[ProbabilityConfig],
+) -> Optional[jft.Model]:
     """
-    Build a zero flux model based on the provided likelihood configuration.
-
-    If no specific configuration for the zero-flux model is found in
-    `likelihood_config`, a default model returning zero is created.
-    Otherwise, it builds a parametric model based on the provided configuration.
+    Build a zero flux model based on the provided PriorConfig.
+    If `prior_config` is None, None is returned.
 
     Parameters
     ----------
     prefix : str
-        A string prefix used to identify and name the parameters associated with
-         the zero-flux model.
-    likelihood_config : dict
-        A configuration dictionary containing model details.
-        The zero-flux model configuration is expected to be under the key
-        specified by `ZERO_FLUX_KEY`.
+        A string prefix used for the prior domain of the zero-flux model.
+    prior_config : Optional[PriorConfig]
+        The prior config, which is used to instantiate the prior probability 
+        call.
 
     Returns
     -------
-    jft.Model
-        A model representing the zero flux configuration.
-        If no configuration is provided, the model returns zero;
-        otherwise, it uses a parametric prior.
+    Optional[jft.Model]
+        A physical model representing the zero-flux of the observation.
+        If `prior_config` is None, None is returned.
     """
-    model_cfg = likelihood_config.get(ZERO_FLUX_KEY, None)
-    if model_cfg is None:
-        return jft.Model(lambda _: 0, domain=dict())
+    if prior_config is None:
+        return None
 
     prefix = '_'.join([prefix, ZERO_FLUX_KEY])
 
-    shape = (1,)
-    prior = build_parametric_prior(prefix, model_cfg['prior'], shape)
-    return jft.Model(prior, domain={prefix: jft.ShapeWithDtype(shape)})
+    prior = build_parametric_prior_from_prior_config(
+        prefix, prior_config, SHAPE)
+    return jft.Model(prior, domain={prefix: jft.ShapeWithDtype(SHAPE)})
