### 2025-01-22 1b26412
Fix plotting norm
