### 2025-01-23 170cbc8
copy resolve
