+++ src/instruments/resolve/re/__init__.py
@@ -2,4 +2,4 @@
 from .sky_model import sky_model_diffuse, sky_model_points, sky_model
 from .response import InterferometryResponse, InterferometryResponseFinuFFT, InterferometryResponseDucc
 
-# from .mosaicing import *
+from .mosaicing import *
