### 2025-01-23 0a45bd6
resolve: copy config easy
