+++ src/__init__.py
@@ -7,6 +7,7 @@
 from .version import __version__
 from .utils import (save_to_pickle, load_from_pickle, get_config, save_to_yaml,
                     copy_config, create_output_directory,
+                    save_config_copy_easy,
                     coord_center, get_stats)
 from .plot import (plot_result, plot_histograms,
                    plot_sample_averaged_log_2d_histogram, plot_rgb)
@@ -48,4 +49,3 @@ from .diagnostics import calculate_nwr, calculate_uwr
 from .convolve import linpatch_convolve, convolve, slice_patches
 from .minimization_parser import MinimizationParser
 from .hashcollector import save_local_packages_hashes_to_txt
-
