### 2025-01-25 952cb92
build psf from config for debugging
