+++ src/instruments/erosita/erosita_response.py
@@ -357,6 +357,61 @@ def build_erosita_response(
     return response_dict
 
 
+def build_psf_from_config(config):
+
+    tel_info = config['telescope']
+    file_info = config['files']
+    psf_info = config['psf']
+    grid_info = config['grid']
+
+    # load calibration directory paths
+    caldb_path = file_info['calibration_path']
+    caldb_dir_name = file_info['caldb_folder_name']
+
+    psf_file_suffix = file_info['psf_filename_suffix']
+
+    psf_filenames = [get_erosita_psf_filenames(caldb_path,
+                                                key,
+                                                psf_filename_suffix=psf_file_suffix,
+                                                caldb_name=caldb_dir_name, )
+                     for key in tel_info['tm_ids']]
+
+    obs_instance = ErositaObservation(file_info['input'],
+                                      file_info['output'],
+                                      file_info['obs_path'])
+
+    center_stats = []
+    for tm_id in tel_info['tm_ids']:
+        tmp_center_stat = obs_instance.get_pointing_coordinates_stats(
+            tm_id,
+            file_info['input'])
+        tmp_center_stat = [tmp_center_stat['RA'][0], tmp_center_stat['DEC'][0]]
+        center_stats.append(tmp_center_stat)
+    center_stats = np.array(center_stats)
+
+    # center with respect to desired pointing center
+    ref_center = np.array(tel_info['pointing_center']) * 3600  # to arcsec
+    d_centers = center_stats - ref_center
+
+    # Set the image pointing to the center
+    image_pointing_center = np.array((tel_info['fov'] / 2.,) * 2)
+    # FIXME: image pointing center should be added somewhere else
+    # FIXME IMHO this image poiting center is not needed?
+    pointing_center = d_centers + image_pointing_center
+
+
+    domain = Domain(tuple([grid_info['edim']] + [grid_info['sdim']] * 2),
+                    tuple([1] + [tel_info['fov'] / grid_info['sdim']] * 2))
+    psf_func, kernel = build_erosita_psf(psf_filenames,
+                                         psf_info['energy'],
+                                         pointing_center,
+                                         domain,
+                                         psf_info['npatch'],
+                                         psf_info['margfrac'])
+
+    return psf_func, kernel
+
+
 def build_erosita_response_from_config(
     config: dict
 ) -> dict:
