### 2025-01-25 c667523
better plotting for log
