### 2025-01-25 8475025
solved pointing problem
