### 2025-01-28 8e040ee
resolve: remove MPI util
