+++ src/instruments/resolve/__init__.py
@@ -17,7 +17,6 @@ from .logger import logger
 # from .library.primary_beams import *
 # from .likelihood import *
 # from .mosaicing import *
-from .mpi import barrier, onlymaster
 # from .mpi_operators import *
 # from .points import PointInserter
 # from .polarization_matrix_exponential import *
