### 2025-01-28 4ae27e2
resovle: set JAX to f64 in demo
