### 2025-01-28 c3f32d7
resolve: cleanup util
