+++ src/instruments/resolve/util.py
@@ -14,12 +14,6 @@
 # Copyright(C) 2020-2022 Max-Planck-Society, Philipp Arras
 # Author: Philipp Arras
 
-import cProfile
-import io
-import pstats
-from pstats import SortKey
-from time import time
-
 import nifty8 as ift
 import numpy as np
 
@@ -86,116 +80,6 @@ def _types_equal(a, b):
     return type(a) == type(b)
 
 
-def divide_where_possible(a, b):
-    # Otherwise one
-    if isinstance(a, ift.Field):
-        dom = a.domain
-        a = a.val
-        dtype = a.dtype
-        if isinstance(b, ift.Field):
-            my_asserteq(b.dtype, a.dtype)
-    elif isinstance(b, ift.Field):
-        dom = b.domain
-        b = b.val
-        dtype = b.dtype
-    else:
-        raise TypeError
-    arr = np.divide(a, b, out=np.ones(dom.shape, dtype=dtype), where=b != 0)
-    return ift.makeField(dom, arr)
-
-
-def imshow(arr, ax=None, **kwargs):
-    import matplotlib.pyplot as plt
-
-    if ax is None:
-        ax = plt.gca()
-    return ax.imshow(arr.T, origin="lower", **kwargs)
-
-
-def rows_to_baselines(antenna_positions, data_field):
-    ua = antenna_positions.unique_antennas()
-    my_assert(np.all(antenna_positions.ant1 < antenna_positions.ant2))
-
-    res = {}
-    for iant1 in range(len(ua)):
-        for iant2 in range(iant1+1, len(ua)):
-            res[f"{iant1}-{iant2}"] = antenna_positions.extract_baseline(iant1, iant2, data_field)
-    return res
-
-
-def assert_sky_domain(dom):
-    """Check that input fulfils resolve's conventions of a sky domain.
-
-    A sky domain is a DomainTuple:
-
-    dom = (pdom, tdom, fdom, sdom)
-
-    where `pdom` is a `PolarizationSpace`, `tdom` and `fdom` are  an `IRGSpace`,
-and `sdom` is a two-dimensional `RGSpace`.
-
-    Parameters
-    ----------
-    dom : DomainTuple
-        Object that is checked to fulfil the properties.
-    """
-    from .irg_space import IRGSpace
-    from .polarization_space import PolarizationSpace
-
-    my_assert_isinstance(dom, ift.DomainTuple)
-    my_asserteq(len(dom), 4)
-    pdom, tdom, fdom, sdom = dom
-    for ii in range(3):
-        my_asserteq(len(dom[ii].shape), 1)
-    my_asserteq(len(sdom.shape), 2)
-    my_assert_isinstance(pdom, PolarizationSpace)
-    my_assert_isinstance(tdom, IRGSpace)
-    my_assert_isinstance(fdom, IRGSpace)
-    my_assert_isinstance(sdom, ift.RGSpace)
-    my_asserteq(len(sdom.shape), 2)
-
-
-def assert_data_domain(dom):
-    """Check that input fulfils resolve's conventions of a data domain.
-
-    A data domain is a DomainTuple:
-
-    dom = (pdom, rdom, fdom)
-
-    where `pdom` is a `PolarizationSpace`, `rdom` is an UnstructuredDomain representing the rows of the measurement set and
-    `fdom` is an `IRGSpace` containing the frequencies.
-
-    Parameters
-    ----------
-    dom : DomainTuple
-        Object that is checked to fulfil the properties.
-    """
-    from .irg_space import IRGSpace
-    from .polarization_space import PolarizationSpace
-
-    my_assert_isinstance(dom, ift.DomainTuple)
[…]
