+++ src/sky_model/parse/resolve_sky.py
@@ -1,0 +1,5 @@
+from dataclasses import dataclass
+
+@dataclass
+class ResolveSkyModel:
+    pass
