### 2025-01-29 fe8498a
resolve: move parsing, grid, and color
