+++ src/instruments/jwst/rotation_and_shift/rotation_and_shift.py
@@ -15,9 +15,9 @@ from .coordinates_correction import CoordinatesWithCorrection
 from .linear_rotation_and_shift import build_linear_rotation_and_shift
 from .nufft_rotation_and_shift import build_nufft_rotation_and_shift
 from .sparse_rotation_and_shift import build_sparse_rotation_and_shift
-from ..grid import Grid
-from ..wcs import subsample_grid_corners_in_index_grid_non_vstack
-from ..wcs.wcs_base import WcsBase
+from ....grid import Grid
+from ....wcs import subsample_grid_corners_in_index_grid_non_vstack
+from ....wcs.wcs_base import WcsBase
 
 from ..parse.rotation_and_shift.rotation_and_shift import(
     LinearConfig, NufftConfig, SparseConfig, RotationAndShiftAlgorithm
