### 2025-01-29 cc514ec
resolve: delete unneeded reconstruction grid class
