+++ src/instruments/jwst/__init__.py
@@ -6,8 +6,6 @@
 # %%
 
 from . import integration_model
-from . import wcs
 
 from .config_handler import *
 from .jwst_data import JwstData
-from .grid import Grid
