### 2025-01-29 0c77fa6
resolve: point sources developed
