+++ src/sky_model/resolve_sky.py
@@ -5,22 +5,22 @@ import nifty8 as ift
 import nifty8.re as jft
 import jax
 from jax import numpy as jnp
+
+from ..parse.sky_model.resolve_sky import ResolvePointSourcesModel
+
+from astropy import units as u
+
 import numpy as np
 
-
 # TODO: replace with new config parsing!
-def _spatial_dom(cfg):
-    nx = cfg.getint("space npix x")
-    ny = cfg.getint("space npix y")
-    dx = str2rad(cfg["space fov x"]) / nx
-    dy = str2rad(cfg["space fov y"]) / ny
-    return ift.RGSpace([nx, ny], [dx, dy])
 
 ARCMIN2RAD = np.pi / 60 / 180
 AS2RAD = ARCMIN2RAD / 60
 DEG2RAD = np.pi / 180
 SPEEDOFLIGHT = 299792458.
 # TODO: replace with new astropy unit handling!
+
+
 def str2rad(s):
     """Convert string of number and unit to radian.
 
@@ -48,6 +48,15 @@ def str2rad(s):
         if unit == kk:
             return float(s[:nn]) * c[kk]
     raise RuntimeError("Unit not understood")
+
+
+def _spatial_dom(cfg):
+    nx = cfg.getint("space npix x")
+    ny = cfg.getint("space npix y")
+    dx = str2rad(cfg["space fov x"]) / nx
+    dy = str2rad(cfg["space fov y"]) / ny
+    return ift.RGSpace([nx, ny], [dx, dy])
+
 
 def build_cf(prefix, conf, shape, dist):
     zmo = conf.getfloat(f"{prefix} zero mode offset")
@@ -83,7 +92,8 @@ def build_cf(prefix, conf, shape, dist):
 
 def sky_model_diffuse(cfg):
     if not cfg["freq mode"] == "single":
-        raise NotImplementedError("FIXME: only implemented for single frequency")
+        raise NotImplementedError(
+            "FIXME: only implemented for single frequency")
     if not cfg["polarization"] == "I":
         raise NotImplementedError("FIXME: only implemented for stokes I")
     sky_dom = _spatial_dom(cfg)
@@ -111,40 +121,35 @@ def jax_insert(x, ptsx, ptsy, bg):
     return bg
 
 
-def sky_model_points(cfg, bg=None):
-    if cfg["freq mode"] == "single":
-        if cfg["polarization"] == "I":
-            if cfg["point sources mode"] == "single":
-                ppos = []
-                sky_dom = _spatial_dom(cfg)
-                s = cfg["point sources locations"]
-                for xy in s.split(","):
-                    x, y = xy.split("$")
-                    ppos.append((str2rad(x), str2rad(y)))
-                ppos = np.array(ppos)
-                dx = np.array(sky_dom.distances)
-                center = np.array(sky_dom.shape) // 2
-                inds = np.unique(np.round(ppos / dx + center).astype(int).T, axis=1)
-                indsx, indsy = inds
-                alpha = cfg.getfloat("point sources alpha")
-                q = cfg.getfloat("point sources q")
-                sky_dom = _spatial_dom(cfg)
+def sky_model_points(
+    sky_dom: ift.RGSpace,
+    resolve_point_souces_model: ResolvePointSourcesModel,
+    bg=None
+):
+    if resolve_point_souces_model.freq_mode == "single":
+        if resolve_point_souces_model.polarization == "I":
+            if resolve_point_souces_model.mode == "single":
+
+                indsx, indsy = resolve_point_souces_model.locations.to_indices(
+                    sky_dom, unit=u.rad)
                 shp = sky_dom.shape
                 full_shp = (1, 1, 1) + shp
+
[…]
