+++ demos/cygnusa_2ghz.cfg
@@ -20,6 +20,6 @@ stokesI diffuse space i0 asperity mean =  0.2
 stokesI diffuse space i0 asperity stddev = 0.2
 
 point sources mode = single
-point sources locations = 0deg$0deg,0.35as$-0.22as
-point sources alpha = 0.5
-point sources q = 0.2
+point sources relative locations = 0deg$0deg,0.35as$-0.22as
+point sources a = 0.5
+point sources scale = 0.2

[…]
