### 2025-01-29 6d25ec9
resolve: docstring
