### 2025-01-31 5780c0f
resolve: WHAT happens if obs.times = None
