### 2025-02-02 2d00d9f
resolve: some unneeded check
