### 2025-02-03 86982b5
fix map pspec plotting
