+++ src/instruments/erosita/erosita_psf_utils.py
@@ -21,11 +21,13 @@ def to_r_phi(cc):
     # FIXME: This assumes that ra is the x-coordinate and dec the y-coordinate
     # and furthermore assumes that the psfs are given in vertical distances
     # (off axis angle). Ensure that this is the correct orientation!
+    # VE: The "-jnp.pi/2" come from the fact that the psfset is changing
+    # as a function of theta and phi = 0 while phi=0 seems to be measured as the angle from the y axis
     r = jnp.sqrt(cc[..., 0]**2 + cc[..., 1]**2)
     phi = jnp.angle(cc[..., 0] + 1.j*cc[..., 1]) - jnp.pi/2.
     return jnp.stack((r, phi), axis = -1)
 
-
+# TODO ADD TEST check orthogonality, works, but push the test!
 def to_ra_dec(rp):
     """
     Transforms form r-phi coordinates to ra-dec (sky) coordinates.
@@ -53,7 +55,7 @@ def get_interpolation_weights(rs, r):
     def _get_wgt_back(i):
         res = jnp.zeros(rs.shape, dtype=float)
         res = res.at[rs.size-1].set(1.)
-        return res    
+        return res
     res += cond(r >= rs[rs.size-1], _get_wgt_back,
                 lambda _: jnp.zeros(rs.shape, dtype=float), 0)
 
@@ -130,6 +132,12 @@ def get_psf(psfs, rs, patch_center_ids, patch_deltas, pointing_center):
             raise ValueError
 
     def psf(ra, dec, dra, ddec):
+        """
+        ra: right ascession of the center a patch
+        dec: declination of the center of a patch
+        dra: relative coordinate grid from "the" center (IMHO should be the poiting center)
+        ddec: relative coordinate grid from "the" center (sie abaf)
+        """
         # Find r and phi corresponding to requested location
         cc = jnp.stack((ra, dec), axis = -1)
         cc -= pointing_center
@@ -142,7 +150,7 @@ def get_psf(psfs, rs, patch_center_ids, patch_deltas, pointing_center):
         wgts = jnp.moveaxis(wgts, -1, 0)
 
         # Rotate requested psf slice to align with patch
-        int_coords = jnp.stack((dra, ddec), axis = -1)
+        int_coords = jnp.stack((dra, ddec), axis=-1)
         int_coords = to_r_phi(int_coords)
         int_coords = jnp.moveaxis(int_coords, -1, 0)
         rp = jnp.moveaxis(rp, -1, 0)
@@ -159,7 +167,6 @@ def get_psf(psfs, rs, patch_center_ids, patch_deltas, pointing_center):
             res += ww*int_res
 
         return res
-
     return psf
 
 
@@ -206,18 +213,35 @@ def psf_interpolator(domain, npatch, psf_infos):
         # centers = (np.array([(i*ss + ss/2 + 0.5)*dd for i in range(npatch)])
         #            for ss, dd in zip(patch_shp, dist))
 
+        # relative coordinates of the PIXEL (CENTERS) corners for all image pixel
         c_p = ((np.arange(ss) - ss/2)*dd for ss, dd in zip(shp, dist))
+        # relative coordinates of the patch centers (relative to the upper left corner?)
         centers = (np.array([(i*ss + ss/2)*dd for i in range(npatch)]) for
                    ss, dd in zip(patch_shp, dist))
 
         c_p = np.meshgrid(*c_p, indexing='ij')
-        d_ra = c_p[0]
-        d_dec = c_p[1]
+        # FIXME IMHO there should be minus here
+        # Flip ra and dec and change sign for ra
+        # OLD
+        # d_ra = c_p[0]
+        # d_ra= c_p[1]
+        # NEW
+        d_dec = c_p[0]  # NEW
[…]
