+++ src/__init__.py
@@ -32,7 +32,8 @@ from .response import build_exposure_function, build_readout_function
 from .instruments.erosita.erosita_response import (build_callable_from_exposure_file,
                                                    build_erosita_psf,
                                                    build_erosita_response_from_config,
-                                                   load_erosita_response)
+                                                   load_erosita_response,
+                                                   build_psf_from_config)
 from .instruments.erosita.erosita_likelihood import (generate_erosita_likelihood,
                                                      generate_erosita_likelihood_from_config)
 from .instruments.erosita.erosita_data import (generate_erosita_data_from_config,
