+++ demos/paper/plot_eROSITA_data.py
@@ -2,6 +2,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import pickle
 from jax import linear_transpose, vmap
+from matplotlib import patches
 import jax.numpy as jnp
 
 import jubik0 as ju
@@ -56,7 +57,7 @@ if __name__ == "__main__":
     correct_exposures_for_effective_area = True
     if correct_exposures_for_effective_area:
         # from src.library.response import calculate_erosita_effective_area
-        ea = ju.calculate_erosita_effective_area(path_to_caldb, tm_ids, e_min, e_max)
+        ea = ju.instruments.erosita.erosita_response.calculate_erosita_effective_area(path_to_caldb, tm_ids, e_min, e_max)
         exposures *= ea[:, :, np.newaxis, np.newaxis]
 
     summed_data = np.sum(data, axis=0)
@@ -70,10 +71,14 @@ if __name__ == "__main__":
     exposure_corrected_data[mask_exp] = 0
     bbox_info = [(28, 16), 28, 160, 'black']
 
+    sat_min = {'log': [1.2e-9, 1.0e-10, 2.0e-10],
+               "lin": [1e-10, 1e-10, 1e-10]}
+    sat_max = {'log': [2.1e-7, 1.5e-7, 1.5e-7],
+               "lin": [2.3e-8, 1.5e-8, 1.e-8]}
     #### LOG Plot
     plot_rgb(exposure_corrected_data,
-             sat_min=[2e-12, 2e-12, 2e-12],
-             sat_max=[1, 1, 1],
+             sat_min=sat_min['log'],
+             sat_max=sat_max['log'],
              log=True,
              title='eROSITA LMC data', fs=18, pixel_measure=112,
              output_file=join(output_dir, 'log_expcor_eRSOITA_data_rgb.png'),
@@ -84,8 +89,8 @@ if __name__ == "__main__":
 
     #### Lin Plot
     plot_rgb(exposure_corrected_data,
-             sat_min=[1e-10, 1e-10, 1e-10],
-             sat_max=[2.3e-8, 1.5e-8, 1.e-8],
+             sat_min=sat_min['lin'],
+             sat_max=sat_min['lin'],
              # log=True,
              title='eROSITA LMC data', fs=18, pixel_measure=112,
              output_file=join(output_dir, 'lin_expcor_eRSOITA_data_rgb.png'),
@@ -119,10 +124,10 @@ if __name__ == "__main__":
              **plotting_kwargs_rec
              )
         if i==0:
-            plot_rgb(exposure_corrected_data_tm[:, 570: 770,  150: 350],
-                     sat_min=[1e-10, 1e-10, 1e-10],
-                     sat_max=[2.3e-8, 1.5e-8, 1.e-8],
-                     # log=True,
+            plot_rgb(exposure_corrected_data_tm[:, 570: 770,  130: 330],
+                     sat_min=sat_min['log'],
+                     sat_max=sat_max['log'],
+                     log=True,
                      title='eROSITA LMC data TM1', fs=32,
                      output_file=join(output_dir,
                                       'zoom_expcor_eRSOITA_data_rgb_tm1.png'),
@@ -131,22 +136,22 @@ if __name__ == "__main__":
                      bbox_info=bbox_info,
                      )
     ### Zoom
-    zoomed_expcor_data = exposure_corrected_data[:, 570: 770,  150: 350]
+    zoomed_expcor_data = exposure_corrected_data[:, 570: 770,  130: 330]
     plot_rgb(zoomed_expcor_data,
-             sat_min=[1e-10, 1e-10, 1e-10],
-             sat_max=[2.3e-8, 1.5e-8, 1.e-8],
[…]
