### 2025-02-11 #7768 @ivastar
Dear @prashjha and @hamogu, thank you for agreeing to review this submission for JOSS! The instructions on the review are at the top of the issue. Please use the suggested command to generate your individual checklist and consult the reviewer instructions for a more detailed explanation on what we are looking for in each item. If you have any questions or concerns, do not hesitate to reach out to me, either tag me here or email me (momcheva@mpia.de).

We are looking for a first round or reviews in ~3 weeks, that takes us to ~March 4th for this review. Please let me (and the author) know if you need more time than this.
