### 2025-02-11 100df43
disentangle linpatch convolve
