### 2025-02-14 072ba63
resolve: one can add sky_beamers
