### 2025-02-16 a238487
resolve: multi messenger metadata
