### 2025-02-17 1d16fdd
resolve: adjust reference bin loading
