### 2025-02-17 88f7641
jwst: lens light plotting bugfix
