### 2025-02-18 NIFTy-PPL/J-UBIK#2 @prashjet
Error running `jwst_demo.py`
Running `python jwst_demo.py` gives the following error:

```
(jubik) prash@air > python jwst_demo.py
Ciao is not sourced or installed. Therefore some operations can't be performed
shapely not installed. Some JWST functions will not work.
Config file saved to: results/jwst_demo/jwst_demo.yaml.
Traceback (most recent call last):
  File "/Users/prash/Astro/Refereeing/JOSS_25_02/J-UBIK/demos/jwst_demo.py", line 162, in <module>
    sky_model = ju.SkyModel(config_path)
  File "/opt/miniconda3/envs/jubik/lib/python3.10/site-packages/jubik0/sky_models.py", line 36, in __init__
    raise TypeError("The config must be a python-dictionary.")
TypeError: The config must be a python-dictionary.
```

Explicitly including the location of the config file gives the same outcome.
