### 2025-02-18 #7768 @prashjet
Hi @wmarg. I wanted to raise two quick points before I can leave my full review:

1. I see there is a "User Guide" section in the docs with the following header,

> This guide is an overview and explains the main conceptual idea behind J-UBIK (JAX-accelerated Universal Bayesian Imaging Kit for X-ray).

but it's otherwise empty. I just wanted to check whether this conceptual overview is accidentally missing? If it does exist, it would be very useful to see!

2. [This error](https://github.com/NIFTy-PPL/J-UBIK/issues/2) is preventing me from running the demo scripts.
