+++ demos/configs/spt0418.yaml
@@ -1,0 +1,425 @@
+cpu: False
+prior_samples: False
+plot_results: True
+nonparametric_lens: True
+max_residuals: 4 # 4 for big run
+no_interactive_plotting: True
+
+
+# Before long run:
+# 1) No prior samples
+# 2) No interactive plotting
+# 3) Frequencies & reference frequency
+# 4) No fixed positions
+# 5) Large Psf
+# 6) Output directory
+# 7) (Optional) Download latest webbpsf and delete stashed Psfs: https://webbpsf.readthedocs.io/en/stable/installation.html
+
+
+sky: 
+  grid:
+    sdim: 384
+    s_padding_ratio: 1.5  # Size of the convergence, i.e. (384*1.5, 6*1.5)
+    # NOTE: SHOULD THIS BE HERE?????
+
+    fov: 6
+    fov_unit: arcsec
+    rotation: 0
+    coordinate_frame: icrs
+    sky_center:
+      ra: 64.66543063107049
+      dec: -47.86462563973049
+      unit: deg
+
+    source_grid:
+      sdim: 384
+      fov: 3
+      s_padding_ratio: 1.0
+
+    energy_unit: eV
+    energy_bin:
+      e_min:
+        # - 0.054 # F2100W
+
+        - 0.068 # F1800W  # mistake here actually 0.064
+
+        # - 0.075 # F1500W
+        # - 0.093 # F1280W  # mistake here actually 0.088
+        # - 0.114 # F1000W
+        # - 0.143 # F770W
+        # - 0.201 # F560W
+
+        - 0.249 # F444W
+
+        # - 0.319 # F356W  # mistake here actually 0.311
+        # - 0.396 # F277W
+        # - 0.557 # F200W
+        # - 0.743 # F150W
+
+        - 0.967 # F115W
+      e_max:
+        # - 0.067 # F2100W
+
+        - 0.075 # F1800W
+
+        # - 0.092 # F1500W
+        # - 0.107 # F1280W
+        # - 0.137 # F1000W
+        # - 0.188 # F770W
+        # - 0.245 # F560W
+
+        - 0.319 # F444W
+
+        # - 0.395 # F356W
+        # - 0.512 # F277W
+        # - 0.707 # F200W
+        # - 0.932 # F150W
+
+        - 1.224 # F115W
+      reference_bin: 1 # F444W
+
+  model:
+    source:
+      light:
+        mean:
+          gaussian_01: # i0*Gauss
+            i0: ['delta', 1.0, 0.0]
+            # i0: ['log_normal', 1.0e-1, 2.0e-1]
+            center: ['normal', 0, 0.1]
+            covariance: ['log_normal', 0.08, 0.02]
+            off_diagonal: ['normal', 0.0, 10.0]
+          gaussian_02: # i0*Gauss
+            i0: ['delta', 1.0, 0.0]
+            # i0: ['log_normal', 2.0e-2, 4.0e-2]
+            center: ['normal', [-0.5, -0.5], 0.1]
+            covariance: ['log_normal', 0.08, 0.02]
+            off_diagonal: ['normal', 0.0, 10.0]
+
+        multifrequency:
+          nifty_mf:
+            zero_mode: [-4.0, 0.1] # normal
+            spatial_amplitude:
+              model: 'matern'
+              scale: [0.2, 0.03]        # lognormal
+              cutoff: [0.1, 0.01]       # lognormal
+              loglogslope: [-4.0, 0.1]  # normal
+            spectral_amplitude:
+              model: 'matern'
+              scale: [0.1, 0.03]        # lognormal
+              cutoff: [0.1, 0.01]       # lognormal
[…]
