### 2025-02-19 NIFTy-PPL/J-UBIK#2 @Eberle-Vincent
@prashjet Thanks for finding this error. We recently changed some conventions in the interface and forgot this demo script.
Should be solved by5f74f43acfa0bb9b8c4adfb99934df72a561e2ec
