### 2025-02-21 956cfc3
86-enforce-float64: enforce float64 precision in init.
