### 2025-02-21 f451433
86-enforce-float64: add note into README.md .
