### 2025-02-24 0a86e3c
resolve: diffuse building works
