+++ src/instruments/resolve/re/__init__.py
@@ -1,5 +1,5 @@
-
-from ....sky_model.resolve_sky import sky_model_diffuse, sky_model_points, sky_model
+from ....sky_model.resolve_sky import sky_model_diffuse, sky_model
+from ....sky_model.resolve_point_sources import resolve_point_sources
 from .response import InterferometryResponse, InterferometryResponseFinuFFT, InterferometryResponseDucc
 
 from .mosaicing import *
