+++ demos/configs/cygnusa_2ghz.cfg
@@ -19,7 +19,7 @@ stokesI diffuse space i0 flexibility stddev = 0.4
 stokesI diffuse space i0 asperity mean =  0.2
 stokesI diffuse space i0 asperity stddev = 0.2
 
-point sources mode = single
+point sources mode = fixed_locations
 point sources relative locations = 0deg$0deg,0.35as$-0.22as
 point sources a = 0.5
 point sources scale = 0.2

[…]
