### 2025-02-24 bcb5c5d
resolve: parse coordinate system change to classmethods
