+++ src/parse/wcs/coordinate_system.py
@@ -2,6 +2,8 @@ from dataclasses import dataclass
 from typing import Optional
 from enum import Enum
 
+from configparser import ConfigParser
+
 FRAME_KEY = 'frame'
 FRAME_DEFAULT = 'icrs'
 FRAME_EQUINOX_KEY = 'equinox'
@@ -13,6 +15,40 @@ class CoordinateSystemModel:
     ctypes: tuple[str, str]
     radesys: str
     equinox: Optional[float] = None
+
+    @classmethod
+    def from_yaml_dict(cls, grid_config: dict):
+        frame = grid_config.get(FRAME_KEY, FRAME_DEFAULT)
+        equinox = grid_config.get(FRAME_KEY)
+
+        coordinate_system = getattr(CoordinateSystems, frame)
+        _check_if_implemented(coordinate_system)
+
+        if (
+            (equinox is not None) and
+            (coordinate_system in [
+             CoordinateSystems.fk4, CoordinateSystems.f55])
+        ):
+            coordinate_system.equinox = equinox
+
+        return coordinate_system.value
+
+    @classmethod
+    def from_config_parser(cls, grid_config: ConfigParser):
+        frame = grid_config.get(FRAME_KEY, FRAME_DEFAULT)
+        equinox = grid_config.get(FRAME_KEY)
+
+        coordinate_system = getattr(CoordinateSystems, frame)
+        _check_if_implemented(coordinate_system)
+
+        if (
+            (equinox is not None) and
+            (coordinate_system in [
+             CoordinateSystems.fk4, CoordinateSystems.f55])
+        ):
+            coordinate_system.equinox = equinox
+
+        return coordinate_system.value
 
 
 class CoordinateSystems(Enum):
@@ -35,38 +71,6 @@ def _check_if_implemented(coordinate_system: str):
     if coordinate_system not in CoordinateSystems:
         raise ValueError(f"Unsupported coordinate system: {coordinate_system}."
                          f"Supported systems {[c for c in CoordinateSystems]}")
-
-
-def yaml_to_coordinate_system(grid_config: dict) -> CoordinateSystemModel:
-    frame = grid_config.get(FRAME_KEY, FRAME_DEFAULT)
-    equinox = grid_config.get(FRAME_KEY)
-
-    coordinate_system = getattr(CoordinateSystems, frame)
-    _check_if_implemented(coordinate_system)
-
-    if (
-        (equinox is not None) and
-        (coordinate_system in [CoordinateSystems.fk4, CoordinateSystems.f55])
-    ):
-        coordinate_system.equinox = equinox
-
-    return coordinate_system.value
-
-
-def cfg_to_coordinate_system(grid_config: dict) -> CoordinateSystemModel:
-    frame = grid_config.get(FRAME_KEY, FRAME_DEFAULT)
-    equinox = grid_config.get(FRAME_KEY)
-
-    coordinate_system = getattr(CoordinateSystems, frame)
-    _check_if_implemented(coordinate_system)
-
-    if (
-        (equinox is not None) and
-        (coordinate_system in [CoordinateSystems.fk4, CoordinateSystems.f55])
-    ):
-        coordinate_system.equinox = equinox
-
-    return coordinate_system.value
 
 
 def yaml_to_frame_name(grid_config: dict) -> str:
