### 2025-02-24 1ed634a
resolve: pytest coordinate system parsing
