### 2025-02-25 23745ce
parsing: check coordinate system
