### 2025-02-25 0e3e033
parser: (Re-)Introducing kl_kwargs.xtol = 1e-12
