### 2025-02-25 b9f4491
resolve: making the string
