### 2025-02-25 a3b739c
miniparser_fixes: fix miniparser, add constants for optimize_kl
