### 2025-02-27 e91274e
resolve: logging, no more color repr and docstring
