### 2025-02-28 b6a5d3c
jwst: build_filter_projector inside filter_projector
