### 2025-02-28 d78d094
docs: docstring and reformating
