### 2025-03-01 NIFTy-PPL/J-UBIK#4 @hamogu
Where do I find community guidelines?
The readme points out that this software will hopefully find a larger user community. Maybe the community can also contribute instruments models and wrappers for other observatories?
Unfortunately, I can't find a statement on community guidelines (which [JOSS also requires for all submissions](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines)). This can be simple, like a small section of a few sentences in the README along the lines of "Please open a GH issue here (link) if you run into trouble" and "we welcome contributions as Github pull request similar to here (link to any larger project or GH own pages that explain how PRs work)".

If you want, you can also write a long page for the docs, but that's not required. (And if that's in the docs already, I apologize - but see the next issue).
