### 2025-03-01 NIFTy-PPL/J-UBIK#5 @hamogu
Can I read the docs without installing the entire package locally?
I see that there is a `docs` folder but it seems to be entirely empty except for an overview of one sentence and a link to API docs (from https://github.com/NIFTy-PPL/J-UBIK/blob/main/docs/source/user/index.rst) that leads to a 404 error (see attached screenshot):

Where do I find the documentation for this package?
