### 2025-03-01 #7768 @hamogu
I'm trying to start my review. First, I want apologize that it has taken me so long to get started, I had COVID two weeks ago and was on vacation last week.

I'm looking for documentation beyond the readme and I can't find anything (see https://github.com/NIFTy-PPL/J-UBIK/issues/4 and https://github.com/NIFTy-PPL/J-UBIK/issues/5). Does that mean that I'm looking at the wrong location? Maybe it's all on readthedocs and there is just a link missing? Or is there no documentation beyond the readme and three example scripts?

In my opinion a software that is this heavy on dependencies (e.g. if I want to run a Chandra example, I need all of ciao, which is, incl the CALDB, > 1.5 GB in addition to JAX, which itself is binary and best executed on GPUs) absolutely needs to have readable documentation that I can look through and decide if this is the software I need before I go through a significant time of trying the get all the dependencies. It's also not exactly and easy installation process. To me, an easy process would be `pip install jubik[chandra,jwst'` (using `pyproject.toml` for specifying dependencies). I understand that the heavy dependencies and the additional paths that need to be set by hand don't make such simple specification feasible, but especially give that it's a bit of work to get everything up and running, I really need to know if it's worth the effort to do that.
