### 2025-03-02 NIFTy-PPL/J-UBIK#9 @hamogu
Pytest fails because it's can't find the test data
When running pytest from the project directory, I get this:

```Python
(J-UBIK) MoritzAirRoseGold ~/code/J-UBIK> pytest
=================================================== test session starts ===================================================
platform darwin -- Python 3.11.9, pytest-8.3.3, pluggy-1.5.0
rootdir: /Users/guenther/code/J-UBIK
configfile: pyproject.toml
plugins: jwst-1.17.1, anyio-4.7.0, asdf-4.1.0
collected 30 items

test/instruments/chandra/test_chandra_observation.py FF                                                             [  6%]
test/test_build_exposure_function.py .                                                                              [ 10%]
test/test_build_exposure_readout_function.py ....                                                                   [ 23%]
test/test_diagnostics.py ..                                                                                         [ 30%]
test/test_minimization_parser.py ................                                                                   [ 83%]
test/test_mock_data.py .                                                                                            [ 86%]
test/test_sky_model.py ..                                                                                           [ 93%]
test/test_slice_patches.py ..                                                                                       [100%]

======================================================== FAILURES =========================================================
___________________________________________________ test_initialization ___________________________________________________

sample_obs_info = {'aspect_sol': '../../chandra_test_data/primary/structure_only_asol.fits', 'bpix_file': '../../chandra_test_data/prima...ta_location': '../../chandra_test_data/', 'event_file': '../../chandra_test_data/primary/structure_only_evt.fits', ...}

    def test_initialization(sample_obs_info):
        npix_s = 1024
        npix_e = 100
        fov = 30.0
        elim = (0.5, 7.0)
        center = (10.684, 41.269)
        energy_ranges = None
        chips_off = (5, 7)

>       chandra_obs = ju.ChandraObservationInformation(
            obsInfo=sample_obs_info,
            npix_s=npix_s,
            npix_e=npix_e,
            fov=fov,
            elim=elim,
            center=center,
            energy_ranges=energy_ranges,
            chips_off=chips_off
        )

test/instruments/chandra/test_chandra_observation.py:30:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
src/instruments/chandra/chandra_observation.py:77: in __init__
    self.obsInfo['ra']       = float(rt.dmkeypar(infile=self.obsInfo['event_file'],keyword='RA_NOM',  echo=True))
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = <CIAO tool: dmkeypar>, args = ()
kwargs = {'echo': True, 'infile': '../../chandra_test_data/primary/structure_only_evt.fits', 'keyword': 'RA_NOM'}
parfile = '/var/folders/r7/f0qh27rn207bwwvt3s7v5gh40000gn/T/tmpenoda1fp.dmkeypar.par', stackfiles = {}, rval = 205
sout = "# dmkeypar (CIAO 4.17.0): ERROR: Could not open file '../../chandra_test_data/primary/structure_only_evt.fits'.\n\n"
sep = '\n  '

    def __call__(self, *args, **kwargs):

        # processing the argument list also validates them
        self._process_argument_list(args, kwargs)

        parfile = self._create_parfile_copy()
        stackfiles = {}
        try:
            stackfiles = self._update_parfile(parfile)
            self._display_command_line()
            _log_par_file_contents(parfile)
            (rval, sout) = self._run(parfile)
            _log_par_file_contents(parfile)

            if rval == 0:
                self.read_params(parfile)
                for pname in stackfiles:
                    oval = self._get_param_value(pname)
                    v5(f"Replacing {self._toolname}.{pname} = {oval}")
                    nval = stk.build(oval)
                    if len(nval) == 1:
                        nval = nval[0]

                    v5(f"  by {nval}")
                    self._set_param_value(pname, nval)

                txt = sout.rstrip()
                if txt == "":
                    retval = None
                else:
                    retval = CIAOPrintableString(txt)

            else:
                sep = "\n  "
                smsg = sep.join(sout.rstrip().split("\n"))
>               raise IOError(f"An error occurred while running '{self._toolname}':{sep}{smsg}")
E               OSError: An error occurred while running 'dmkeypar':
E                 # dmkeypar (CIAO 4.17.0): ERROR: Could not open file '../../chandra_test_data/primary/structure_only_evt.fits'.

../../mambaforge/envs/J-UBIK/lib/python3.11/site-packages/ciao_contrib/runtool.py:1863: OSError
______________________________________________________ test_get_data ______________________________________________________

sample_obs_info = {'aspect_sol': '../../chandra_test_data/primary/structure_only_asol.fits', 'bpix_file': '../../chandra_test_data/prima...ta_location': '../../chandra_test_data/', 'event_file': '../../chandra_test_data/primary/structure_only_evt.fits', ...}

    def test_get_data(sample_obs_info):
        npix_s = 1024
        npix_e = 100
        fov = 30.0
        elim = (0.5, 7.0)
        center = (10.684, 41.269)
        energy_ranges = None
        chips_off = (5, 7)

>       chandra_obs = ju.ChandraObservationInformation(
            obsInfo=sample_obs_info,
            npix_s=npix_s,
            npix_e=npix_e,
            fov=fov,
            elim=elim,
            center=center,
            energy_ranges=energy_ranges,
            chips_off=chips_off
        )

test/instruments/chandra/test_chandra_observation.py:64:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
src/instruments/chandra/chandra_observation.py:77: in __init__
    self.obsInfo['ra']       = float(rt.dmkeypar(infile=self.obsInfo['event_file'],keyword='RA_NOM',  echo=True))
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = <CIAO tool: dmkeypar>, args = ()
kwargs = {'echo': True, 'infile': '../../chandra_test_data/primary/structure_only_evt.fits', 'keyword': 'RA_NOM'}
parfile = '/var/folders/r7/f0qh27rn207bwwvt3s7v5gh40000gn/T/tmp1er8skf9.dmkeypar.par', stackfiles = {}, rval = 205
sout = "# dmkeypar (CIAO 4.17.0): ERROR: Could not open file '../../chandra_test_data/primary/structure_only_evt.fits'.\n\n"
sep = '\n  '

    def __call__(self, *args, **kwargs):

        # processing the argument list also validates them
        self._process_argument_list(args, kwargs)

        parfile = self._create_parfile_copy()
        stackfiles = {}
        try:
            stackfiles = self._update_parfile(parfile)
            self._display_command_line()
            _log_par_file_contents(parfile)
            (rval, sout) = self._run(parfile)
            _log_par_file_contents(parfile)

            if rval == 0:
                self.read_params(parfile)
                for pname in stackfiles:
                    oval = self._get_param_value(pname)
                    v5(f"Replacing {self._toolname}.{pname} = {oval}")
                    nval = stk.build(oval)
                    if len(nval) == 1:
                        nval = nval[0]

                    v5(f"  by {nval}")
                    self._set_param_value(pname, nval)

                txt = sout.rstrip()
                if txt == "":
                    retval = None
                else:
                    retval = CIAOPrintableString(txt)

            else:
                sep = "\n  "
                smsg = sep.join(sout.rstrip().split("\n"))
>               raise IOError(f"An error occurred while running '{self._toolname}':{sep}{smsg}")
E               OSError: An error occurred while running 'dmkeypar':
E                 # dmkeypar (CIAO 4.17.0): ERROR: Could not open file '../../chandra_test_data/primary/structure_only_evt.fits'.

../../mambaforge/envs/J-UBIK/lib/python3.11/site-packages/ciao_contrib/runtool.py:1863: OSError
================================================= short test summary info =================================================
FAILED test/instruments/chandra/test_chandra_observation.py::test_initialization - OSError: An error occurred while running 'dmkeypar':
FAILED test/instruments/chandra/test_chandra_observation.py::test_get_data - OSError: An error occurred while running 'dmkeypar':
============================================== 2 failed, 28 passed in 9.84s ===============================================
(J-UBIK) MoritzAirRoseGold ~/code/J-UBIK>
```

The root cause is that the location of the file to be loaded is hardcoded into the test relative to some directory, but this directory is neither the directory where the test is in, nor the root directory of the project where I'm running pytest; maybe the test files got moved around without updating the test?
