### 2025-03-02 945c4c5
Move Chandra test datafiles and set path for them using pytest.request
fixes #9
