### 2025-03-02 NIFTy-PPL/J-UBIK#11 @hamogu
Test failure because `energy_range is None`
Even after #10 (ensuring that all the test data files are found), I still see a failure in the test suite:
```Python
=================================================== test session starts ===================================================
platform darwin -- Python 3.11.9, pytest-8.3.3, pluggy-1.5.0
rootdir: /Users/guenther/code/J-UBIK
configfile: pyproject.toml
plugins: jwst-1.17.1, anyio-4.7.0, asdf-4.1.0
collected 2 items

test_chandra_observation.py .F
>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> captured stdout >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>

Observation Nr. 0 Information:
------------------------
Nominal pointing: RA = 225.9635, DEC = -41.8538, ROLL = 280.1516             degrees
Observation Duration = 68086.7870 s

Considered range in detector coordinates x = -58.2 - 2.8,            y = -35.7 - 25.3 pixels
Considered energy range = 0.5 to 7.0 keV
------------------------

Observation Nr. 0 Binning:
------------------------
Number of events considered = 0
------------------------
>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> traceback >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
AttributeError: 'NoneType' object has no attribute 'log'

sample_obs_info = {'aim_dec': -41.8537780709751, 'aim_ra': 225.96345533067, 'aspect_sol': '/Users/guenther/code/J-UBIK/test/instruments/chandra/chandra_test_data/primary/structure_only_asol.fits', 'asphist_res_xy': None, ...}
request = <FixtureRequest for <Function test_get_data>>

    def test_get_data(sample_obs_info, request):
        npix_s = 1024
        npix_e = 100
        fov = 30.0
        elim = (0.5, 7.0)
        center = (10.684, 41.269)
        energy_ranges = None
        chips_off = (5, 7)

        chandra_obs = ju.ChandraObservationInformation(
            obsInfo=sample_obs_info,
            npix_s=npix_s,
            npix_e=npix_e,
            fov=fov,
            elim=elim,
            center=center,
            energy_ranges=energy_ranges,
            chips_off=chips_off
        )

>       data_array = chandra_obs.get_data(
            f"{request.path.parent}/chandra_test_data/generated_data.fits"
        )

test_chandra_observation.py:79:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = <jubik0.instruments.chandra.chandra_observation.ChandraObservationInformation object at 0x157b30410>
outfile = '/Users/guenther/code/J-UBIK/test/instruments/chandra/chandra_test_data/generated_data.fits'

    def get_data(self, outfile):

        """
        Obtain the observed photon counts on a 3D grid over spatial and energy coordinates.

        Parameters:
        -----------
        outfile: string
            fits file to which CXC saves the filtered event list

        Returns:
        --------
        data: np.array
            event counts on a 3D grid

        """

        # filter w/ cxc: spatial and energy cuts
        # creates an event list w/ only those detections that make the cuts
        infile  = self.obsInfo['event_file']
        infile += '[x={0:.6f}:{1:.6f},'.format(self.obsInfo['x_min'], self.obsInfo['x_max'])
        infile += 'y={0:.6f}:{1:.6f},'.format(self.obsInfo['y_min'], self.obsInfo['y_max'])
        infile += 'energy={0:.4f}:{1:.4f}]'.format(1.e3*self.obsInfo['energy_min'], 1.e3*self.obsInfo['energy_max'])

        rt.dmcopy(infile=infile, outfile=outfile, clobber='yes')

        # bin events on a numpy array
        # see https://cxc.cfa.harvard.edu/ciao/data_products_guide/events.html for L2 event file columns
        with fits.open(outfile) as dat_filtered:
            evts = dat_filtered['EVENTS'].data
        evts = np.array([evts['x'], evts['y'], np.log(1.e-3*evts['energy'])])
        evts = evts.transpose()
        if self.obsInfo['energy_ranges']:
            bins = (self.obsInfo['npix_s'],  self.obsInfo['npix_s'], np.log(self.obsInfo['energy_ranges']))
        else:
            bins = (self.obsInfo['npix_s'],  self.obsInfo['npix_s'], self.obsInfo['npix_e'])
        ranges = ((self.obsInfo['x_min'],self.obsInfo['x_max']),
                  (self.obsInfo['y_min'],self.obsInfo['y_max']),
                  (np.log(self.obsInfo['energy_min']), np.log(self.obsInfo['energy_max'])))

        data, edges = np.histogramdd(evts, bins=bins, range=ranges, density=False, weights=None)
        data = data.transpose((1,0,2)).astype(int)
        self.obsInfo['ntot_binned'] = np.sum(data)

        message_binning(self.obsInfo)
>       energy_ranges = np.log(self.obsInfo['energy_ranges'])
E       TypeError: loop of ufunc does not support argument 0 of type NoneType which has no callable log method

../../../src/instruments/chandra/chandra_observation.py:198: TypeError
>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> entering PDB >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>

>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> PDB post_mortem (IO-capturing turned off) >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
> /Users/guenther/code/J-UBIK/src/instruments/chandra/chandra_observation.py(198)get_data()
-> energy_ranges = np.log(self.obsInfo['energy_ranges'])
(Pdb) args
self = <jubik0.instruments.chandra.chandra_observation.ChandraObservationInformation object at 0x157b30410>
outfile = '/Users/guenther/code/J-UBIK/test/instruments/chandra/chandra_test_data/generated_data.fits'
(Pdb) q

================================================= short test summary info =================================================
FAILED test_chandra_observation.py::test_get_data - TypeError: loop of ufunc does not support argument 0 of type NoneType which has no callable log method
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! _pytest.outcomes.Exit: Quitting debugger !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
============================================== 1 failed, 1 passed in 34.99s ===============================================

```
