### 2025-03-02 #7768 @hamogu
Just to be clear: The almost complete lack of presentation and the existing failures of a test suite (https://github.com/NIFTy-PPL/J-UBIK/issues/9, https://github.com/NIFTy-PPL/J-UBIK/issues/11) that is already thin (e.g. it doesn't test eROSTITA or JWST and don't actually run through a real inference as far as I can see) make this package non-compliant with JOSS requirements in my view. I don't think that solving those issues is hard, but it will require some time commitment.
