### 2025-03-02 #7768 @hamogu
Note that my review is not complete: [JOSS guidelines ask me to "verify the core functionality of the software."](https://joss.readthedocs.io/en/latest/review_criteria.html#functionality) but the lack of documentation has made it impossible for me to write any reasonable input yaml file for any of my own examples; all I can do is to try to run the demos, but to me "core functionality" means that users should be able to apply the software to their own datasets as well.
I'll continue when I know how to do that.
