### 2025-03-03 002d054
jwst: demo always 64bit
