+++ demos/configs/spt0418.yaml
@@ -303,7 +303,7 @@ telescope:
       #     rotation: ['delta', -1.95823434e-05, 0]
 
 files:
-  res_dir: './results/jwst_lens/cluster/test_04_threedata'
+  res_dir: './results/jwst_lens/cluster/test_06_resolve'
 
   pretrain_path: './results/jwst_lens/mfreworked/all_data_13_HybridMassFromStart/config.yaml'
 
@@ -396,8 +396,8 @@ minimization:
 
     samples:
       switches: [0, 3, 15, 30]
-      # n_samples: [3]
-      n_samples: [5]
+      n_samples: [3]
+      # n_samples: [5]
       mode: [linear_resample, nonlinear_resample, nonlinear_resample, nonlinear_update]
       # mode: [linear_resample]
       # Possible modes:
