### 2025-03-03 NIFTy-PPL/J-UBIK#14 @prashjet
Documentation
`J-UBIK` needs more comprehensive documentation. Here is my thoughts on why that's the case, and some suggestions on how to do that.

### What documentation is already present?

Currently, the example script headers provide the main information on how to use `J-UBIK`. The brief descriptions provided in these headers don't answer some basic questions about what the configuration settings settings mean. To give one example, in `jwst_demo.py`, for the `priors: diffuse: spatial: offset` settings, I don't know what this offset is referring to. Is this an offset of the diffuse component from the point sources? An offset of the diffuse component from the image center? What units should these take?

The header of `chandra_demo.py` points [here](https://doi.org/10.1051/0004-6361/202347750) and [here](https://doi.org/10.21105/joss.06593) for information on adjusting the  `J-UBIK` configuration options, but it's not obvious to me where in these sources I should look to find relevant information. The main functionality of `J-UBIK` should be documented in its own, standalone, documentation. This very well could involve repeating some information which can be found in other sources and/or linking to _specific_ parts of external sources where appropriate.

As it stands, there is insufficient detail for me to understand what `J-UBIK` does and how I can use it.

### What should the documentation contain?

The [JOSS review checklist](https://github.com/openjournals/joss-reviews/issues/7768#issuecomment-2665761774) provides some general guidelines for what to include. In addition to this, given the scope of `J-UBIK` as described in the JOSS paper, I think there are three specific topics which ought to be comprehensively described and clearly sign-posted:

1. What is the underlying probabilistic model?
   - the structure of the prior model and specific choices of probability distributions should be explicitly defined
   - use variable names consistent with the `priors` section of the configuration file
   - ideally, the effect of changing the prior settings should be demonstrated e.g. using images sampled from the generative prior
2. Inference algorithms
   - which algorithms are available?
   - what are their pros and cons?
   - how do I interpret the output?
3. a guide for how to work with different instruments/telescopes

This information should be decoupled from the example scripts so it is easy to locate. As far as possible, I would like to be able to run simple examples which can illustrate the above points without installing large external dependencies. Images sampled from the generative model could be useful here instead of real datasets with large associated dependencies.

The current example scripts could then be used not to reiterate the meaning of the configuration settings, but rather to discuss any details which are particularly relevant for JWST/Chandra/eRosita data. Ideally, this would be formatted as a notebook rendered within the documentation itself, allowing for more discursive text alongside the code. At the very least, the output plots of the example scripts should be adequately labelled and explained.

### Where should the documentation live?

There is already a template for `sphinx` documentation in `J-UBIK`, containing a "User Guide" section with the following header:

> This guide is an overview and explains the main conceptual idea behind J-UBIK (JAX-accelerated Universal Bayesian Imaging Kit for X-ray).

This would be the simplest and most convenient place to put more documentation. It should be structured into relevant sections, possibly containing code snippets and/or notebooks to illustrate the software. You can then render the documentation (either as a pdf or html) and the rendered version should clearly accessible from the Github repository. Ideally, you would have an html version hosted online (e.g. at [readthedocs](https://about.readthedocs.com/) with a link to this documentation website in the `README` file.
