### 2025-03-03 #7768 @prashjet
I think `J-UBIK` certainly ticks the box for "substantial scholarly effort" required for JOSS, however it needs much more comprehensive documentation in order for me and other potential users to understand what it does and test its functionality. Some more detailed discussion is [here](https://github.com/NIFTy-PPL/J-UBIK/issues/14).
