### 2025-03-04 e66ae24
resolve: parametric prior can return a model
