### 2025-03-04 a5cfb22
resolve: introduce some mf models
