### 2025-03-04 ba8218b
resolve: making resolve demo work
