+++ src/parse/instruments/resolve/data/data_modify/__init__.py
@@ -1,0 +1,1 @@
+from .modify_observation import ObservationModify, WeightModify
