### 2025-03-05 #7768 @ivastar
@prashjet, @hamogu given that you have both encountered issues with documentation and trying to run examples, maybe we can pass the ball back to the authors at this point?

@wmarg could you please review the comments above and address them? Since they are a bit more substantial, rather than setting a timeline on my end, please let me know what would be reasonable time for you to address them.
