### 2025-03-06 3541512
parsing: rewrite to new model setup
