+++ src/instruments/resolve/multimessanger/radio_sky_extractor.py
@@ -1,0 +1,185 @@
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+# Copyright(C) 2019-2020 Max-Planck-Society
+# Author: Julian Rüstig
+
+from ....grid import Grid
+from ....color import ColorRanges
+from ..constants import RESOLVE_SKY_UNIT
+
+import nifty8.re as jft
+
+import jax
+import jax.numpy as jnp
+
+import astropy.units as u
+
+from typing import Union
+
+
+def build_extract_sky(sky_key: str | None):
+    if sky_key is None:
+        return lambda x: x
+    return lambda x: x[sky_key]
+
+
+def build_radio_slicing(index_of_last_radio_bin: int | None):
+    '''Build a Grid corresponding to the radio part of the sky.
+
+    ***Warning: The radio part of  the sky is assumed to be in the first
+    indices of the sky.***
+
+    Parameters
+    ----------
+    index_of_last_radio_bin: int | None
+        The index of the last radio bin. We assume that the radio sky is in the
+        first part of sky frequencies.
+    '''
+
+    # TODO: Make the radio sclicing independent from the assumption that the
+    # first indices correspond to the radio part of the sky.
+
+    if index_of_last_radio_bin is None:
+        return lambda x: x
+
+    def radio_sky_extractor(tree):
+        return jax.tree.map(lambda x: x[:index_of_last_radio_bin+1], tree)
+
+    return radio_sky_extractor
+
+
+def build_unit_conversion(sky_unit: u.Unit | None):
+    if sky_unit is None:
+        return lambda x: x
+
+    conversion_factor = sky_unit.to(RESOLVE_SKY_UNIT)
+
+    def unit_conversion(tree):
+        return jax.tree.map(lambda x: x*conversion_factor, tree)
+
+    return unit_conversion
+
+
+def resolve_transpose(tree):
+    return jax.tree.map(lambda x: jnp.transpose(x, (0, 1, 2, 4, 3)), tree)
+
+
+def build_radiofy_sky(target_shape):
+    if len(target_shape) == 2:
+        return lambda tree: jax.tree.map(lambda x: x[None, None, None], tree)
+
+    elif len(target_shape) == 3:
+        return lambda tree: jax.tree.map(lambda x: x[None, None], tree)
+
+    elif len(target_shape) == 4:
+        return lambda tree: jax.tree.map(lambda x: x[None], tree)
+
+    elif len(target_shape) == 5:
+        return lambda tree: jax.tree.map(lambda x: x, tree)
+
+    else:
+        raise ValueError(f'Shape: {target_shape} is not compatible with '
+                         'radio sky.')
+
+
+class RadioSkyExtractor(jft.Model):
+    def __init__(
+        self,
+        domain: Union[jft.ShapeWithDtype, dict],
+        extract_sky: callable,
+        slice_radio_sky: callable,
+        unit_conversion: callable,
+        radiofy: callable,
+        transpose: callable,
+    ):
+        self.extract = extract_sky
+        self.slicing = slice_radio_sky
+        self.conv = unit_conversion
+        self.radiofy = radiofy
+        self.trans = transpose
+        super().__init__(domain=domain)
+
[…]
