### 2025-03-06 150c514
resolve: restructure resolve paths
