### 2025-03-06 32b34d3
resolve: input mosaicing, and likelihood building
