+++ src/instruments/resolve/likelihood/__init__.py
@@ -15,8 +15,4 @@
 # Author: Julian Rüstig
 
 
-from ....sky_model.resolve_sky import sky_model_diffuse, sky_model
-from ....sky_model.resolve_point_sources import resolve_point_sources
-from .response import InterferometryResponse, InterferometryResponseFinuFFT, InterferometryResponseDucc
-
-from .mosaicing import *
+from .radio_likelihood import build_radio_likelihood
