### 2025-03-07 bbc02df
jwst: remove not needed subsample_centers
