### 2025-03-07 27d37dd
parsing: Bugfix in spatial-model
