### 2025-03-07 71a86a0
jwst: test subsample centers
