### 2025-03-07 2014be1
jwst: test nufft interpolation
