### 2025-03-07 80eb1d5
resolve: mosaicing is in main resolve
