### 2025-03-08 ad4e06d
resolve: simplify mosaic likelihood building
