### 2025-03-08 634da74
resolve: casting to dtype32 is possible
