### 2025-03-08 c396b32
resolve: introduce strong typing, for parsing
