### 2025-03-08 c823332
resolve: not testing the jax finufft, since not dependency
