### 2025-03-08 f3fa40c
jwst: delete pretrain
