### 2025-03-08 824d15a
parsing: introducing atomic and hybrid sky-model
