### 2025-03-09 723a828
resolve: more robust SkyCenter setup
