+++ src/instruments/resolve/mosaicing/sky_beamer.py
@@ -155,10 +155,8 @@ def build_jft_sky_beamer(
         o_phase_center = SkyCoord(direction.phase_center[0]*u.rad,
                                   direction.phase_center[1]*u.rad,
                                   frame=sky_center.frame)
-        r = sky_center.separation(o_phase_center)
-        phi = sky_center.position_angle(o_phase_center)
-        center_y = r.to(u.rad).value * np.cos(phi.to(u.rad).value)
-        center_x = r.to(u.rad).value * np.sin(phi.to(u.rad).value)
+        center_x, center_y = calculate_phase_offset_to_image_center(
+            sky_center, o_phase_center)
 
         x = sky_coords.separation(o_phase_center)
         x = x.to(u.rad).value
@@ -186,6 +184,28 @@ def build_jft_sky_beamer(
         )
 
     return SkyBeamerJft(sky_shape_with_dtype, beam_directions)
+
+
+def calculate_phase_offset_to_image_center(
+    sky_center: SkyCoord,
+    phase_center: SkyCoord,
+):
+    '''Calculate the relative shift of the phase center to the sky center
+    (reconstruction center) in radians.
+
+    Parameters
+    ----------
+    sky_center: astropy.SkyCoord
+        The world coordinate of the sky center.
+    phase_center: astropy.SkyCoord
+        The world coordinate of the phase center of the observation.
+    '''
+    r = sky_center.separation(phase_center)
+    phi = sky_center.position_angle(phase_center)
+    # FIXME: center (x, y) switch maybe because of the ducc0 fft?
+    center_y = r.to(u.rad).value * np.cos(phi.to(u.rad).value)
+    center_x = r.to(u.rad).value * np.sin(phi.to(u.rad).value)
+    return center_x, center_y
 
 
 def _get_mean_frequency(ff, n_freq_bins, f_binbounds, observation):
