### 2025-03-09 0e6d5ab
resolve: init restrict and average stokes i
