### 2025-03-11 24abf25
Final paper plotting
