### 2025-03-11 NIFTy-PPL/J-UBIK#5 @Eberle-Vincent
@hamogu ,Thanks for pointing this out. Actually, the API is supposed to be built by the CI. The reason it is not is due to a recent change in sphinx which resolves the symlink and therefore prevents the API from being built properly.

We have fixed this by setting the maximum version of sphinx for the time being. In the long run, sphinx will hopefully allow the use of symlinks again.

You can now see the API again here:
https://ift.pages.mpcdf.de/j-ubik/user/index.html
