### 2025-03-11 NIFTy-PPL/J-UBIK#5 @hamogu
I would suggest to put a link to that into the README.

Github also has a "About" (in your repro, click on the little cogwheel in the top right, of course you can't see that in my screenshot because I don't have admin rights to your repro) for each project where you can set the name, topic (that will help when someone searches on GitHub, e.g. "astronomy" "baysian analysis", whatever you like), and a "homepage". Many projects point to their docs as "homepage".
