### 2025-03-11 09fb401
Joss paper plotting
