+++ demos/configs/eROSITA_demo.yaml
@@ -20,6 +20,7 @@ files:
 
 telescope:
   tm_ids : [1, 2] #, 3, 4, 6]
+  pointing_center: [83.77602098929445, -69.31087419594445]
   rebin: 80 # typical value used in analysis; equals 4 arcsec resolution
   fov: 2048
   pattern: 15
@@ -44,7 +45,7 @@ plotting:
   dpi: 300
 
 grid:
-  sdim: 512
+  sdim:
   edim: 3 # FIXME: remove this parameter
   s_padding_ratio: 1.1075 # leads to a 1134, which is a "good size" padded space
   e_padding_ratio: 1.
