+++ src/instruments/resolve/multimessanger/__init__.py
@@ -1,2 +1,3 @@
 from .radio_sky_extractor import (
-    RadioSkyExtractor, build_radio_sky_extractor, build_radio_grid)
+    RadioSkyExtractor, build_radio_sky_extractor, build_radio_grid,
+    build_radiofy_sky)
