### 2025-03-17 a908eb8
jwst: correct index boundaries, explicit indexing, No interpolation-transpose
