### 2025-03-17 c704516
jwst: testing reconstruction shapes
