+++ demos/configs/spt0418.yaml
@@ -18,11 +18,12 @@ no_interactive_plotting: False
 
 sky: 
   grid:
-    sdim: 384
+    sdim: 512
     # NOTE: SHOULD THIS BE HERE?????
     s_padding_ratio: 1.5  # Size of the convergence, i.e. (384*1.5, 6*1.5)
 
-    fov: 6.0arcsec
+    # fov: 6.0arcsec
+    fov: 5.0arcsec
     rotation: 0.0deg
     coordinate_frame: icrs
     sky_center:
@@ -37,41 +38,41 @@ sky:
     energy_unit: eV
     energy_bin:
       e_min:
-        # - 0.054 # F2100W
+        - 0.054 # F2100W
 
         - 0.068 # F1800W  # mistake here actually 0.064
 
-        # - 0.075 # F1500W
-        # - 0.093 # F1280W  # mistake here actually 0.088
-        # - 0.114 # F1000W
-        # - 0.143 # F770W
-        # - 0.201 # F560W
+        - 0.075 # F1500W
+        - 0.093 # F1280W  # mistake here actually 0.088
+        - 0.114 # F1000W
+        - 0.143 # F770W
+        - 0.201 # F560W
 
         - 0.249 # F444W
 
-        # - 0.319 # F356W  # mistake here actually 0.311
-        # - 0.396 # F277W
-        # - 0.557 # F200W
-        # - 0.743 # F150W
+        - 0.319 # F356W  # mistake here actually 0.311
+        - 0.396 # F277W
+        - 0.557 # F200W
+        - 0.743 # F150W
 
         - 0.967 # F115W
       e_max:
-        # - 0.067 # F2100W
+        - 0.067 # F2100W
 
         - 0.075 # F1800W
 
-        # - 0.092 # F1500W
-        # - 0.107 # F1280W
-        # - 0.137 # F1000W
-        # - 0.188 # F770W
-        # - 0.245 # F560W
+        - 0.092 # F1500W
+        - 0.107 # F1280W
+        - 0.137 # F1000W
+        - 0.188 # F770W
+        - 0.245 # F560W
 
         - 0.319 # F444W
 
-        # - 0.395 # F356W
-        # - 0.512 # F277W
-        # - 0.707 # F200W
-        # - 0.932 # F150W
+        - 0.395 # F356W
+        - 0.512 # F277W
+        - 0.707 # F200W
+        - 0.932 # F150W
 
         - 1.224 # F115W
       reference_bin: 1 # F444W
@@ -323,63 +324,63 @@ files:
 
     f1800w:
       - './data/spt0418/f1800w/jw01355015001_13101_00001_mirimage/jw01355015001_13101_00001_mirimage_cal.fits'
-      - './data/spt0418/f1800w/jw01355015001_13101_00002_mirimage/jw01355015001_13101_00002_mirimage_cal.fits'
-      - './data/spt0418/f1800w/jw01355015001_13101_00003_mirimage/jw01355015001_13101_00003_mirimage_cal.fits'
-      - './data/spt0418/f1800w/jw01355015001_13101_00004_mirimage/jw01355015001_13101_00004_mirimage_cal.fits'
+      # - './data/spt0418/f1800w/jw01355015001_13101_00002_mirimage/jw01355015001_13101_00002_mirimage_cal.fits'
+      # - './data/spt0418/f1800w/jw01355015001_13101_00003_mirimage/jw01355015001_13101_00003_mirimage_cal.fits'
+      # - './data/spt0418/f1800w/jw01355015001_13101_00004_mirimage/jw01355015001_13101_00004_mirimage_cal.fits'
 
[…]
