+++ demos/configs/spt0418.yaml
@@ -417,5 +417,5 @@ minimization:
     kl_minimization:
       switches: [0, 10, 15]
       # kl_absdelta: [Null, 1.0e-2, 1.0e-4]
-      kl_xtol: 1e-12
+      kl_xtol: 1.0e-12
       kl_maxiter: [15, 20, 25]
