### 2025-03-17 55089bc
jwst: deleting probably unused index_grid & relative_coordinates
