### 2025-03-17 b3d5db6
jwst: resume without nonlinear resample
