+++ demos/configs/spt0418.yaml
@@ -385,7 +385,7 @@ files:
 
 minimization:
     key: 20250225
-    resume: False
+    resume: True
     n_total_iterations: 25
 
     delta:
@@ -393,7 +393,7 @@ minimization:
       values: [5.e-6, 1.e-6, 1.e-7]
 
     samples:
-      switches: [0, 3, 15, 30]
+      switches: [0, 15, 30]
       n_samples: [3]
       # n_samples: [5]
       mode: [linear_resample, nonlinear_resample, nonlinear_resample, nonlinear_update]
