### 2025-03-17 074639b
resolve: testing with all freqs
