### 2025-03-17 ee55ca8
jwst: make finufft interpolation possible
