### 2025-03-17 6ba3ddc
resolve: nicer demo, SkyDomain not needed anymore
