+++ src/instruments/resolve/data/observation.py
@@ -30,7 +30,8 @@ class BaseObservation:
     @property
     def _dom(self):
         pol_dom = self.polarization.space
-        dom = [pol_dom] + [ift.UnstructuredDomain(ss) for ss in self._vis.shape[1:]]
+        dom = [pol_dom] + \
+            [ift.UnstructuredDomain(ss) for ss in self._vis.shape[1:]]
         return ift.makeDomain(dom)
 
     @property
@@ -272,7 +273,8 @@ class SingleDishObservation(BaseObservation):
     def load(file_name):
         dct = dict(np.load(file_name))
         pol = Polarization.from_list(dct["polarization"])
-        pointings = Directions.from_list([dct["pointings0"], dct["pointings1"]])
+        pointings = Directions.from_list(
+            [dct["pointings0"], dct["pointings1"]])
         return SingleDishObservation(
             pointings, dct["vis"], dct["weight"], pol, dct["freq"]
         )
@@ -330,7 +332,8 @@ class Observation(BaseObservation):
         my_assert(np.all(np.isfinite(weight)))
         my_assert(np.all(np.diff(freq)))
 
-        my_assert(not (is_single_precision(vis.dtype) ^ is_single_precision(weight.dtype)))
+        my_assert(not (is_single_precision(vis.dtype)
+                  ^ is_single_precision(weight.dtype)))
 
         vis.flags.writeable = False
         weight.flags.writeable = False
@@ -453,34 +456,42 @@ class Observation(BaseObservation):
         if comm is None:
             local_imaging_bands = range(n_imaging_bands)
         else:
-            lo, hi = ift.utilities.shareRange( n_imaging_bands, comm.Get_size(), comm.Get_rank())
+            lo, hi = ift.utilities.shareRange(
+                n_imaging_bands, comm.Get_size(), comm.Get_rank())
             local_imaging_bands = range(lo, hi)
         full_obs = Observation.load(file_name)
         obs_list = []
         for ii in local_imaging_bands:
-            slc = slice(*ift.utilities.shareRange(len(global_freqs), n_imaging_bands, ii))
+            slc = slice(
+                *ift.utilities.shareRange(len(global_freqs), n_imaging_bands, ii))
             obs_list.append(full_obs.get_freqs_by_slice(slc))
         nu0 = global_freqs.mean()
         return obs_list, nu0
 
     def to_double_precision(self):
         return Observation(self._antpos,
-                           self._vis.astype(np.complex128, casting="same_kind", copy=False),
-                           self._weight.astype(np.float64, casting="same_kind", copy=False),
+                           self._vis.astype(
+                               np.complex128, casting="same_kind", copy=False),
+                           self._weight.astype(
+                               np.float64, casting="same_kind", copy=False),
                            self._polarization, self._freq, self._auxiliary_tables)
 
     def to_single_precision(self):
         return Observation(self._antpos,
-                           self._vis.astype(np.complex64, casting="same_kind", copy=False),
-                           self._weight.astype(np.float32, casting="same_kind", copy=False),
+                           self._vis.astype(
+                               np.complex64, casting="same_kind", copy=False),
+                           self._weight.astype(
+                               np.float32, casting="same_kind", copy=False),
                            self._polarization, self._freq, self._auxiliary_tables)
 
     def is_single_precision(self):
-        assert not (is_single_precision(self._weight.dtype) ^ is_single_precision(self._vis.dtype))
[…]
