+++ src/parse/instruments/resolve/response.py
@@ -8,54 +8,12 @@ from typing import Optional, Union
 from configparser import ConfigParser
 
 
-@dataclass
-class SkyDomain:
-    # Polarization
-    polarization_labels: list[str]
-    # Time
-    times: list[float]
-    # Frequencies
-    frequencies: list[float]
-    # Spatial
-    npix_x: int
-    npix_y: int
-    pixsize_x: float
-    pixsize_y: float
-    # Optional
-    center_x: Optional[float] = 0.
-    center_y: Optional[float] = 0.
-
-
-def sky_domain_from_grid(
-    grid: Grid,
-    center: Optional[tuple[float]] = [0., 0.]
-) -> SkyDomain:
-    SPECTRAL_UNIT = u.Hz
-    SPATIAL_UNIT = u.rad
-
-    shape = grid.spatial.shape
-    distances = grid.spatial.distances_in(SPATIAL_UNIT)
-    frequencies = grid.spectral.binbounds_in(SPECTRAL_UNIT)
-
-    return SkyDomain(
-        polarization_labels=grid.polarization_labels,
-        times=grid.times,
-        frequencies=frequencies,
-        npix_x=shape[0],
-        npix_y=shape[1],
-        pixsize_x=distances[0],
-        pixsize_y=distances[1],
-        center_x=center[0],
-        center_y=center[1],
-    )
-
-
-BACKEND_KEY = 'backend'
 FINUFFT_KEYS = ['finufft']
 DUCC_KEYS = ['ducc', 'ducc0']
-EPSILON_KEY = 'epsilon'
-POLARIZATION_KEY = 'no_polarization'
 
+EPSILON_KEY = 'epsilon'
+
+BACKEND_KEY = 'backend'
 DO_WGRIDDING_KEY = 'do_wgridding'
 NTHREADS_KEY = 'nthreads'
 VERBOSITY_KEY = 'verbosity'
@@ -70,6 +28,15 @@ class Ducc0Settings:
 
     @classmethod
     def from_yaml_dict(cls, yaml_dict: dict):
+        f'''Read ducc0 settings from yaml_dict.
+
+        Parameters
+        ----------
+        {EPSILON_KEY}: float
+        {DO_WGRIDDING_KEY}: bool
+        {NTHREADS_KEY}: int
+        {VERBOSITY_KEY}: bool
+        '''
         epsilon = yaml_dict[EPSILON_KEY]
         do_wgridding = yaml_dict[DO_WGRIDDING_KEY]
         nthreads = yaml_dict[NTHREADS_KEY]
@@ -83,7 +50,25 @@ class Ducc0Settings:
 
     @classmethod
     def from_config_parser(cls, config: ConfigParser):
-        raise NotImplementedError
+        f'''Read ducc0 settings from yaml_dict.
+
+        Parameters
+        ----------
+        {EPSILON_KEY}: float
+        {DO_WGRIDDING_KEY}: bool | None
+        {NTHREADS_KEY}: int | None
+        {VERBOSITY_KEY}: bool | None
+        '''
+        epsilon = eval(config[EPSILON_KEY])
+        do_wgridding = eval(config.get(DO_WGRIDDING_KEY, False))
+        nthreads = eval(config.get(NTHREADS_KEY, 1))
+        verbosity = eval(config.get(VERBOSITY_KEY, False))
+        return Ducc0Settings(
+            epsilon=epsilon,
+            do_wgridding=do_wgridding,
+            nthreads=nthreads,
+            verbosity=verbosity,
+        )
 
 
 @dataclass
@@ -93,23 +78,41 @@ class FinufftSettings:
 
     @classmethod
     def from_yaml_dict(cls, yaml_dict: dict):
[…]
