+++ src/instruments/resolve/data/data_modify/__init__.py
@@ -1,0 +1,2 @@
+from .modify_observation import modify_observation
+from .restrict_and_average_to_stokesi import restrict_and_average_to_stokes_i
