+++ src/utils.py
@@ -167,6 +167,15 @@ def copy_config(filename, path_to_yaml_file=None,
         print(f"Config file saved to: {join(output_dir, filename)}.")
 
 
+def save_config_copy_easy(path_to_file: str, path_to_save_file: str):
+    from shutil import copy, SameFileError
+    try:
+        copy(path_to_file, path_to_save_file)
+        print(f"Config file saved to: {path_to_save_file}.")
+    except SameFileError:
+        pass
+
+
 def create_output_directory(directory_name):
     """Convenience function to create directories.
     Parameters

[…]
