### 2025-03-18 5821540
jwst: setup of linear and finufft interpolation with 'xy' indexing
