+++ test/instruments/jwst/rotation_and_shift/test_linear.py
@@ -2,38 +2,61 @@ import numpy as np
 from jubik0.instruments.jwst.rotation_and_shift.linear_rotation_and_shift import build_linear_rotation_and_shift
 
 
-def test_field_sameaxis():
+def test_field_sameaxis_ij():
     maxx, maxy = 256, 256
     field = np.zeros((maxx, maxy))
     field[100:150, 100:150] = 1
     field[200:250, 100:150] = 1
 
-    xx, yy = np.array(np.meshgrid(np.arange(0, maxx, 1),
-                                  np.arange(0, maxy, 1),
-                                  indexing='ij')
-                      )
+    xy = np.array(np.meshgrid(np.arange(0, maxx, 1),
+                              np.arange(0, maxy, 1),
+                              indexing='ij'))
     rs = build_linear_rotation_and_shift(1, 1)
-    field_mapped = rs(field, np.array((xx, yy)))
+    field_mapped = rs(field, xy)
 
     assert np.allclose(field, field_mapped, atol=1e-5)
 
 
-def test_field_differentaxis():
+def test_field_sameaxis_xy():
+    maxx, maxy = 256, 256
+    field = np.zeros((maxx, maxy))
+    field[100:150, 100:150] = 1
+    field[200:250, 100:150] = 1
+
+    xy = np.array(np.meshgrid(np.arange(0, maxx, 1),
+                              np.arange(0, maxy, 1),
+                              indexing='xy'))
+    rs = build_linear_rotation_and_shift(1, 1, indexing='xy')
+    field_mapped = rs(field, xy)
+
+    assert np.allclose(field, field_mapped, atol=1e-5)
+
+
+def test_field_differentaxis_ij():
     maxx, maxy = 256, 325
     field = np.zeros((maxx, maxy))
     field[100:150, 100:150] = 1
     field[200:250, 100:150] = 1
 
-    xx, yy = np.array(np.meshgrid(np.arange(0, maxx, 1),
-                                  np.arange(0, maxy, 1),
-                                  indexing='ij'))
+    xy = np.array(np.meshgrid(np.arange(0, maxx, 1),
+                              np.arange(0, maxy, 1),
+                              indexing='ij'))
     rs = build_linear_rotation_and_shift(1, 1)
-    field_mapped = rs(field, np.array((xx, yy)))
+    field_mapped = rs(field, xy)
     assert np.allclose(field, field_mapped, atol=1e-4)
 
-    xx, yy = np.array(np.meshgrid(np.arange(0, maxx, 1),
-                                  np.arange(0, maxy, 1),
-                                  indexing='xy'))
-    rs = build_linear_rotation_and_shift(1, 1)
-    field_mapped = rs(field.T, np.array((yy, xx))).T
+
+def test_field_differentaxis_xy():
+    maxx, maxy = 256, 325
+    field = np.zeros((maxx, maxy))
+    field[100:150, 100:150] = 1
+    field[200:250, 100:150] = 1
+
+    xy = np.array(np.meshgrid(np.arange(0, maxx, 1),
+                              np.arange(0, maxy, 1),
+                              indexing='xy'))
+    rs = build_linear_rotation_and_shift(1, 1, indexing='xy')
+    # field_mapped = rs(field.T, np.array((yy, xx))).T
+    field_mapped = rs(field, xy)
+
     assert np.allclose(field, field_mapped, atol=1e-4)
