+++ demos/configs/spt0418.yaml
@@ -1,7 +1,7 @@
 cpu: False
 prior_samples: False
 plot_results: True
-nonparametric_lens: True
+nonparametric_lens: False
 max_residuals: 4 # 4 for big run
 no_interactive_plotting: False
 
@@ -39,44 +39,44 @@ sky:
     energy_unit: eV
     energy_bin:
       e_min:
-        - 0.054 # F2100W
+        # - 0.054 # F2100W
 
         - 0.068 # F1800W  # mistake here actually 0.064
-
-        - 0.075 # F1500W
-        - 0.093 # F1280W  # mistake here actually 0.088
-        - 0.114 # F1000W
-        - 0.143 # F770W
-        - 0.201 # F560W
+        #
+        # - 0.075 # F1500W
+        # - 0.093 # F1280W  # mistake here actually 0.088
+        # - 0.114 # F1000W
+        # - 0.143 # F770W
+        # - 0.201 # F560W
 
         - 0.249 # F444W
 
-        - 0.319 # F356W  # mistake here actually 0.311
-        - 0.396 # F277W
-        - 0.557 # F200W
-        - 0.743 # F150W
+        # - 0.319 # F356W  # mistake here actually 0.311
+        # - 0.396 # F277W
+        # - 0.557 # F200W
+        # - 0.743 # F150W
 
         - 0.967 # F115W
       e_max:
-        - 0.067 # F2100W
+        # - 0.067 # F2100W
 
         - 0.075 # F1800W
 
-        - 0.092 # F1500W
-        - 0.107 # F1280W
-        - 0.137 # F1000W
-        - 0.188 # F770W
-        - 0.245 # F560W
+        # - 0.092 # F1500W
+        # - 0.107 # F1280W
+        # - 0.137 # F1000W
+        # - 0.188 # F770W
+        # - 0.245 # F560W
 
         - 0.319 # F444W
 
-        - 0.395 # F356W
-        - 0.512 # F277W
-        - 0.707 # F200W
-        - 0.932 # F150W
+        # - 0.395 # F356W
+        # - 0.512 # F277W
+        # - 0.707 # F200W
+        # - 0.932 # F150W
 
         - 1.224 # F115W
-      reference_bin: 7 # F444W
+      reference_bin: 1 # F444W
 
   model:
     source:
@@ -305,23 +305,23 @@ telescope:
       #     rotation: ['delta', -1.95823434e-05, 0]
 
 files:
-  res_dir: './results/jwst_lens/cluster/test_09_nonparametriclens'
+  res_dir: './results/jwst_lens/cluster/test_10_3data'
 
   filter:
 
-    f2100w:
-      - './data/spt0418/f2100w/jw01355015001_15101_00001_mirimage/jw01355015001_15101_00001_mirimage_cal.fits'
-      - './data/spt0418/f2100w/jw01355015001_15101_00002_mirimage/jw01355015001_15101_00002_mirimage_cal.fits'
-      - './data/spt0418/f2100w/jw01355015001_15101_00003_mirimage/jw01355015001_15101_00003_mirimage_cal.fits'
-      - './data/spt0418/f2100w/jw01355015001_15101_00004_mirimage/jw01355015001_15101_00004_mirimage_cal.fits'
-      - './data/spt0418/f2100w/jw01355015001_15101_00005_mirimage/jw01355015001_15101_00005_mirimage_cal.fits'
-      - './data/spt0418/f2100w/jw01355015001_15101_00006_mirimage/jw01355015001_15101_00006_mirimage_cal.fits'
[…]
