### 2025-03-18 a23d8dd
imaging_resolve_jax: simplify output file names
