### 2025-03-18 8e5180d
imaging_resolve: format code
