+++ demos/imaging_resolve.py
@@ -14,13 +14,13 @@ import jubik0 as ju
 import jubik0.instruments.resolve as rve
 import jubik0.instruments.resolve.re as jrve
 from jubik0.parse.grid import GridModel
-from jubik0.instruments.resolve.data.data_modify.restrict_to_testing_percentage import restrict_to_testing_percentage
-from jubik0.parse.instruments.resolve.response import (
-    Ducc0Settings, FinufftSettings)
+from jubik0.instruments.resolve.data.data_modify.restrict_to_testing_percentage import (
+    restrict_to_testing_percentage,
+)
+from jubik0.parse.instruments.resolve.response import Ducc0Settings, FinufftSettings
 
 
-
-jax.config.update('jax_default_device', jax.devices('cpu')[0])
+jax.config.update("jax_default_device", jax.devices("cpu")[0])
 jax.config.update("jax_enable_x64", True)
 
 seed = 42
@@ -30,19 +30,17 @@ cfg = configparser.ConfigParser()
 cfg.read("./demos/configs/cygnusa_2ghz.cfg")
 
 # choose between ducc0 and finufft backend
-response = 'ducc0'
+response = "ducc0"
 # response = "finufft"
 backend_settings = dict(
-    ducc0=Ducc0Settings(
-        epsilon=1e-9, do_wgridding=False, nthreads=1, verbosity=False),
+    ducc0=Ducc0Settings(epsilon=1e-9, do_wgridding=False, nthreads=1, verbosity=False),
     finufft=FinufftSettings(epsilon=1e-9),
 )[response]
 
 # # NOTE : The observation can also be loaded and modified via the config file.
 # backend_settings = config_parser_to_response_settings(cfg['data'])
 
-obs = rve.Observation.load(
-    "./data/resolve_test/CYG-ALL-2052-2MHZ_RESOLVE_float64.npz")
+obs = rve.Observation.load("./data/resolve_test/CYG-ALL-2052-2MHZ_RESOLVE_float64.npz")
 obs = obs.restrict_to_stokesi()
 obs = obs.average_stokesi()
 # scale weights, as they are wrong for this specific dataset
@@ -58,16 +56,16 @@ obs = restrict_to_testing_percentage(obs, 0.01)
 
 sky, additional = jrve.sky_model(cfg["sky"])
 
-gm = GridModel.from_config_parser(cfg['sky'])
+gm = GridModel.from_config_parser(cfg["sky"])
 gm.spatial_model.wcs_model.center = obs.direction.to_sky_coord()
 grid = ju.Grid.from_grid_model(gm)
 
 
-R_new = jrve.InterferometryResponse(
-    obs, grid, backend_settings=backend_settings)
+R_new = jrve.InterferometryResponse(obs, grid, backend_settings=backend_settings)
 
 
-def signal_response(x): return R_new(sky(x))
+def signal_response(x):
+    return R_new(sky(x))
 
 
 nll = jft.Gaussian(obs.vis_val, obs.weight_val).amend(signal_response)
@@ -79,8 +77,7 @@ def callback(samples, opt_state):
     post_sr_mean = jft.mean(tuple(sky(s) for s in samples))
     plt.imshow(post_sr_mean[0, 0, 0, :, :].T, origin="lower", norm=LogNorm())
     plt.colorbar()
-    plt.savefig(
-        join(output_dir, f"iteration_{opt_state.nit}.png"))
+    plt.savefig(join(output_dir, f"iteration_{opt_state.nit}.png"))
     plt.close()
 
 
