+++ src/instruments/resolve/re/response.py
@@ -115,8 +115,11 @@ def InterferometryResponse(
     pixsize_x, pixsize_y = sky_grid.spatial.distances_in(SPATIAL_UNIT)
     center_y, center_x = calculate_phase_offset_to_image_center(
         sky_grid.spatial.center,
-        sky_grid.spatial.center if observation.direction is None
-        else observation.direction.to_sky_coord()
+        (
+            sky_grid.spatial.center
+            if observation.direction is None
+            else observation.direction.to_sky_coord()
+        ),
     )
 
     # build responses for: time binds, freq bins
@@ -124,12 +127,11 @@ def InterferometryResponse(
     row_indices, freq_indices = [], []
     for t in range(n_times):
         sr_tmp, t_tmp, f_tmp = [], [], []
-        if tuple(bb_times[t: t + 2]) == (-np.inf, np.inf):
+        if tuple(bb_times[t : t + 2]) == (-np.inf, np.inf):
             oo = observation
             tind = slice(None)
         else:
-            oo, tind = observation.restrict_by_time(
-                bb_times[t], bb_times[t + 1], True)
+            oo, tind = observation.restrict_by_time(bb_times[t], bb_times[t + 1], True)
         for f in range(n_freqs):
             ooo, find = oo.restrict_by_freq(bb_freqs[f], bb_freqs[f + 1], True)
             if any(np.array(ooo.vis.shape) == 0):
@@ -150,7 +152,7 @@ def InterferometryResponse(
                         center_y=center_y,
                     )
                 elif isinstance(backend_settings, FinufftSettings):
-                    print('Using Finufft')
+                    print("Using Finufft")
                     rrr = InterferometryResponseFinuFFT(
                         observation=ooo,
                         pixsize_x=pixsize_x,
@@ -160,9 +162,11 @@ def InterferometryResponse(
                         center_y=center_y,
                     )
                 else:
-                    err = ("backend_settings must be an instance of "
-                           "`Ducc0Settings` or `FinufftSettings`, not "
-                           f"{backend_settings}")
+                    err = (
+                        "backend_settings must be an instance of "
+                        "`Ducc0Settings` or `FinufftSettings`, not "
+                        f"{backend_settings}"
+                    )
                     raise ValueError(err)
 
             sr_tmp.append(rrr)
@@ -194,8 +198,7 @@ def InterferometryResponse(
                         continue
                     inp = sky[pp, tt, ff]
                     r = op(inp)
-                    res = res.at[pp, row_indices[tt][ff],
-                                 freq_indices[tt][ff]].set(r)
+                    res = res.at[pp, row_indices[tt][ff], freq_indices[tt][ff]].set(r)
         return convert_polarization(res, inp_pol, out_pol)
 
     return apply_R
@@ -235,12 +238,7 @@ def InterferometryResponseDucc(
 
 
 def InterferometryResponseFinuFFT(
-    observation,
-    pixsize_x,
-    pixsize_y,
-    epsilon,
-    center_x=None,
-    center_y=None
+    observation, pixsize_x, pixsize_y, epsilon, center_x=None, center_y=None
 ):
     from jax_finufft import nufft2
 
@@ -249,8 +247,7 @@ def InterferometryResponseFinuFFT(
     vol = pixsize_x * pixsize_y
     speedoflight = 299792458.0
 
-    uvw = np.transpose(
-        (uvw[..., None] * freq / speedoflight), (0, 2, 1)).reshape(-1, 3)
[…]
