### 2025-03-18 5b5ea10
resolve_sky: formatting
