+++ src/instruments/jwst/jwst_data.py
@@ -7,6 +7,7 @@
 
 from astropy import units
 from astropy.coordinates import SkyCoord
+
 try:
     from jwst import datamodels
 except ImportError:
@@ -22,35 +23,35 @@ from .wcs.wcs_jwst_data import WcsJwstData
 nircam_filters = dict(
     # https://jwst-docs.stsci.edu/jwst-mid-infrared-instrument/miri-instrumentation/miri-filters-and-dispersers#gsc.tab=0
     # Pivot λ (μm), BW Δλ (μm), Effective response, Blue λ- (µm), Red λ+ (µm)
-    F070W=(0.704,	0.128,	0.237,	0.624,	0.781),
-    F090W=(0.901,	0.194,	0.318,	0.795,	1.005),
-    F115W=(1.154,	0.225,	0.333,	1.013,	1.282),
-    F140M=(1.404,	0.142,	0.434,	1.331,	1.479),
-    F150W=(1.501,	0.318,	0.476,	1.331,	1.668),
-    F162M=(1.626,	0.168,	0.469,	1.542,	1.713),
-    F164N=(1.644,	0.020,	0.385,	1.635,	1.653),
-    F150W2=(1.671,	1.227,	0.489,	1.007,	2.38),
-    F182M=(1.845,	0.238,	0.505,	1.722,	1.968),
-    F187N=(1.874,	0.024,	0.434,	1.863,	1.885),
-    F200W=(1.990,	0.461,	0.525,	1.755,	2.227),
-    F210M=(2.093,	0.205,	0.522,	1.992,	2.201),
-    F212N=(2.120,	0.027,	0.420,	2.109,	2.134),
-    F250M=(2.503,	0.181,	0.370,	2.412,	2.595),
-    F277W=(2.786,	0.672,	0.412,	2.423,	3.132),
-    F300M=(2.996,	0.318,	0.432,	2.831,	3.157),
-    F322W3=(3.247,	1.339,	0.499,	2.432,	4.013),
-    F323N=(3.237,	0.038,	0.290,	3.217,	3.255),
-    F335M=(3.365,	0.347,	0.480,	3.177,	3.537),
-    F356W=(3.563,	0.787,	0.530,	3.135,	3.981),
-    F360M=(3.621,	0.372,	0.515,	3.426,	3.814),
-    F405N=(4.055,	0.046,	0.418,	4.030,	4.076),
-    F410M=(4.092,	0.436,	0.499,	3.866,	4.302),
-    F430M=(4.280,	0.228,	0.526,	4.167,	4.398),
-    F444W=(4.421,	1.024,	0.533,	3.881,	4.982),
-    F460M=(4.624,	0.228,	0.460,	4.515,	4.747),
-    F466N=(4.654,	0.054,	0.320,	4.629,	4.681),
-    F470N=(4.707,	0.051,	0.316,	4.683,	4.733),
-    F480M=(4.834,	0.303,	0.447,	4.662,	4.973),
+    F070W=(0.704, 0.128, 0.237, 0.624, 0.781),
+    F090W=(0.901, 0.194, 0.318, 0.795, 1.005),
+    F115W=(1.154, 0.225, 0.333, 1.013, 1.282),
+    F140M=(1.404, 0.142, 0.434, 1.331, 1.479),
+    F150W=(1.501, 0.318, 0.476, 1.331, 1.668),
+    F162M=(1.626, 0.168, 0.469, 1.542, 1.713),
+    F164N=(1.644, 0.020, 0.385, 1.635, 1.653),
+    F150W2=(1.671, 1.227, 0.489, 1.007, 2.38),
[…]
