+++ src/instruments/jwst/rotation_and_shift/__init__.py
@@ -8,4 +8,6 @@
 from .linear_rotation_and_shift import build_linear_rotation_and_shift
 from .nufft_rotation_and_shift import build_nufft_rotation_and_shift
 from .build_rotation_and_shift import (
-    build_rotation_and_shift_model, RotationAndShiftModel)
+    build_rotation_and_shift_model,
+    RotationAndShiftModel,
+)

[…]
