+++ src/instruments/resolve/likelihood/radio_likelihood.py
@@ -24,7 +24,9 @@ from ....grid import Grid
 from ..data.data_loading import load_and_modify_data_from_objects
 from ..multimessanger import build_radio_sky_extractor, build_radio_grid
 from ..mosaicing.sky_beamer import build_jft_sky_beamer
-from ..telescopes.primary_beam import build_primary_beam_pattern_from_beam_pattern_config
+from ..telescopes.primary_beam import (
+    build_primary_beam_pattern_from_beam_pattern_config,
+)
 from .mosaic_likelihood import build_likelihood_from_sky_beamer
 from .cast_to_dtype import cast_to_dtype
 
@@ -45,20 +47,20 @@ def build_radio_likelihood(
     sky_grid: Grid,
     sky_model: jft.Model,
     last_radio_bin: int | None,
-    sky_key: str = 'sky',
+    sky_key: str = "sky",
     sky_unit: u.Unit | None = None,
-    direction_key: str = 'PHASE_DIR',
+    direction_key: str = "PHASE_DIR",
 ):
+    response_settings = yaml_to_response_settings(cfg["radio_response"])
 
     radio_sky_extractor = build_radio_sky_extractor(
         last_radio_bin,
         sky_model,
         sky_key=sky_key,
-        sky_unit=sky_unit)
+        sky_unit=sky_unit,
+        transpose=response_settings.transpose,
+    )
     radio_grid = build_radio_grid(last_radio_bin, sky_grid)
-
-    response_backend_settings = yaml_to_response_settings(
-        cfg['radio_response'])
 
     if not isinstance(data_names, list):
         assert isinstance(data_names, str)
@@ -67,25 +69,30 @@ def build_radio_likelihood(
     likelihoods = []
     sky_beamers = []
     for data_name in data_names:
-        logger.info(f'Loading data: {data_name}')
+        logger.info(f"Loading data: {data_name}")
 
-        dl = DataLoading.from_yaml_dict(cfg['alma_data'][data_name])
-        dm = ObservationModify.from_yaml_dict(cfg['alma_data'][data_name])
-        observations = list(load_and_modify_data_from_objects(
-            sky_frequencies=radio_grid.spectral.binbounds_in(u.Hz),
-            data_loading=dl,
-            observation_modify=dm))
+        dl = DataLoading.from_yaml_dict(cfg["alma_data"][data_name])
+        dm = ObservationModify.from_yaml_dict(cfg["alma_data"][data_name])
+        observations = list(
+            load_and_modify_data_from_objects(
+                sky_frequencies=radio_grid.spectral.binbounds_in(u.Hz),
+                data_loading=dl,
+                observation_modify=dm,
+            )
+        )
 
         # TODO: The following lines have to be simplified.
         beam_func = build_primary_beam_pattern_from_beam_pattern_config(
-            BeamPatternConfig.from_yaml_dict(cfg['alma_data'][data_name]['dish']))
+            BeamPatternConfig.from_yaml_dict(cfg["alma_data"][data_name]["dish"])
+        )
 
         _sky_beamer = build_jft_sky_beamer(
             sky_shape_with_dtype=radio_sky_extractor.target,
             sky_fov=sky_grid.spatial.fov,
             sky_center=sky_grid.spatial.center,
             sky_frequency_binbounds=sky_grid.spectral.binbounds_in(
-                RESOLVE_SPECTRAL_UNIT),
+                RESOLVE_SPECTRAL_UNIT
+            ),
             observations=observations,
             beam_func=beam_func,
             direction_key=direction_key,
@@ -96,28 +103,30 @@ def build_radio_likelihood(
         for field_name, beam_direction in _sky_beamer.beam_directions.items():
             for o in observations:
[…]
