### 2025-03-18 abefcd9
87-consistent-formatting: using the ruff formatter & linte
