+++ src/sky_model/resolve_point_sources.py
@@ -14,16 +14,15 @@ def jax_insert(x, ptsx, ptsy, bg):
 
 
 def resolve_point_sources(
-    sky_dom: ift.RGSpace,
-    resolve_point_souces_model: ResolvePointSourcesModel,
-    bg=None
+    sky_dom: ift.RGSpace, resolve_point_souces_model: ResolvePointSourcesModel, bg=None
 ):
     if resolve_point_souces_model.freq_mode == "single":
         if resolve_point_souces_model.polarization == "I":
             if resolve_point_souces_model.mode == "fixed_locations":
 
                 indsx, indsy = resolve_point_souces_model.locations.to_indices(
-                    sky_dom, unit=u.rad)
+                    sky_dom, unit=u.rad
+                )
                 shp = sky_dom.shape
                 full_shp = (1, 1, 1) + shp
 
@@ -36,14 +35,17 @@ def resolve_point_sources(
                 if bg is None:
                     bg = jnp.zeros(full_shp)
 
-                    def pts_func(x): return jax_insert(
-                        inv_gamma(x), ptsx=indsx, ptsy=indsy, bg=bg
-                    )
+                    def pts_func(x):
+                        return jax_insert(inv_gamma(x), ptsx=indsx, ptsy=indsy, bg=bg)
+
                     dom = inv_gamma.domain
                 else:
-                    def pts_func(x): return jax_insert(
-                        inv_gamma(x), ptsx=indsx, ptsy=indsy, bg=bg(x)
-                    )
+
+                    def pts_func(x):
+                        return jax_insert(
+                            inv_gamma(x), ptsx=indsx, ptsy=indsy, bg=bg(x)
+                        )
+
                     dom = {**inv_gamma.domain, **bg.domain}
                 pts_model = jft.Model(pts_func, domain=dom)
                 additional = {}
