+++ src/wcs/wcs_astropy.py
@@ -5,8 +5,7 @@
 
 # %%
 from .wcs_base import WcsBase
-from ..parse.wcs.coordinate_system import (
-    CoordinateSystemModel, CoordinateSystems)
+from ..parse.wcs.coordinate_system import CoordinateSystemModel, CoordinateSystems
 from ..parse.wcs.spatial_model import SpatialModel
 
 import numpy as np
@@ -31,7 +30,9 @@ class WcsAstropy(WCS, WcsBase):
         shape: tuple[int, int],
         fov: tuple[u.Quantity, u.Quantity],
         rotation: u.Quantity = 0.0 * u.deg,
-        coordinate_system: Optional[CoordinateSystemModel] = CoordinateSystems.icrs.value
+        coordinate_system: Optional[
+            CoordinateSystemModel
+        ] = CoordinateSystems.icrs.value,
     ):
         """
         Create FITS header, use it to instantiate an WcsAstropy.
@@ -54,7 +55,7 @@ class WcsAstropy(WCS, WcsBase):
 
         self.shape = shape
         self.fov = fov
-        self.distances = [f.to(u.deg)/s for f, s in zip(fov, shape)]
+        self.distances = [f.to(u.deg) / s for f, s in zip(fov, shape)]
         self.center = center
 
         # Calculate rotation matrix
@@ -74,29 +75,30 @@ class WcsAstropy(WCS, WcsBase):
 
         # Build the header dictionary
         header = {
-            'WCSAXES': 2,
-            'CTYPE1': coordinate_system.ctypes[0],
-            'CTYPE2': coordinate_system.ctypes[1],
-            'CRPIX1': shape[0] / 2 + 0.5,
-            'CRPIX2': shape[1] / 2 + 0.5,
-            'CRVAL1': lon,
-            'CRVAL2': lat,
-            'CDELT1': -fov[0].to(u.deg).value / shape[0],
-            'CDELT2': fov[1].to(u.deg).value / shape[1],
-            'PC1_1': pc11,
-            'PC1_2': pc12,
-            'PC2_1': pc21,
-            'PC2_2': pc22,
-            'RADESYS': coordinate_system.radesys,
-            'CUNIT1': 'deg',
-            'CUNIT2': 'deg',
+            "WCSAXES": 2,
+            "CTYPE1": coordinate_system.ctypes[0],
+            "CTYPE2": coordinate_system.ctypes[1],
+            "CRPIX1": shape[0] / 2 + 0.5,
+            "CRPIX2": shape[1] / 2 + 0.5,
+            "CRVAL1": lon,
+            "CRVAL2": lat,
+            "CDELT1": -fov[0].to(u.deg).value / shape[0],
+            "CDELT2": fov[1].to(u.deg).value / shape[1],
+            "PC1_1": pc11,
+            "PC1_2": pc12,
+            "PC2_1": pc21,
+            "PC2_2": pc22,
+            "RADESYS": coordinate_system.radesys,
+            "CUNIT1": "deg",
+            "CUNIT2": "deg",
         }
 
         # Set equinox if needed for FK4/FK5
         if coordinate_system.radesys in [
-                CoordinateSystems.fk4.value.radesys,
-                CoordinateSystems.fk5.value.radesys]:
-            header['EQUINOX'] = coordinate_system.equinox
+            CoordinateSystems.fk4.value.radesys,
+            CoordinateSystems.fk5.value.radesys,
+        ]:
+            header["EQUINOX"] = coordinate_system.equinox
 
         super().__init__(header)
 
@@ -111,9 +113,7 @@ class WcsAstropy(WCS, WcsBase):
         )
 
     # TODO: Check output axis, RENAME index_from_world_location
-    def wl_from_index(
[…]
