+++ src/wcs/wcs_jwst_data.py
@@ -24,17 +24,16 @@ class WcsJwstData(WcsBase):
         try:
             from gwcs import WCS
         except ImportError:
-            raise ImportError("gwcs not installed."
-                              "Please install via 'pip install gwcs'.")
+            raise ImportError(
+                "gwcs not installed." "Please install via 'pip install gwcs'."
+            )
 
         if not isinstance(wcs, WCS):
-            raise TypeError('wcs must be a gwcs.WCS')
+            raise TypeError("wcs must be a gwcs.WCS")
 
         self.wcs = wcs
 
-    def wl_from_index(
-        self, index: ArrayLike
-    ) -> Union[SkyCoord, List[SkyCoord]]:
+    def wl_from_index(self, index: ArrayLike) -> Union[SkyCoord, List[SkyCoord]]:
         """
         Convert pixel coordinates to world coordinates.
 
