+++ src/wcs/wcs_subsample_centers.py
@@ -47,23 +47,28 @@ def subsample_grid_centers_in_index_grid_non_vstack(
         each axis.
     """
     tbsg_pixcenter_indices = to_be_subsampled_grid_wcs.index_grid_from_wl_extrema(
-        world_extrema, indexing=indexing)
+        world_extrema, indexing=indexing
+    )
 
-    ps = np.arange(0.5/subsample, 1, 1/subsample) - 0.5
+    ps = np.arange(0.5 / subsample, 1, 1 / subsample) - 0.5
     ms = np.vstack(np.array(np.meshgrid(ps, ps, indexing=indexing)).T)
 
-    subsample_centers = np.zeros((
-        tbsg_pixcenter_indices.shape[0],
-        tbsg_pixcenter_indices.shape[1] * subsample,
-        tbsg_pixcenter_indices.shape[2] * subsample,
-    ))
+    subsample_centers = np.zeros(
+        (
+            tbsg_pixcenter_indices.shape[0],
+            tbsg_pixcenter_indices.shape[1] * subsample,
+            tbsg_pixcenter_indices.shape[2] * subsample,
+        )
+    )
     for ii, ps in enumerate(ms):
         xx = ii % subsample
         yy = ii // subsample
         subsample_centers[:, xx::subsample, yy::subsample] = (
-            tbsg_pixcenter_indices + ps[:, None, None])
+            tbsg_pixcenter_indices + ps[:, None, None]
+        )
 
-    wl_subsample_centers = to_be_subsampled_grid_wcs.wl_from_index(
-        [subsample_centers])[0]
+    wl_subsample_centers = to_be_subsampled_grid_wcs.wl_from_index([subsample_centers])[
+        0
+    ]
 
     return index_grid_wcs.index_from_wl(wl_subsample_centers)[0]
