+++ src/wcs/wcs_base.py
@@ -18,10 +18,9 @@ class WcsBase(ABC):
     coordinates. Inherited classes need to provide a `wl_from_index` and an
     `index_from_wl` method.
     """
+
     @abstractmethod
-    def wl_from_index(
-        self, index: ArrayLike
-    ) -> Union[SkyCoord, List[SkyCoord]]:
+    def wl_from_index(self, index: ArrayLike) -> Union[SkyCoord, List[SkyCoord]]:
         """
         Convert pixel coordinates to world coordinates.
 
@@ -54,9 +53,7 @@ class WcsBase(ABC):
         pass
 
     def index_from_wl_extrema(
-        self,
-        world_extrema: SkyCoord,
-        shape_check: Optional[Tuple[int, int]] = None
+        self, world_extrema: SkyCoord, shape_check: Optional[Tuple[int, int]] = None
     ) -> Tuple[int, int, int, int]:
         """
         Find the minimum and maximum pixel indices of the bounding box that
@@ -82,9 +79,9 @@ class WcsBase(ABC):
 
         if shape_check is not None:
             check = (
-                np.any(edges_dgrid < 0) or
-                np.any(edges_dgrid >= shape_check[0]) or
-                np.any(edges_dgrid >= shape_check[1])
+                np.any(edges_dgrid < 0)
+                or np.any(edges_dgrid >= shape_check[0])
+                or np.any(edges_dgrid >= shape_check[1])
             )
             if check:
                 o = f"""One of the wcs world_extrema is outside the data grid
@@ -102,7 +99,7 @@ class WcsBase(ABC):
         self,
         world_extrema: SkyCoord,
         indexing: str,
-        shape_check: Tuple[int, int] | None = None
+        shape_check: Tuple[int, int] | None = None,
     ) -> np.typing.ArrayLike:
         """
         Find the pixel indices of the bounding box that contain the world
@@ -126,9 +123,12 @@ class WcsBase(ABC):
             the edge points.
         """
 
-        minx, maxx, miny, maxy = self.index_from_wl_extrema(
-            world_extrema, shape_check)
+        minx, maxx, miny, maxy = self.index_from_wl_extrema(world_extrema, shape_check)
 
-        return np.array(np.meshgrid(np.arange(minx, maxx+1, 1),
-                                    np.arange(miny, maxy+1, 1),
-                                    indexing=indexing))
+        return np.array(
+            np.meshgrid(
+                np.arange(minx, maxx + 1, 1),
+                np.arange(miny, maxy + 1, 1),
+                indexing=indexing,
+            )
+        )
