+++ src/parse/instruments/resolve/response.py
@@ -8,15 +8,15 @@ from typing import Optional, Union
 from configparser import ConfigParser
 
 
-FINUFFT_KEYS = ['finufft']
-DUCC_KEYS = ['ducc', 'ducc0']
+FINUFFT_KEYS = ["finufft"]
+DUCC_KEYS = ["ducc", "ducc0"]
 
-EPSILON_KEY = 'epsilon'
+EPSILON_KEY = "epsilon"
 
-BACKEND_KEY = 'backend'
-DO_WGRIDDING_KEY = 'do_wgridding'
-NTHREADS_KEY = 'nthreads'
-VERBOSITY_KEY = 'verbosity'
+BACKEND_KEY = "backend"
+DO_WGRIDDING_KEY = "do_wgridding"
+NTHREADS_KEY = "nthreads"
+VERBOSITY_KEY = "verbosity"
 
 
 @dataclass
@@ -28,7 +28,7 @@ class Ducc0Settings:
 
     @classmethod
     def from_yaml_dict(cls, yaml_dict: dict):
-        f'''Read ducc0 settings from yaml_dict.
+        f"""Read ducc0 settings from yaml_dict.
 
         Parameters
         ----------
@@ -36,7 +36,7 @@ class Ducc0Settings:
         {DO_WGRIDDING_KEY}: bool
         {NTHREADS_KEY}: int
         {VERBOSITY_KEY}: bool
-        '''
+        """
         epsilon = yaml_dict[EPSILON_KEY]
         do_wgridding = yaml_dict[DO_WGRIDDING_KEY]
         nthreads = yaml_dict[NTHREADS_KEY]
@@ -50,7 +50,7 @@ class Ducc0Settings:
 
     @classmethod
     def from_config_parser(cls, config: ConfigParser):
-        f'''Read ducc0 settings from yaml_dict.
+        f"""Read ducc0 settings from yaml_dict.
 
         Parameters
         ----------
@@ -58,7 +58,7 @@ class Ducc0Settings:
         {DO_WGRIDDING_KEY}: bool | None
         {NTHREADS_KEY}: int | None
         {VERBOSITY_KEY}: bool | None
-        '''
+        """
         epsilon = eval(config[EPSILON_KEY])
         do_wgridding = eval(config.get(DO_WGRIDDING_KEY, False))
         nthreads = eval(config.get(NTHREADS_KEY, 1))
@@ -78,31 +78,31 @@ class FinufftSettings:
 
     @classmethod
     def from_yaml_dict(cls, yaml_dict: dict):
-        f'''Read finufft settings from yaml_dict.
+        f"""Read finufft settings from yaml_dict.
 
         Parameters
         ----------
         {EPSILON_KEY}: float
-        '''
+        """
         epsilon = yaml_dict[EPSILON_KEY]
         return FinufftSettings(epsilon=epsilon)
 
     @classmethod
     def from_config_parser(cls, config: ConfigParser):
-        f'''Read finufft settings from yaml_dict.
+        f"""Read finufft settings from yaml_dict.
 
         Parameters
         ----------
         {EPSILON_KEY}: float
-        '''
+        """
         epsilon = eval(config[EPSILON_KEY])
         return FinufftSettings(epsilon=epsilon)
 
 
 def yaml_to_response_settings(
-    response_dict: dict
+    response_dict: dict,
 ) -> Union[Ducc0Settings, FinufftSettings]:
-    f'''Read the yaml file in order to parse to Backend settings.
+    f"""Read the yaml file in order to parse to Backend settings.
     These can either be `Ducc0Settings` or `FinufftSettings`. 
 
     Parameters
@@ -112,7 +112,7 @@ def yaml_to_response_settings(
     Note
     ----
[…]
