### 2025-03-20 4d2c77b
resolve: ruff format resolve
