### 2025-03-20 188b801
resolve: ruff format resolve
