+++ src/instruments/resolve/likelihood/mosaic_likelihood.py
@@ -1,5 +1,4 @@
-from ....parse.instruments.resolve.response import (
-    Ducc0Settings, FinufftSettings)
+from ....parse.instruments.resolve.response import Ducc0Settings, FinufftSettings
 
 from ....grid import Grid
 from ..re.response import InterferometryResponse
@@ -19,7 +18,7 @@ def build_likelihood_from_sky_beamer(
     backend_settings: Union[Ducc0Settings, FinufftSettings],
     cast_to_dtype: Callable | None = None,
 ):
-    '''First, builds response operator, which takes the `field_name` from the
+    """First, builds response operator, which takes the `field_name` from the
     sky_beamer operator and calculates the visibilities corresponding to the
     observation.
     Second, builds the likelihood operator corresponding to this observation.
@@ -37,7 +36,7 @@ def build_likelihood_from_sky_beamer(
     The likelihood which takes the beam-corrected sky corresponding to the
     `observation`, which gets transformed to visibility space and compared to
     the visibilities in the observation.
-    '''
+    """
 
     sky2vis = InterferometryResponse(
         observation=observation,
@@ -49,7 +48,6 @@ def build_likelihood_from_sky_beamer(
         response = jft.wrap(lambda x: sky2vis(cast_to_dtype(x), field_name))
 
     likelihood = jft.Gaussian(
-        observation.vis.val,
-        noise_cov_inv=lambda x: x*observation.weight.val
+        observation.vis.val, noise_cov_inv=lambda x: x * observation.weight.val
     )
     return likelihood.amend(response, domain=jft.Vector(sky_beamer.target))
