+++ src/instruments/resolve/data/direction.py
@@ -30,7 +30,7 @@ class Direction:
         reference year of the equinox
     """
 
-    def __init__(self, phase_center, equinox, name=''):
+    def __init__(self, phase_center, equinox, name=""):
         my_asserteq(len(phase_center), 2)
         self._pc = phase_center
         self._e = float(equinox)
@@ -55,11 +55,12 @@ class Direction:
         return [*self._pc, self._e]
 
     def to_sky_coord(self):
-        assert self._e in ['J2000', 2000., 2000], (
+        assert self._e in ["J2000", 2000.0, 2000], (
             f"The equinox has to be `J2000`, provided equinox {self._e}.\n"
             "The icrs and fk5 correspond at j2000. Here, we make the explicit "
-            "assumption that the observations are recorded in the icrs frame.")
-        return SkyCoord(*(self._pc*u.rad), frame='icrs')
+            "assumption that the observations are recorded in the icrs frame."
+        )
+        return SkyCoord(*(self._pc * u.rad), frame="icrs")
 
     @staticmethod
     def from_list(lst):
