+++ src/instruments/resolve/mosaicing/sky_wcs.py
@@ -31,7 +31,7 @@ def build_astropy_wcs(
     fov: Tuple[Unit, Unit],
     rotation: Unit = 0.0 * units.deg,
 ) -> WCS:
-    '''
+    """
     Specify the Astropy wcs.
 
     Parameters
@@ -47,7 +47,7 @@ def build_astropy_wcs(
 
     rotation : units.Quantity
         The rotation of the grid WCS with respect to the ICRS system, in degrees.
-    '''
+    """
 
     # Create a WCS object
     w = WCS(naxis=2)
@@ -61,8 +61,10 @@ def build_astropy_wcs(
 
     # Set up ICRS system
     w.wcs.crpix = [shape[0] / 2 + 0.5, shape[1] / 2 + 0.5]
-    w.wcs.cdelt = [-fov[0].to(units.deg).value / shape[0],
-                   fov[1].to(units.deg).value / shape[1]]
+    w.wcs.cdelt = [
+        -fov[0].to(units.deg).value / shape[0],
+        fov[1].to(units.deg).value / shape[1],
+    ]
     w.wcs.crval = [center.ra.deg, center.dec.deg]
     w.wcs.ctype = ["RA---TAN", "DEC--TAN"]
     w.wcs.pc = np.array([[pc11, pc12], [pc21, pc22]])
