+++ src/instruments/resolve/data/data_modify/frequency_averaging.py
@@ -9,7 +9,7 @@ def freq_average_by_bins(obs: Observation, n_freq_chuncks: int | None):
     if n_freq_chuncks is None:
         return obs
 
-    logger.info(f'Frequency averaging observation to {n_freq_chuncks}')
+    logger.info(f"Frequency averaging observation to {n_freq_chuncks}")
 
     splitted_freqs = np.array_split(obs.freq, n_freq_chuncks)
     return freq_average_by_fmin_fmax(obs, splitted_freqs)
@@ -20,7 +20,7 @@ def freq_average_by_fdom_and_n_freq_chunks(
     obs: Observation,
     n_freq_chuncks: int,
 ):
-    '''Create a new observation with frequencies averaged. The frequencies of
+    """Create a new observation with frequencies averaged. The frequencies of
     the new observation will be averaged into `n_freq_chuncks` according to the
     ranges of the sky_frequencies.
 
@@ -32,12 +32,14 @@ def freq_average_by_fdom_and_n_freq_chunks(
         The observation to be modified.
     n_freq_chuncks: int
         The number of frequency chuncks.
-    '''
+    """
     if n_freq_chuncks is None:
         return obs
 
-    fmin_fmax_array = [(sky_frequencies[ii], sky_frequencies[ii+1])
-                       for ii in range(len(sky_frequencies)-1)]
+    fmin_fmax_array = [
+        (sky_frequencies[ii], sky_frequencies[ii + 1])
+        for ii in range(len(sky_frequencies) - 1)
+    ]
     splitted_obs_freq = []
     for ff in fmin_fmax_array:
         obs_freq = obs.restrict_by_freq(ff[0], ff[-1]).freq
@@ -45,16 +47,19 @@ def freq_average_by_fdom_and_n_freq_chunks(
             continue
         splitted_obs_freq.append(obs_freq)
 
-    logger.info('Frequency averaging observation to (N_ObsInSky, N_Chunks) = '
-                f'({len(splitted_obs_freq)}, {n_freq_chuncks})')
+    logger.info(
+        "Frequency averaging observation to (N_ObsInSky, N_Chunks) = "
+        f"({len(splitted_obs_freq)}, {n_freq_chuncks})"
+    )
 
     splitted_freqs = []
     for ofreq in splitted_obs_freq:
         # TODO : Make it robust against bad sky frequency choices.
         tmp_splits = np.array_split(ofreq, n_freq_chuncks)
         for tmp in tmp_splits:
-            assert tmp[0] != tmp[-1], ('Frequency chunking not of data not '
-                                       'compatible with sky frequencies.')
+            assert tmp[0] != tmp[-1], (
+                "Frequency chunking not of data not compatible with sky frequencies."
+            )
             splitted_freqs.append(np.array([tmp[0], tmp[-1]]))
 
     return freq_average_by_fmin_fmax(obs, splitted_freqs)
@@ -72,7 +77,7 @@ def freq_average_by_fmin_fmax(
     for obsi in splitted_obs:
         new_vis = np.mean(obsi.vis.val, axis=2, keepdims=True)
         cov = 1 / obsi.weight.val
-        new_cov = np.sum(cov, axis=2, keepdims=True) / (obsi.vis.shape[2]**2)
+        new_cov = np.sum(cov, axis=2, keepdims=True) / (obsi.vis.shape[2] ** 2)
         new_weight = 1 / new_cov
         new_freq = np.array([np.mean(obsi.freq)])
         new_obs = Observation(
@@ -81,7 +86,8 @@ def freq_average_by_fmin_fmax(
             new_weight,
             obsi.polarization,
             new_freq,
-            obs._auxiliary_tables)
[…]
