+++ src/instruments/resolve/data/data_modify/revert_frequencies.py
@@ -3,7 +3,7 @@ from nifty8.logger import logger
 
 
 def revert_frequencies(obs: Observation) -> Observation:
-    '''This reverts the frequencies and returns an observation'''
+    """This reverts the frequencies and returns an observation"""
     logger.info("Reverting frequencies")
     return Observation(
         obs.antenna_positions,
@@ -11,5 +11,5 @@ def revert_frequencies(obs: Observation) -> Observation:
         obs.weight.val[:, :, ::-1],
         obs.polarization,
         obs.freq[::-1],
-        auxiliary_tables=obs._auxiliary_tables
+        auxiliary_tables=obs._auxiliary_tables,
     )
