### 2025-03-20 8cb0fc0
resolve jr: ruff format
