+++ src/instruments/resolve/data/ms_import.py
@@ -30,7 +30,9 @@ def ms_table(path):
     try:
         from casacore.tables import table
     except ImportError:
-        raise ImportError("You need to install python-casacore for working with measurement sets")
+        raise ImportError(
+            "You need to install python-casacore for working with measurement sets"
+        )
     return table(path, readonly=True, ack=False)
 
 
@@ -49,9 +51,16 @@ def ms2observations_all(ms, data_column):
                 yield oo
 
 
-def ms2observations(ms, data_column, with_calib_info, spectral_window,
-                    polarizations="all", channels=slice(None), ignore_flags=False,
-                    field=None):
+def ms2observations(
+    ms,
+    data_column,
+    with_calib_info,
+    spectral_window,
+    polarizations="all",
+    channels=slice(None),
+    ignore_flags=False,
+    field=None,
+):
     """Read and convert a given measurement set into an array of :class:`Observation`
 
     If WEIGHT_SPECTRUM is available this column is used for weighting.
@@ -137,27 +146,42 @@ def ms2observations(ms, data_column, with_calib_info, spectral_window,
         auxtables = {}
         auxtables["ANTENNA"] = _import_aux_table(ms, "ANTENNA")
         auxtables["STATE"] = _import_aux_table(ms, "STATE")
-        auxtables["SPECTRAL_WINDOW"] = _import_aux_table(ms, "SPECTRAL_WINDOW", row=spectral_window, skip=["ASSOC_NATURE", "ASSOC_SPW_ID"])
+        auxtables["SPECTRAL_WINDOW"] = _import_aux_table(
+            ms,
+            "SPECTRAL_WINDOW",
+            row=spectral_window,
+            skip=["ASSOC_NATURE", "ASSOC_SPW_ID"],
+        )
         sf = _source_and_field_table(ms, spectral_window, ifield)
         if sf is None:
             print(f"Field {ifield} cannot be found in SOURCE table")
             observations.append(None)
             continue
         auxtables = {**auxtables, **sf}
-        source_name = auxtables['SOURCE']['NAME'][0]
+        source_name = auxtables["SOURCE"]["NAME"][0]
         if field is not None and source_name != field:
             continue
         print(f"Work on Field {ifield}: {source_name}")
-        mm = read_ms_i(ms, data_column, ifield, spectral_window, pol_ind, pol_summation,
-                       with_calib_info, channels, ignore_flags)
+        mm = read_ms_i(
+            ms,
+            data_column,
+            ifield,
+            spectral_window,
+            pol_ind,
+            pol_summation,
+            with_calib_info,
+            channels,
+            ignore_flags,
+        )
         if mm is None:
             print(f"{ms}, field #{ifield} is empty or fully flagged")
             observations.append(None)
             continue
         # POINTING table is ignored for now
         antpos = AntennaPositions(mm["uvw"], mm["ant1"], mm["ant2"], mm["time"])
-        obs = Observation(antpos, mm["vis"], mm["wgt"], polobj, mm["freq"],
-                          auxiliary_tables=auxtables)
+        obs = Observation(
+            antpos, mm["vis"], mm["wgt"], polobj, mm["freq"], auxiliary_tables=auxtables
+        )
         observations.append(obs)
     return observations
 
@@ -185,8 +209,17 @@ def _determine_weighting(ms):
     return fullwgt, weightcol
 
 
-def read_ms_i(name, data_column, field, spectral_window, pol_indices, pol_summation,
-              with_calib_info, channels, ignore_flags):
+def read_ms_i(
+    name,
+    data_column,
+    field,
+    spectral_window,
+    pol_indices,
+    pol_summation,
[…]
