### 2025-03-20 3b48545
miniparser_fixes: add constants to test.
