### 2025-03-21 d12ba2a
miniparser_fixes: raise error if n_dof are calculated from a likelhood without a proper domain.
