### 2025-03-24 c165a76
88-new-mf-model: add new multi-frequency model. Missing frequency deviations.
