+++ src/spectral_behavior.py
@@ -1,0 +1,253 @@
+# Copyright(C) 2025
+# SPDX-License-Identifier: GPL-2.0+ OR BSD-2-Clause
+# Authors:  Julian Rüstig
+
+from abc import ABC, abstractmethod
+from functools import reduce
+
+import jax.numpy as jnp
+from jax.typing import ArrayLike
+
+from .scaled_excitations import ScaledExcitations
+
+from nifty8.re.model import Model
+from nifty8.re.tree_math.vector import Vector
+
+
+class HarmonicLogSpectralBehavior(Model, ABC):
+    @property
+    @abstractmethod
+    def relative_log_frequencies(self):
+        '''The spectral axis is fixed to be the first index of the output of
+        the model. The `relative_log_frequencies` cast the output arrays of the
+        models to conform with this convention by following the numpy casting
+        conventions.'''
+        pass
+
+    @abstractmethod
+    def mean(self, p) -> list[ArrayLike]:
+        pass
+
+    @abstractmethod
+    def fluctuations(self, p) -> list[ArrayLike]:
+        pass
+
+    @abstractmethod
+    def fluctuations_with_frequencies(self, parameters) -> ArrayLike:
+        pass
+
+    @abstractmethod
+    def mean_with_frequencies(self, parameters) -> ArrayLike:
+        pass
+
+    @abstractmethod
+    def remove_degeneracy_of_spectral_deviations(
+        self,
+        deviations: ArrayLike
+    ) -> ArrayLike:
+        pass
+
+
+class SingleHarmonicLogSpectralBehavior(HarmonicLogSpectralBehavior, ABC):
+    '''Abstract base class for a model with single spectral behavior.
+    Hence, mean is a single number and fluctations is an Array.
+
+    Note
+    ----
+    This can be used to speed up some calculations in the
+    `CorrelatedMultiFrequencySky` class.
+    '''
+    pass
+
+
+class SpectralIndex(SingleHarmonicLogSpectralBehavior):
+    def __init__(
+        self,
+        log_frequencies: ArrayLike,
+        mean: Model,
+        spectral_scaled_excitations: ScaledExcitations,
+        reference_frequency_index: int,
+    ):
+        '''Spectral Index spectral behavior. Special case of the
+        `SpectralPolynomial`. However, since it has a single parameter, it can
+        be used to speed up evaluations.
+
+        .. math::
+            \\log F(\\nu, A, ) = A^{spec.}
+            * \\left\\{ \\mathrm{fluc.}^A \\xi^A \\nu  \\right\\} 
+            + \\mathrm{mean}^A \\nu
+
+        Parameters
+        ----------
+        log_frequencies: ArrayLike
+            The log of the frequencies. The relative frequencies are calculated
+            internally. See `reference_frequency_index`.
+        mean: Model
+            The means of the polynomial model. See equation above.
+        spectral_scaled_excitations: Model
+            The fluctuations already applied to the xis. Hence, the
+            combination of \\mathrm{fluc.}^A \\xi^A.
+        reference_frequency_index: int
+            The index of the reference frequency. Used in order to calculate
+            the relative frequencies.
+        '''
+        # NOTE : The model assumes that spectral axis is fixed to be the first
+        # index of the output model. The slicing tuple matches the shape of the
+        # relative_log_frequencies using the numpy slicing convention.
+        slicing_tuple = (
+            (slice(None),) + (None,)*len(spectral_scaled_excitations.target.shape))
+        self._relative_log_frequencies = (
+            log_frequencies - log_frequencies[reference_frequency_index]
+        )[slicing_tuple]
+
+        self._mean = mean
+        self.spectral_scaled_excitations = spectral_scaled_excitations
+
[…]
