### 2025-03-24 50c6947
88-new-mf-model: add nd_gauss_markov for the spectral deviations.
