+++ src/gauss_markov.py
@@ -1,0 +1,199 @@
+#!/usr/bin/env python3
+
+# SPDX-License-Identifier: BSD-2-Clause
+
+from functools import partial
+from typing import Callable, Union, Optional
+
+import jax.numpy as jnp
+import numpy as np
+from jax import Array, vmap
+from jax.tree_util import tree_map
+
+from nifty8.re.model import Initializer, LazyModel, Model
+from nifty8.re.prior import LogNormalPrior, NormalPrior
+from nifty8.re.tree_math import ShapeWithDtype, random_like
+
+
+def _isscalar(x):
+    return jnp.ndim(x) == 0
+
+
+
+def nd_wiener_process(
+        xi: Array,
+        x0: Array,
+        sigma: Union[float, Array],
+        dt: Array,
+):
+    """Implements the Wiener process (WP)."""
+    amp = jnp.sqrt(dt) * sigma * xi
+    return jnp.cumsum(jnp.vstack((x0[None, ...], amp)), axis=0)
+
+
+class NdGaussMarkovProcess(Model):
+    def __init__(
+        self,
+        process: Callable,
+        x0: Union[float, Array, LazyModel],
+        dt: Union[float, Array],
+        name='xi',
+        N_steps: Optional[int] = None,
+        **kwargs
+    ):
+        if _isscalar(dt):
+            if N_steps is None:
+                msg = "`N_steps` is None and `dt` is not a sequence"
+                raise NotImplementedError(msg)
+            dt = np.ones(N_steps) * dt
+        shp = dt.shape + jnp.shape(
+            x0.target if isinstance(x0, LazyModel) else x0
+        )
+        domain = {name: ShapeWithDtype(shp)}
+        init = Initializer(
+            tree_map(lambda x: partial(random_like, primals=x), domain)
+        )
+        if isinstance(x0, LazyModel):
+            domain = domain | x0.domain
+            init = init | x0.init
+        self.x0 = x0
+        for _, a in kwargs.items():
+            if isinstance(a, LazyModel):
+                domain = domain | a.domain
+                init = init | a.init
+        self.kwargs = kwargs
+        self.name = name
+        self.process = process
+        slicing_tuple = (slice(None),) + (None,) * len(x0.shape)
+        self.dt = jnp.array(dt)[slicing_tuple]
+
+        super().__init__(domain=domain, init=init)
+
+    def __call__(self, x):
+        xi = x[self.name]
+        xx = self.x0(x) if isinstance(self.x0, LazyModel) else self.x0
+        tmp = {
+            k: a(x) if isinstance(a, LazyModel) else a
+            for k, a in self.kwargs.items()
+        }
+        return self.process(xi=xi, x0=xx, dt=self.dt, **tmp)
+
+
+def build_wiener_process(
+    x0: Union[tuple, float, Array, LazyModel],
+    sigma: Union[tuple, float, Array, LazyModel],
+    dt: Union[float, Array],
+    name: str = 'wp',
+    n_steps: int = None
+) -> NdGaussMarkovProcess:
+    """Implements the Wiener process (WP).
+
+    The WP in continuous time takes the form:
+
+    .. math::
+        d/dt x_t = sigma xi_t ,
+
+    where `xi_t` is continuous time white noise.
+
+    Parameters:
+    -----------
+    x0: tuple, float, or LazyModel
+        Initial position of the WP. Can be passed as a fixed value, or a
+        generative Model. Passing a tuple is a shortcut to set a normal prior
+        with mean and std equal to the first and second entry of the tuple
+        respectively on `x0`.
+    sigma: tuple, float, Array, LazyModel
+        Standard deviation of the WP. Analogously to `x0` may also be passed on
[…]
