### 2025-03-24 463f1fd
88-new-mf-model: add tests, refactoring.
