+++ test/test_mf_model.py
@@ -1,0 +1,250 @@
+#!/usr/bin/env python3
+
+# SPDX-License-Identifier: GPL-2.0+ OR BSD-2-Clause
+# Authors: Matteo Guardiani & Julian Rüstig
+
+import jax.random as random
+import pytest
+from numpy.testing import assert_allclose
+import numpy as np
+
+import nifty8.re as jft
+import jubik0 as ju
+
+from nifty8.re.correlated_field import make_grid, NonParametricAmplitude
+from jubik0.frequency_deviations import build_frequency_deviations_model_with_degeneracies
+from jubik0.spectral_behavior import SpectralIndex
+from jubik0.mf_model import build_scaled_excitations
+
+pmp = pytest.mark.parametrize
+
+# TODO: reintroduce this test when amplitudes are implemented
+# @pmp("shape", [(10,), (10, 10)])
+# @pmp("settings, amplitude_model",
+#     [
+#         (dict(
+#             fluctuations=(1.0, 0.02),
+#             loglogavgslope=(-2.0, 0.1),
+#             flexibility=None,
+#             asperity=None,),
+#          "non_parametric",
+#         ),
+#         (dict(
+#             scale=(1.0, 0.02),
+#             cutoff=(0.1, 0.01),
+#             loglogslope=(-4, 0.1),),
+#         "matern")
+#     ]
+# )
+# @pmp("renormalize_amplitude", [True, False])
+# @pmp("prefix", ["test", "", None])
+# @pmp("kind", ["amplitude", "power"])
+# def test_build_amplitude_model(
+#     shape,
+#     settings,
+#     amplitude_model,
+#     renormalize_amplitude,
+#     prefix,
+#     kind,
+# ):
+#     dist = .1
+#     grid = _make_grid(shape, dist, "fourier")
+#     amplitude = build_amplitude_model(grid,
+#                                       settings,
+#                                       amplitude_model,
+#                                       renormalize_amplitude,
+#                                       prefix,
+#                                       kind,
+#                                       )
+#     assert amplitude
+#     assert amplitude.domain
+
+
+@pmp("shape", [(10,), (10, 10)])
+@pmp("distances", [0.1])
+@pmp("log_frequencies", [np.array((0.1,))])
+@pmp("reference_frequency_index", [0])
+@pmp("zero_mode", [jft.Model(lambda p: 0., domain={"zero_mode": None})])
+@pmp("spectral_index_mean", [jft.NormalPrior(0., 1., name="spectral_index_mean")])
+@pmp("spectral_index_fluctuations", [jft.LogNormalPrior(0.1, 1., name="spectral_index_fluctuations")])
+@pmp("deviations_settings", [dict(process='wiener', sigma=(1., 0.1),), None])
+def test_correlated_multi_frequency_sky_init(
+    shape,
+    distances,
+    log_frequencies,
+    reference_frequency_index,
+    zero_mode,
+    spectral_index_mean,
+    spectral_index_fluctuations,
+    deviations_settings,
+):
+    grid = make_grid(shape, distances, "fourier")
+    spatial_amplitude = NonParametricAmplitude(
+        grid, jft.lognormal_prior(.1, .01), None)
+
+    spatial_fluctuations = jft.LogNormalPrior(
+        0.1, 10., name="spatial_fluctuations")
+
+    spectral_behavior = SpectralIndex(
+        log_frequencies=log_frequencies,
+        mean=spectral_index_mean,
+        spectral_scaled_excitations=spectral_index_fluctuations,
+        reference_frequency_index=reference_frequency_index,
+    )
+
[…]
