### 2025-03-24 5e33e75
resolve: revert -> reverse
