+++ src/instruments/resolve/data/data_modify/reverse_frequencies.py
@@ -2,9 +2,9 @@ from ..observation import Observation
 from nifty8.logger import logger
 
 
-def revert_frequencies(obs: Observation) -> Observation:
-    """This reverts the frequencies and returns an observation"""
-    logger.info("Reverting frequencies")
+def reverse_frequencies(obs: Observation) -> Observation:
+    """This reverses the frequencies and returns an observation"""
+    logger.info("Reverse frequencies")
     return Observation(
         obs.antenna_positions,
         obs.vis.val[:, :, ::-1],
