### 2025-03-24 89c3597
resolve: restrict and average to stokes I, are two different things
