### 2025-03-24 9f9895e
resolve: restrict and average to stokes I, are two different things
