+++ demos/configs/cygnusa_2ghz.cfg
@@ -32,6 +32,7 @@ data path = ./data/resolve_test/
 data templates = CYG-ALL-2052-2MHZ_RESOLVE_float64.npz
 data testing percentage = 0.1
 restrict to stokes I = True
+average to stokes I = True
 data weight modify percentage = 0.1
 
 backend = ducc0
