### 2025-03-25 c373944
mf_model: rework paths and stuff
