+++ src/sky_model/multifrequency/gauss_markov.py
@@ -3,13 +3,12 @@
 # SPDX-License-Identifier: BSD-2-Clause
 
 from functools import partial
-from typing import Callable, Union, Optional
+from typing import Callable, Optional, Union
 
 import jax.numpy as jnp
 import numpy as np
 from jax import Array, vmap
 from jax.tree_util import tree_map
-
 from nifty8.re.model import Initializer, LazyModel, Model
 from nifty8.re.prior import LogNormalPrior, NormalPrior
 from nifty8.re.tree_math import ShapeWithDtype, random_like
@@ -19,12 +18,11 @@ def _isscalar(x):
     return jnp.ndim(x) == 0
 
 
-
 def nd_wiener_process(
-        xi: Array,
-        x0: Array,
-        sigma: Union[float, Array],
-        dt: Array,
+    xi: Array,
+    x0: Array,
+    sigma: Union[float, Array],
+    dt: Array,
 ):
     """Implements the Wiener process (WP)."""
     amp = jnp.sqrt(dt) * sigma * xi
@@ -37,22 +35,18 @@ class NdGaussMarkovProcess(Model):
         process: Callable,
         x0: Union[float, Array, LazyModel],
         dt: Union[float, Array],
-        name='xi',
+        name="xi",
         N_steps: Optional[int] = None,
-        **kwargs
+        **kwargs,
     ):
         if _isscalar(dt):
             if N_steps is None:
                 msg = "`N_steps` is None and `dt` is not a sequence"
                 raise NotImplementedError(msg)
             dt = np.ones(N_steps) * dt
-        shp = dt.shape + jnp.shape(
-            x0.target if isinstance(x0, LazyModel) else x0
-        )
+        shp = dt.shape + jnp.shape(x0.target if isinstance(x0, LazyModel) else x0)
         domain = {name: ShapeWithDtype(shp)}
-        init = Initializer(
-            tree_map(lambda x: partial(random_like, primals=x), domain)
-        )
+        init = Initializer(tree_map(lambda x: partial(random_like, primals=x), domain))
         if isinstance(x0, LazyModel):
             domain = domain | x0.domain
             init = init | x0.init
@@ -73,8 +67,7 @@ class NdGaussMarkovProcess(Model):
         xi = x[self.name]
         xx = self.x0(x) if isinstance(self.x0, LazyModel) else self.x0
         tmp = {
-            k: a(x) if isinstance(a, LazyModel) else a
-            for k, a in self.kwargs.items()
+            k: a(x) if isinstance(a, LazyModel) else a for k, a in self.kwargs.items()
         }
         return self.process(xi=xi, x0=xx, dt=self.dt, **tmp)
 
@@ -83,8 +76,8 @@ def build_wiener_process(
     x0: Union[tuple, float, Array, LazyModel],
     sigma: Union[tuple, float, Array, LazyModel],
     dt: Union[float, Array],
-    name: str = 'wp',
-    n_steps: int = None
+    name: str = "wp",
+    n_steps: int = None,
 ) -> NdGaussMarkovProcess:
     """Implements the Wiener process (WP).
 
@@ -121,9 +114,9 @@ def build_wiener_process(
     i.E. `sigma_t = sigma_i for t_i <= t < t_{i+1}`.
     """
     if isinstance(x0, tuple):
-        x0 = NormalPrior(x0[0], x0[1], name=name + '_x0')
+        x0 = NormalPrior(x0[0], x0[1], name=name + "_x0")
     if isinstance(sigma, tuple):
-        sigma = LogNormalPrior(sigma[0], sigma[1], name=name + '_sigma')
[…]
