### 2025-03-25 e268a15
mf-model: Rename the CorrelatedMF -> SpectralProductMFSky
