+++ src/__init__.py
@@ -28,7 +28,7 @@ from .instruments.chandra.chandra_data import (generate_chandra_data,
                                                create_chandra_data_from_config)
 from .instruments.chandra.chandra_response import build_chandra_response_from_config
 from .instruments.chandra.chandra_likelihood import generate_chandra_likelihood_from_config
-from .sky_model.multifrequency.mf_model import (
+from .sky_model.multifrequency.spectral_product_mf_sky import (
     SpectralProductMFSky,
     build_default_mf_model,
 )
