### 2025-03-25 53edb2d
mf_model: docstring adjust
