### 2025-03-25 180edff
mf_model: delete unneeded add_xi_properties
