+++ src/sky_model/multifrequency/spectral_product_utils/scaled_excitations.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: GPL-2.0+ OR BSD-2-Clause
 # Authors: Matteo Guardiani & Julian Rüstig
 
-from .multifrequency_utils import build_distribution_or_default
+from .distribution_or_default import build_distribution_or_default
 
 from nifty8.re.model import Model
 from nifty8.re.num.stats_distributions import lognormal_prior, normal_prior
