+++ src/sky_model/multifrequency/spectral_product_utils/normalized_amplitude_model.py
@@ -11,7 +11,7 @@ from nifty8.re.correlated_field import (
 )
 from nifty8.re.num.stats_distributions import lognormal_prior, normal_prior
 
-from .multifrequency_utils import check_demands
+from .check_demands import check_demands
 
 
 def _set_default_or_call(arg: Union[callable, tuple, list] | None, default: callable):
