### 2025-03-25 dc3c927
88-new-mf-model: remove unnecessary test.
