### 2025-03-25 59ef07d
jwst: making `ij` indexing work
