### 2025-03-25 698d956
jwst: plotting update
