+++ src/instruments/jwst/plotting/plotting.py
@@ -30,14 +30,15 @@ def get_plot(
     sky_model_with_keys: jft.Model,
     cfg: dict,
     parametric_flag: bool,
-    sky_key: str = 'sky',
+    sky_key: str = "sky",
     residual_ylen_offset: int = 0,
 ):
     if isinstance(filter_projector.domain, jft.ShapeWithDtype):
         sky_model_new = sky_model
     else:
-        sky_model_new = jft.Model(lambda x: sky_model(x)[sky_key],
-                                  domain=sky_model.domain)
+        sky_model_new = jft.Model(
+            lambda x: sky_model(x)[sky_key], domain=sky_model.domain
+        )
 
     ll_alpha, ll_nonpar, sl_alpha, sl_nonpar = get_alpha_nonpar(lens_system)
 
@@ -47,8 +48,9 @@ def get_plot(
             norm_source=LogNorm,
             norm_lens=LogNorm,
             # norm_source_alpha=LogNorm,
-            norm_source_nonparametric=LogNorm,
+            # norm_source_nonparametric=LogNorm,
             # norm_mass=LogNorm,
+            share_source=False,
         ),
         lens_system=lens_system,
         grid=grid,
@@ -60,7 +62,7 @@ def get_plot(
         sky=FieldPlottingConfig(norm=LogNorm),
         data=FieldPlottingConfig(norm=LogNorm),
         display_pointing=False,
-        xmax_residuals=cfg.get('max_residuals', 4),
+        xmax_residuals=cfg.get("max_residuals", 4),
         ylen_offset=residual_ylen_offset,
     )
     plot_residual = build_plot_sky_residuals(
@@ -73,7 +75,10 @@ def get_plot(
     )
 
     plot_color = build_color_components_plotting(
-        lens_system.source_plane_model.light_model.nonparametric, results_directory, substring='source')
+        lens_system.source_plane_model.light_model.nonparametric,
+        results_directory,
+        substring="source",
+    )
 
     plot_source = build_plot_source(
         results_directory,
@@ -88,7 +93,7 @@ def get_plot(
         source_light_alpha=sl_alpha,
         source_light_parametric=lens_system.source_plane_model.light_model.parametric,
         source_light_nonparametric=sl_nonpar,
-        attach_name=''
+        attach_name="",
     )
 
     return plot_source, plot_residual, plot_lens, plot_color
@@ -101,39 +106,37 @@ def plot_prior(
     plot_residuals: bool,
     plot_source: Union[bool, callable],
     plot_lens: Union[bool, callable],
-
     data_dict: dict,
     parametric_flag: bool,
-    sky_key: str = 'sky',
+    sky_key: str = "sky",
     residual_ylen_offset: int = 0,
 ):
     from charm_lensing.lens_system import build_lens_system
-    from jubik0.instruments.jwst.config_handler import (
-        insert_spaces_in_lensing_new)
+    from jubik0.instruments.jwst.config_handler import insert_spaces_in_lensing_new
     import yaml
+
     test_key, _ = random.split(random.PRNGKey(42), 2)
 
-    cfg = yaml.load(open(config_path, 'r'), Loader=yaml.SafeLoader)
-    insert_spaces_in_lensing_new(cfg['sky'])
+    cfg = yaml.load(open(config_path, "r"), Loader=yaml.SafeLoader)
+    insert_spaces_in_lensing_new(cfg["sky"])
 
-    lens_system = build_lens_system(cfg['sky'])
-    if cfg['nonparametric_lens']:
+    lens_system = build_lens_system(cfg["sky"])
+    if cfg["nonparametric_lens"]:
         sky_model = lens_system.get_forward_model_full()
         parametric_flag = False
     else:
         sky_model = lens_system.get_forward_model_parametric()
         parametric_flag = True
[…]
