### 2025-03-26 1d670e0
88-mf-model: wrong order in building the NonParametricAmplitude (test)
