+++ src/instruments/jwst/parse/rotation_and_shift/coordinates_correction.py
@@ -5,18 +5,20 @@
 
 # %
 from ..parametric_model.parametric_prior import (
-    ProbabilityConfig, transform_setting_to_prior_config)
+    ProbabilityConfig,
+    transform_setting_to_prior_config,
+)
 
 from astropy import units as u
 from dataclasses import dataclass
 from typing import Optional
 
 
-ROTATION_UNIT_KEY = 'rotation_unit'
-ROTATION_UNIT_DEFAULT = 'rad'
-SHIFT_UNIT_KEY = 'shift_unit'
-SHIFT_UNIT_DEFAULT = 'arcsec'
-DEFAULT_KEY = 'default'
+ROTATION_UNIT_KEY = "rotation_unit"
+ROTATION_UNIT_DEFAULT = "rad"
+SHIFT_UNIT_KEY = "shift_unit"
+SHIFT_UNIT_DEFAULT = "arcsec"
+DEFAULT_KEY = "default"
 
 NOT_FILTER_KEYS = [
     ROTATION_UNIT_KEY,
@@ -26,8 +28,8 @@ NOT_FILTER_KEYS = [
     DEFAULT_KEY,
 ]
 
-SHIFT_KEY = 'shift'
-ROTATION_KEY = 'rotation'
+SHIFT_KEY = "shift"
+ROTATION_KEY = "rotation"
 
 
 @dataclass
@@ -39,71 +41,80 @@ class CoordiantesCorrectionPriorConfig:
 
     def shift_in(self, unit: u.Unit):
         priors = self.shift.to_list()
-        priors[1] = (priors[1]*self.shift_unit).to(unit).value
-        if len(priors) == 4:
-            priors[2] = (priors[2]*self.shift_unit).to(unit).value
+        priors[1] = (priors[1] * self.shift_unit).to(unit).value
+        if priors[2] is not None:
+            priors[2] = (priors[2] * self.shift_unit).to(unit).value
         return transform_setting_to_prior_config(priors)
 
     def rotation_in(self, unit: u.Unit):
         priors = self.rotation.to_list()
-        priors[1] = (priors[1]*self.rotation_unit).to(unit).value
-        if len(priors) == 4:
-            priors[2] = (priors[2]*self.rotation_unit).to(unit).value
+        priors[1] = (priors[1] * self.rotation_unit).to(unit).value
+        if priors[2] is not None:
+            priors[2] = (priors[2] * self.rotation_unit).to(unit).value
         return transform_setting_to_prior_config(priors)
 
 
 @dataclass
 class CoordiantesCorrectionConfig:
-    '''This class saves the `CoordiantesCorrectionPrior`s for the coordinate
+    """This class saves the `CoordiantesCorrectionPrior`s for the coordinate
     correction. Specifics can be provided for the different filters and their
     multiple datasets (dithers, etc.). If a specific filter gets a None, the
     coordinates from the data are fixed to the provided values.
-    '''
-    default: CoordiantesCorrectionPriorConfig
-    filters: dict[str, list[Optional[CoordiantesCorrectionPriorConfig]]]
+    """
 
-    def get_filter_or_default(self, filter_name: str, data_index: int) -> ProbabilityConfig:
-        '''Returns the PriorConfig for the `filter_name` and `data_index` or the default.
+    default: CoordiantesCorrectionPriorConfig
+    name_settings: dict[str, list[Optional[CoordiantesCorrectionPriorConfig]]]
+
+    def get_name_setting_or_default(
+        self, filter_name: str, data_index: int | None = None
+    ) -> ProbabilityConfig:
+        """Returns the PriorConfig for the `filter_name` and `data_index` or the default.
 
         Parameters
         ----------
         filter_name : str
             The filter in question.
-        '''
+        """
[…]
