### 2025-03-28 267dbf4
jwst: world_location_to_index & index_to_world_location
