### 2025-03-28 06a95e0
jwst: restructure inheritence & world_to_pixel, etc.
