### 2025-03-31 3fc4173
jwst: comments, & constant mode in finufft interpolation
