### 2025-03-31 04a8d31
jwst: renaming subsample_centers, & further cosmetics
