+++ src/instruments/jwst/jwst_data.py
@@ -8,7 +8,7 @@
 from ...color import Color, ColorRange
 from ...grid import Grid
 from ...wcs.wcs_jwst_data import WcsJwstData
-from ...wcs.wcs_subsample_centers import subsample_grid_centers_in_index_grid_non_vstack
+from ...wcs.wcs_subsample_centers import subsample_grid_centers_in_index_grid
 from .masking import get_mask_from_index_centers
 
 from astropy import units
@@ -238,7 +238,7 @@ def load_jwst_data_mask_std(
     # TODO: Use a simpler algorithm. Check that its consistent with the
     # subsampling of the rotation and shift model.
 
-    centers = subsample_grid_centers_in_index_grid_non_vstack(
+    centers = subsample_grid_centers_in_index_grid(
         world_extrema=world_corners,
         to_be_subsampled_grid_wcs=jwst_data.wcs,
         index_grid_wcs=grid.spatial,
