+++ src/instruments/jwst/integration/__init__.py
@@ -4,5 +4,3 @@
 # Copyright(C) 2024 Max-Planck-Society
 
 # %%
-
-from .sum_integration import build_sum
