### 2025-03-31 1b2f760
jwst: restructure rotation_and_shift config parsing
