### 2025-03-31 6bb1b51
jwst: remove flux conversion from rotation&shift
