+++ src/instruments/jwst/rotation_and_shift/nufft_rotation_and_shift.py
@@ -16,35 +16,19 @@ from functools import reduce
 
 
 def build_nufft_rotation_and_shift(
-    sky_dvol: ArrayLike,
-    sub_dvol: ArrayLike,
     sky_shape: tuple[int, int],
-    sky_distances: Union[tuple[Quantity, Quantity], tuple[float, float]],
     out_shape: tuple[int, int],
-    indexing: str = 'ij',
-    sky_as_brightness: bool = False
+    indexing: str = "ij",
     mode: str = "constant",
 ) -> Callable[[ArrayLike], ArrayLike]:
-    """
-    Builds non-uniform FFT interpolation model.
+    """Builds non-uniform FFT interpolation model.
 
     Parameters
     ----------
-    sky_dvol: float
-        The volume of the sky/reconstruction pixels
-    sub_dvol: float
-        The volume of the subsample pixels.
-        Typically, the data pixel is subsampled.
     sky_shape: Tuple[int, int]
         The shape of the reconstruction array (sky shape)
-    sky_distances: Tuple[Quantity, Quantity]
-        The sky_distances are needed to check for the consistency of the
-        `xy_conversion`.
     out_shape: Tuple[int, int]
         The shape of the subsample array.
-    sky_as_brightness: bool
-        If True, the sky will be treated as a brightness distribution.
-        This is the same as setting sky_dvol = 1.
     mode: str
         The mode of the interpolation. ['wrap', 'constant']
 
@@ -60,18 +44,6 @@ def build_nufft_rotation_and_shift(
         flux(x, y) = sky(x, y) * sky_dvol
     """
     from jax_finufft import nufft2
-
-    if isinstance(sky_distances[0], Quantity):
-        assert reduce(lambda x, y: x.value == y.value, sky_distances)
-    else:
-        assert reduce(lambda x, y: x == y, sky_distances)
-    assert len(sky_distances) == len(out_shape) == len(sky_shape) == 2
-
-    # The conversion factor from sky to subpixel
-    # (flux = sky_brightness * flux_conversion)
-    if sky_as_brightness:
-        sky_dvol = 1
-    flux_conversion = sub_dvol / sky_dvol
 
     xy_conversion = 2 * np.pi / np.array(sky_shape)[:, None]
 
@@ -94,12 +66,11 @@ def build_nufft_rotation_and_shift(
                 raise ValueError("mode must either be `wrap` or `constant`.")
 
             out = nufft2(f_field, xy_finufft[0], xy_finufft[1]).real
-            return reshape(out, out_shape) * flux_conversion
-    elif indexing == 'xy':
 
             if mode == "constant":
                 out = jnp.where(mask, 0.0, out)
 
+            return jnp.reshape(out, out_shape)
 
     elif indexing == "xy":
         out_shape = out_shape[1], out_shape[0]
@@ -115,12 +86,11 @@ def build_nufft_rotation_and_shift(
                 raise ValueError("mode must either be `wrap` or `constant`.")
 
             out = nufft2(f_field, xy_finufft[1], xy_finufft[0]).real
-            return reshape(out, out_shape).T * flux_conversion
 
             if mode == "constant":
                 out = jnp.where(mask, 0.0, out)
 
-            return jnp.reshape(out, out_shape)  # * flux_conversion
+            return jnp.reshape(out, out_shape)
     else:
         raise ValueError("Need either provide `ij` or `xy` indexing.")
 
