### 2025-03-31 6661c3d
jwst: introducing flux_conversion as part of JWST model
