+++ src/instruments/jwst/jwst_response.py
@@ -5,26 +5,27 @@
 
 # %
 
-from .integration_model import build_sum
-from .jwst_psf import build_psf_operator
-from .rotation_and_shift import build_rotation_and_shift_model, RotationAndShiftModel
-from .rotation_and_shift.coordinates_correction import (
-    build_coordinates_correction_from_grid,
-)
-from .zero_flux_model import build_zero_flux_model
-from .masking.build_mask import build_mask
+from typing import Callable, Optional
+
+import nifty8.re as jft
+from astropy import units as u
+from numpy.typing import ArrayLike
 
 from ...wcs import subsample_grid_centers_in_index_grid
-
+from .integration.unit_conversion import build_unit_conversion
+from .integration.integration import integration_factory
+from .jwst_psf import build_psf_operator
+from .masking.build_mask import build_mask
 from .parse.parametric_model.parametric_prior import ProbabilityConfig
 from .parse.rotation_and_shift.coordinates_correction import (
     CoordiantesCorrectionPriorConfig,
 )
-
-from typing import Callable, Optional
-
-import nifty8.re as jft
-from numpy.typing import ArrayLike
+from .rotation_and_shift import RotationAndShiftModel, build_rotation_and_shift_model
+from .rotation_and_shift.coordinates_correction import (
+    build_coordinates_correction_from_grid,
+)
+from .zero_flux_model import build_zero_flux_model
+from ...grid import Grid
 
 
 class JwstResponse(jft.Model):
@@ -40,11 +41,12 @@ class JwstResponse(jft.Model):
     def __init__(
         self,
         sky_domain: dict,
-        rotation_and_shift: Optional[RotationAndShiftModel],
+        rotation_and_shift: RotationAndShiftModel,
         psf: Callable[[ArrayLike], ArrayLike],
+        unit_conversion: Callable[[ArrayLike], ArrayLike],
         integrate: Callable[[ArrayLike], ArrayLike],
         transmission: float,
-        zero_flux_model: Optional[jft.Model],
+        zero_flux_model: jft.Model | None,
         mask: Callable[[ArrayLike], ArrayLike],
     ):
         """
@@ -63,6 +65,8 @@ class JwstResponse(jft.Model):
         psf : callable
             A function that applies a point spread function (PSF) to the
             input data.
+        unit_conversion : callable
+            A function that transforms the unit of the sky to the data unit.
         integrate : callable
             A function that performs integration on the input data.
         transmission : float
@@ -85,6 +89,7 @@ class JwstResponse(jft.Model):
 
         self.rotation_and_shift = rotation_and_shift
         self.psf = psf
+        self.unit_conversion = unit_conversion
         self.integrate = integrate
         self.transmission = transmission
         self.zero_flux_model = zero_flux_model
@@ -98,6 +103,7 @@ class JwstResponse(jft.Model):
     def __call__(self, x):
         out = self.rotation_and_shift(x)
         out = self.psf(out)
+        out = self.unit_conversion(out)
         out = self.integrate(out)
         out = out * self.transmission
         if self.zero_flux_model is not None:
@@ -110,12 +116,14 @@ def build_jwst_response(
     sky_domain: dict,
     data_identifier: str,
     data_subsample: int,
-    rotation_and_shift_kwargs: Optional[dict],
-    shift_and_rotation_correction_prior: Optional[CoordiantesCorrectionPriorConfig],
-    psf_kernel: Optional[ArrayLike],
+    rotation_and_shift_kwargs: dict | None,
[…]
