+++ demos/lens_jwst.py
@@ -1,69 +1,69 @@
 import argparse
+from sys import exit
 
-import nifty8.re as jft
-from jax import random, config
 import jax.numpy as jnp
-
+import nifty8.re as jft
 import numpy as np
+from astropy import units as u
+from charm_lensing.lens_system import build_lens_system
+from jax import config, random
 from matplotlib.colors import LogNorm
 
 import jubik0 as ju
-
-from charm_lensing.lens_system import build_lens_system
-# from charm_lensing.physical_models.multifrequency_models.nifty_mf import build_nifty_mf_from_grid
-
-from jubik0.instruments.jwst.config_handler import load_yaml_and_save_info
-from jubik0.instruments.jwst.config_handler import (
-    insert_spaces_in_lensing_new)
-from jubik0.likelihood import connect_likelihood_to_model
-from jubik0.instruments.jwst.jwst_likelihoods import build_jwst_likelihoods
-
-from jubik0.instruments.jwst.plotting.plotting import get_plot, plot_prior
-
-from jubik0.parse.grid import GridModel
 from jubik0.grid import Grid
 
-from sys import exit
+# from charm_lensing.physical_models.multifrequency_models.nifty_mf import build_nifty_mf_from_grid
+from jubik0.instruments.jwst.config_handler import (
+    insert_spaces_in_lensing_new,
+    load_yaml_and_save_info,
+)
+from jubik0.instruments.jwst.jwst_likelihoods import build_jwst_likelihoods
+from jubik0.instruments.jwst.plotting.plotting import get_plot, plot_prior
+from jubik0.likelihood import connect_likelihood_to_model
+from jubik0.parse.grid import GridModel
 
-SKY_KEY = 'sky'
-config.update('jax_enable_x64', True)
+SKY_KEY = "sky"
+SKY_UNIT = u.MJy / u.sr
+config.update("jax_enable_x64", True)
 
 parser = argparse.ArgumentParser()
 parser.add_argument(
     "config",
     help="Config File",
     type=str,
-    nargs='?',
+    nargs="?",
     const=1,
-    default='./demos/configs/spt0418.yaml')
+    default="./demos/configs/spt0418.yaml",
+)
 args = parser.parse_args()
 config_path = args.config
 
 cfg, results_directory = load_yaml_and_save_info(config_path)
 
-if cfg['cpu']:
+if cfg["cpu"]:
     from jax import devices
-    config.update('jax_default_device', devices('cpu')[0])
 
-if cfg['no_interactive_plotting']:
+    config.update("jax_default_device", devices("cpu")[0])
+
+if cfg["no_interactive_plotting"]:
     import matplotlib
-    matplotlib.use('Agg')
 
-grid = Grid.from_grid_model(GridModel.from_yaml_dict(cfg['sky']['grid']))
+    matplotlib.use("Agg")
+
+grid = Grid.from_grid_model(GridModel.from_yaml_dict(cfg["sky"]["grid"]))
 
 # insert_ubik_energy_in_lensing(cfg, zsource=4.2)
-insert_spaces_in_lensing_new(cfg['sky'])
-lens_system = build_lens_system(cfg['sky'])
+insert_spaces_in_lensing_new(cfg["sky"])
+lens_system = build_lens_system(cfg["sky"])
 
 
-if cfg['nonparametric_lens']:
+if cfg["nonparametric_lens"]:
     sky_model = lens_system.get_forward_model_full()
     parametric_flag = False
 else:
     sky_model = lens_system.get_forward_model_parametric()
     parametric_flag = True
-sky_model = jft.Model(jft.wrap_left(sky_model, SKY_KEY),
[…]
