### 2025-04-01 bb6792f
jwst: rename world_extrema -> world_corners
