### 2025-04-01 067bdcf
jwst: introducing masking
