### 2025-04-01 7829e90
jwst: getting rid of optional typing
