+++ src/instruments/jwst/parse/parametric_model/parametric_prior.py
@@ -1,16 +1,16 @@
 from abc import ABC
-from typing import Union, Optional
+from typing import Union
 from dataclasses import dataclass
 
-DISTRIBUTION_KEY = 'distribution'
+DISTRIBUTION_KEY = "distribution"
 
 
 class ProbabilityConfig(ABC):
     distribution: str
-    transformation: Optional[str]
+    transformation: str | None
 
     def to_list(self) -> list:
-        '''Return the ProbabilityConfig as a tuple.'''
+        """Return the ProbabilityConfig as a tuple."""
         # NOTE: __annotations__ keeps the order of the fields of the
         # make_dataclass used inside create_distribution_config_class
         return list(getattr(self, field) for field in self.__annotations__.keys())
@@ -21,7 +21,7 @@ class DefaultPriorConfig(ProbabilityConfig):
     distribution: str
     mean: float
     sigma: float
-    transformation: Optional[str] = None
+    transformation: str | None = None
 
 
 @dataclass
@@ -29,41 +29,44 @@ class UniformPriorConfig(ProbabilityConfig):
     distribution: str
     min: float
     max: float
-    transformation: Optional[str] = None
+    transformation: str | None = None
 
 
 @dataclass
 class DeltaPriorConfig(ProbabilityConfig):
     distribution: str
     mean: float
-    _: Optional[float] = None  # This is not needed however it's convinient to
+    _: float | None = None  # This is not needed however it's convinient to
     # comply with the signature of the other Configs.
-    transformation: Optional[str] = None
+    transformation: str | None = None
 
 
 def transform_setting_to_prior_config(settings: Union[dict, tuple]):
     """Transforms the parameter distribution `settings` into a ProbabilityConfig."""
 
-    distribution = (settings[DISTRIBUTION_KEY] if isinstance(settings, dict)
-                    else settings[0])
-    distribution = (distribution.lower() if type(distribution) == str
-                    else distribution)
+    distribution = (
+        settings[DISTRIBUTION_KEY] if isinstance(settings, dict) else settings[0]
+    )
+    distribution = distribution.lower() if type(distribution) == str else distribution
 
     match distribution:
-        case 'uniform':
+        case "uniform":
             return (
-                UniformPriorConfig(**settings) if isinstance(settings, dict)
+                UniformPriorConfig(**settings)
+                if isinstance(settings, dict)
                 else UniformPriorConfig(*settings)
             )
 
-        case 'delta' | None:
+        case "delta" | None:
             return (
-                DeltaPriorConfig(**settings) if isinstance(settings, dict)
+                DeltaPriorConfig(**settings)
+                if isinstance(settings, dict)
                 else DeltaPriorConfig(*settings)
             )
 
         case _:
             return (
-                DefaultPriorConfig(**settings) if isinstance(settings, dict)
+                DefaultPriorConfig(**settings)
+                if isinstance(settings, dict)
                 else DefaultPriorConfig(*settings)
             )
