+++ src/instruments/jwst/parse/jwst_psf.py
@@ -1,16 +1,16 @@
 from dataclasses import dataclass
 
 
-WEBBPSF_PATH_KEY = 'webbpsf_path'
-PSF_LIBRARY_PATH_KEY = 'psf_library'
-PSF_ARCSEC_KEY = 'psf_arcsec_extension'
-NORMALIZE_KEY = 'normalize'
-NORMALIZE_DEFAULT = 'last'
+WEBBPSF_PATH_KEY = "webbpsf_path"
+PSF_LIBRARY_PATH_KEY = "psf_library"
+PSF_ARCSEC_KEY = "psf_arcsec_extension"
+NORMALIZE_KEY = "normalize"
+NORMALIZE_DEFAULT = "last"
 
 
 @dataclass
 class PsfKernelConfig:
-    '''The PsfKernelConfig is a data model for holding metadata for
+    """The PsfKernelConfig is a data model for holding metadata for
     the evaluation of the psf kernel.
 
     webbpsf_path : str
@@ -24,7 +24,8 @@ class PsfKernelConfig:
     normalize : str
         The normalization method for the PSF. Default is 'last',
         but other methods may be supported by `webbpsf`.
-    '''
+    """
+
     webbpsf_path: str
     psf_library_path: str
     psf_arcsec: float
@@ -34,15 +35,15 @@ class PsfKernelConfig:
 def yaml_to_psf_kernel_config(
     psf_config: dict,
 ):
-    '''Read the PsfKernelConfig from the yaml config.
+    """Read the PsfKernelConfig from the yaml config.
 
-    psf_config: dict
-        The dictionary holding:
-            - webbpsf_path
-            - psf_library_path
-            - psf_arcsec_extension
-            - normalize (Optional)
-    '''
+        psf_config: dict
+            The dictionary holding:
+                - webbpsf_path
+                - psf_library_path
+                - psf_arcsec_extension
+                - normalize | None
+    j"""
 
     return PsfKernelConfig(
         webbpsf_path=psf_config[WEBBPSF_PATH_KEY],
