### 2025-04-12 32ed649
jwst: test parametric source lensed
