### 2025-04-14 d7f0851
jwst: plotting update
