+++ src/instruments/jwst/plotting/plotting.py
@@ -1,69 +1,59 @@
-from .jwst_plotting import (
-    get_alpha_nonpar,
-    build_plot_lens_system,
-    build_color_components_plotting,
-    build_plot_sky_residuals,
-    build_plot_source,
-    FieldPlottingConfig,
-    ResidualPlottingConfig,
-)
-from ..filter_projector import FilterProjector
+from typing import Union
+from os.path import join
+
+import matplotlib.pyplot as plt
+import nifty8.re as jft
+import numpy as np
+from jax import random
+
 from ....grid import Grid
 from ....parse.grid import GridModel
-
-import nifty8.re as jft
-
-from charm_lensing.lens_system import LensSystem
-
-from jax import random
-from matplotlib.colors import LogNorm
-
-from typing import Union
+from ..filter_projector import FilterProjector
+from .plotting_base import find_closest_factors, get_position_or_samples_of_model
+from .residuals import build_plot_sky_residuals
+from ..parse.plotting import (
+    LensSystemPlottingConfig,
+    ResidualPlottingConfig,
+    FieldPlottingConfig,
+)
+from .plotting_lens_system import build_plot_lens_system
+from .plot_source import build_plot_source
 
 
 def get_plot(
     results_directory: str,
     grid: Grid,
-    lens_system: LensSystem,
+    lens_system,
     filter_projector: FilterProjector,
     data_dict: dict,
-    sky_model: jft.Model,
     sky_model_with_keys: jft.Model,
-    cfg: dict,
-    parametric_flag: bool,
-    sky_key: str = "sky",
+    parametric_lens: bool,
+    parametric_source: bool = False,
+    max_residuals: int = 4,
     residual_ylen_offset: int = 0,
 ):
-    if isinstance(filter_projector.domain, jft.ShapeWithDtype):
-        sky_model_new = sky_model
-    else:
-        sky_model_new = jft.Model(
-            lambda x: sky_model(x)[sky_key], domain=sky_model.domain
-        )
+    from charm_lensing.lens_system import LensSystem
 
-    ll_alpha, ll_nonpar, sl_alpha, sl_nonpar = get_alpha_nonpar(lens_system)
+    lens_system: LensSystem = lens_system
 
+    plotting_config_lens = LensSystemPlottingConfig()
+    plotting_config_lens.source.combined.norm = "log"
+    plotting_config_lens.lens_light.combined.norm = "log"
+    plotting_config_lens.share_source_vmin_vmax = False
     plot_lens = build_plot_lens_system(
         results_directory,
-        plotting_config=dict(
-            norm_source=LogNorm,
-            norm_lens=LogNorm,
-            # norm_source_alpha=LogNorm,
-            # norm_source_nonparametric=LogNorm,
-            # norm_mass=LogNorm,
-            share_source=False,
-        ),
+        plotting_config=plotting_config_lens,
         lens_system=lens_system,
         grid=grid,
-        lens_light_alpha_nonparametric=(ll_alpha, ll_nonpar),
-        source_light_alpha_nonparametric=(sl_alpha, sl_nonpar),
+        parametric_lens=parametric_lens,
+        parametric_source=parametric_source,
     )
 
     residual_plotting_config = ResidualPlottingConfig(
-        sky=FieldPlottingConfig(norm=LogNorm),
-        data=FieldPlottingConfig(norm=LogNorm),
+        sky=FieldPlottingConfig(norm="log", vmin=1e-4, vmax=1e2),
+        data=FieldPlottingConfig(norm="log", vmin=1e-3),
         display_pointing=False,
-        xmax_residuals=cfg.get("max_residuals", 4),
+        xmax_residuals=max_residuals,
         ylen_offset=residual_ylen_offset,
     )
[…]
