+++ src/instruments/jwst/plotting/plotting_base.py
@@ -1,0 +1,313 @@
+from typing import Union, Any
+
+import matplotlib.pyplot as plt
+import nifty8.re as jft
+import numpy as np
+
+from nifty8.re.library.mf_model import CorrelatedMultiFrequencySky
+
+from ..parse.plotting import FieldPlottingConfig
+from ..filter_projector import FilterProjector
+from ..rotation_and_shift.coordinates_correction import CoordinatesWithCorrection
+
+
+def plot_data_data_model_residuals(
+    ims: list,
+    axes: list,
+    data_key: str,
+    data: np.ndarray,
+    data_model: np.ndarray,
+    std: np.ndarray,
+    plotting_config: FieldPlottingConfig,
+):
+    """Plot three panels (data, model, data-model).
+
+    Parameters
+    ----------
+
+    """
+    max_d = plotting_config.get_max(data)
+    min_d = plotting_config.get_min(data)
+
+    axes[0].set_title(f"Data {data_key}")
+    axes[1].set_title("Data model")
+    axes[2].set_title("Data - Data model")
+    ims[0] = axes[0].imshow(
+        data,
+        norm=plotting_config.norm,
+        vmin=min_d,
+        vmax=max_d,
+        **plotting_config.rendering,
+    )
+    ims[1] = axes[1].imshow(
+        data_model,
+        norm=plotting_config.norm,
+        vmin=min_d,
+        vmax=max_d,
+        **plotting_config.rendering,
+    )
+    ims[2] = axes[2].imshow(
+        (data - data_model) / std,
+        vmin=-3,
+        vmax=3,
+        cmap="RdBu_r",
+        **plotting_config.rendering,
+    )
+
+    return ims
+
+
+def display_text(ax: plt.Axes, text: dict, **kwargs):
+    """Display text on plot
+    ax: matplotlib axis
+    text: dict or str (default: {'s': str, 'color': 'white'})
+    kwargs:
+    - keyword: str
+        options: 'top_left' (default), 'top_right', 'bottom_left', 'bottom_right'
+    - x_offset_ticker: float (default: 0)
+    - y_offset_ticker: float (default: 0)
+    """
+    keyword = kwargs.get("keyword", "top_left")
+    x_offset_ticker = kwargs.get("x_offset_ticker", 0)
+    y_offset_ticker = kwargs.get("y_offset_ticker", 0)
+
+    if type(text) is str:
+        text = dict(
+            s=text,
+            color="white",
+            bbox=dict(facecolor="black", alpha=0.5, edgecolor="none"),
+        )
+
+    if keyword == "top_left":
+        ax.text(
+            x=0.05 + x_offset_ticker * 0.05,
+            y=0.95 - y_offset_ticker * 0.05,
+            ha="left",
+            va="top",
+            transform=ax.transAxes,
+            **text,
+        )
+    elif keyword == "top_right":
+        ax.text(
+            x=0.95 - x_offset_ticker * 0.05,
+            y=0.95 - y_offset_ticker * 0.05,
+            ha="right",
+            va="top",
+            transform=ax.transAxes,
+            **text,
+        )
+    elif keyword == "bottom_left":
+        ax.text(
+            x=0.05 + x_offset_ticker * 0.05,
+            y=0.05 + y_offset_ticker * 0.05,
+            ha="left",
+            va="bottom",
+            transform=ax.transAxes,
+            **text,
+        )
+    elif keyword == "bottom_right":
+        ax.text(
+            x=0.95 - x_offset_ticker * 0.05,
+            y=0.05 + y_offset_ticker * 0.05,
+            ha="right",
+            va="bottom",
+            transform=ax.transAxes,
+            **text,
+        )
+    else:
[…]
