### 2025-04-14 13aab2a
jwst: use default factory
