### 2025-04-15 e19cf2c
jwst: introduce minimization abstraction
