### 2025-04-15 e8db567
jwst: introduce different stages of minimization
