### 2025-04-15 59b3276
jwst: try with imaging first
