+++ demos/lens_jwst.py
@@ -151,7 +151,7 @@ kl_settings_imaging = KLSettings(
     random_key=random.PRNGKey(cfg_mini.get("key", 42)),
     outputdir=join(results_directory, "imaging"),
     minimization=mini_parser,
-    n_total_iterations=10,
+    n_total_iterations=20,
     callback=plot_imaging,
     # constants=[p for p in likelihood_imaging.domain.tree if "nifty_mf" in p],
     # point_estimates=[p for p in likelihood_imaging.domain.tree if "nifty_mf" in p],
