+++ test/instruments/jwst/rotation_and_shift/test_linear.py
@@ -1,5 +1,7 @@
 import numpy as np
-from jubik0.instruments.jwst.rotation_and_shift.linear_rotation_and_shift import build_linear_rotation_and_shift
+from jubik0.instruments.jwst.rotation_and_shift.linear_rotation_and_shift import (
+    build_linear_rotation_and_shift,
+)
 
 
 def test_field_sameaxis_ij():
@@ -8,10 +10,10 @@ def test_field_sameaxis_ij():
     field[100:150, 100:150] = 1
     field[200:250, 100:150] = 1
 
-    xy = np.array(np.meshgrid(np.arange(0, maxx, 1),
-                              np.arange(0, maxy, 1),
-                              indexing='ij'))
-    rs = build_linear_rotation_and_shift(1, 1)
+    xy = np.array(
+        np.meshgrid(np.arange(0, maxx, 1), np.arange(0, maxy, 1), indexing="ij")
+    )
+    rs = build_linear_rotation_and_shift(indexing="ij", order=1)
     field_mapped = rs(field, xy)
 
     assert np.allclose(field, field_mapped, atol=1e-5)
@@ -23,10 +25,10 @@ def test_field_sameaxis_xy():
     field[100:150, 100:150] = 1
     field[200:250, 100:150] = 1
 
-    xy = np.array(np.meshgrid(np.arange(0, maxx, 1),
-                              np.arange(0, maxy, 1),
-                              indexing='xy'))
-    rs = build_linear_rotation_and_shift(1, 1, indexing='xy')
+    xy = np.array(
+        np.meshgrid(np.arange(0, maxx, 1), np.arange(0, maxy, 1), indexing="xy")
+    )
+    rs = build_linear_rotation_and_shift(indexing="xy", order=1)
     field_mapped = rs(field, xy)
 
     assert np.allclose(field, field_mapped, atol=1e-5)
@@ -38,10 +40,10 @@ def test_field_differentaxis_ij():
     field[100:150, 100:150] = 1
     field[200:250, 100:150] = 1
 
-    xy = np.array(np.meshgrid(np.arange(0, maxx, 1),
-                              np.arange(0, maxy, 1),
-                              indexing='ij'))
-    rs = build_linear_rotation_and_shift(1, 1)
+    xy = np.array(
+        np.meshgrid(np.arange(0, maxx, 1), np.arange(0, maxy, 1), indexing="ij")
+    )
+    rs = build_linear_rotation_and_shift(indexing="ij", order=1)
     field_mapped = rs(field, xy)
     assert np.allclose(field, field_mapped, atol=1e-4)
 
@@ -52,10 +54,10 @@ def test_field_differentaxis_xy():
     field[100:150, 100:150] = 1
     field[200:250, 100:150] = 1
 
-    xy = np.array(np.meshgrid(np.arange(0, maxx, 1),
-                              np.arange(0, maxy, 1),
-                              indexing='xy'))
-    rs = build_linear_rotation_and_shift(1, 1, indexing='xy')
+    xy = np.array(
+        np.meshgrid(np.arange(0, maxx, 1), np.arange(0, maxy, 1), indexing="xy")
+    )
+    rs = build_linear_rotation_and_shift(indexing="xy", order=1)
     # field_mapped = rs(field.T, np.array((yy, xx))).T
     field_mapped = rs(field, xy)
 
