### 2025-04-15 b0ddeda
jwst: testing finufft but only ij works
