+++ test/instruments/jwst/rotation_and_shift/test_linear.py
@@ -58,7 +58,6 @@ def test_field_differentaxis_xy():
         np.meshgrid(np.arange(0, maxx, 1), np.arange(0, maxy, 1), indexing="xy")
     )
     rs = build_linear_rotation_and_shift(indexing="xy", order=1)
-    # field_mapped = rs(field.T, np.array((yy, xx))).T
     field_mapped = rs(field, xy)
 
     assert np.allclose(field, field_mapped, atol=1e-4)
