### 2025-04-16 259cb03
jwst: more comprehensible energy information in plots
