### 2025-04-16 527ad6e
jwst: testing && second galaxy mask
