+++ demos/configs/spt0418.yaml
@@ -40,44 +40,44 @@ sky:
     energy_unit: eV
     energy_bin:
       e_min:
-        - 0.054 # F2100W
+        # - 0.054 # F2100W
 
         - 0.068 # F1800W  # mistake here actually 0.064
 
-        - 0.075 # F1500W
-        - 0.093 # F1280W  # mistake here actually 0.088
-        - 0.114 # F1000W
-        - 0.143 # F770W
-        - 0.201 # F560W
+        # - 0.075 # F1500W
+        # - 0.093 # F1280W  # mistake here actually 0.088
+        # - 0.114 # F1000W
+        # - 0.143 # F770W
+        # - 0.201 # F560W
 
         - 0.249 # F444W
 
-        - 0.319 # F356W  # mistake here actually 0.311
-        - 0.396 # F277W
-        - 0.557 # F200W
-        - 0.743 # F150W
+        # - 0.319 # F356W  # mistake here actually 0.311
+        # - 0.396 # F277W
+        # - 0.557 # F200W
+        # - 0.743 # F150W
 
         - 0.967 # F115W
       e_max:
-        - 0.067 # F2100W
+        # - 0.067 # F2100W
 
         - 0.075 # F1800W
 
-        - 0.092 # F1500W
-        - 0.107 # F1280W
-        - 0.137 # F1000W
-        - 0.188 # F770W
-        - 0.245 # F560W
+        # - 0.092 # F1500W
+        # - 0.107 # F1280W
+        # - 0.137 # F1000W
+        # - 0.188 # F770W
+        # - 0.245 # F560W
 
         - 0.319 # F444W
 
-        - 0.395 # F356W
-        - 0.512 # F277W
-        - 0.707 # F200W
-        - 0.932 # F150W
+        # - 0.395 # F356W
+        # - 0.512 # F277W
+        # - 0.707 # F200W
+        # - 0.932 # F150W
 
         - 1.224 # F115W
-      reference_bin: 7 # F444W
+      reference_bin: 1 # F444W
 
   model:
     source:
@@ -231,6 +231,11 @@ telescope:
     ra: [4h18m39.42s, 4h18m39.46s, 4h18m39.46s, 4h18m39.41s]
     dec: [-47d51m52.3s, -47d51m52.4s, -47d51m53.3s, -47d51m53.2s]
 
+  corner_mask_02:
+    ra : [4h18m39.59s, 4h18m39.61s, 4h18m39.59s, 4h18m39.58s]
+    dec: [-47d51m54.6s, -47d51m54.9s, -47d51m54.9s, -47d51m54.7s]
+
+
   psf:
     webbpsf_path: './data/JWST/WebbPsf/webbpsf-data-LATEST/'
     psf_library: './data/JWST/psf_library'
@@ -320,19 +325,19 @@ files:
 
   filter:
 
-    f2100w:
-      - './data/spt0418/f2100w/jw01355015001_15101_00001_mirimage/jw01355015001_15101_00001_mirimage_cal.fits'
-      - './data/spt0418/f2100w/jw01355015001_15101_00002_mirimage/jw01355015001_15101_00002_mirimage_cal.fits'
-      - './data/spt0418/f2100w/jw01355015001_15101_00003_mirimage/jw01355015001_15101_00003_mirimage_cal.fits'
-      - './data/spt0418/f2100w/jw01355015001_15101_00004_mirimage/jw01355015001_15101_00004_mirimage_cal.fits'
-      - './data/spt0418/f2100w/jw01355015001_15101_00005_mirimage/jw01355015001_15101_00005_mirimage_cal.fits'
[…]
