### 2025-04-16 d93a5b7
jwst: trying some parametric fixign
