### 2025-04-16 f06bf01
jwst: tmp; put minimum flux into lens system
