### 2025-04-16 585ff0d
jwst: fixing plotting bug, not test finufft
