### 2025-04-17 4c97f6c
jwst: bugfix plotting
