+++ demos/configs/spt0418.yaml
@@ -299,7 +299,7 @@ telescope:
       shift_unit: arcsec
       default:
         rotation: ['normal', 0, 1.0e-3]  # With great power comes great responsibility !!
-        shift: ['normal', 0, 1.0e-1]
+        shift: ['normal', 0, 1.0e1]
 
       # f1800w:
       #   0: 
@@ -346,7 +346,7 @@ telescope:
       #     rotation: ['delta', -1.95823434e-05, 0]
 
 files:
-  res_dir: './results/jwst_lens/cluster/06_fixpointing'
+  res_dir: './results/jwst_lens/cluster/06_fixpointing_extreme_shift_prior'
 
   filter:
 
