+++ demos/configs/spt0418.yaml
@@ -299,7 +299,7 @@ telescope:
       shift_unit: arcsec
       default:
         rotation: ['normal', 0, 1.0e-3]  # With great power comes great responsibility !!
-        shift: ['normal', 0, 1.0e1]
+        shift: ['normal', 0, 1.0e-1]
 
       # f1800w:
       #   0: 
@@ -346,7 +346,7 @@ telescope:
       #     rotation: ['delta', -1.95823434e-05, 0]
 
 files:
-  res_dir: './results/jwst_lens/cluster/06_fixpointing_extreme_shift_prior'
+  res_dir: './results/jwst_lens/cluster/06_fixpointing_moresampling'
 
   filter:
 
@@ -431,15 +431,15 @@ minimization:
     n_total_iterations: 20
 
     delta:
-      switches: [0, 10, 15]
-      values: [5.e-6, 1.e-6, 1.e-7]
+      switches: [0, 4, 8, 15]
+      values: [5.e-6, 1.e-6, 5.e-7, 1.e-7]
 
     samples:
-      switches: [0, 15, 30]
+      switches: [0, 4, 8, 30]
       n_samples: [3]
       # n_samples: [5]
       # mode: [linear_resample, nonlinear_resample, nonlinear_resample, nonlinear_update]
-      mode: [linear_resample, nonlinear_resample]
+      mode: [linear_resample, linear_resample, nonlinear_resample]
       # Possible modes:
       # - "linear_sample",
       # - "linear_resample",
@@ -449,7 +449,7 @@ minimization:
 
 
       # lin_absdelta: [Null, Null, 0.00001]
-      lin_maxiter: [80, 250, 350]
+      lin_maxiter: [80, 250, 350, 350]
 
       # nonlin_xtol: [Null, 0.1, 0.01]
       nonlin_maxiter: [7, 13, 18]
