+++ demos/lens_jwst.py
@@ -166,7 +166,7 @@ kl_settings_fixpointing = KLSettings(
     random_key=random.PRNGKey(cfg_mini.get("key", 42)),
     outputdir=join(results_directory, "fixpointing"),
     minimization=mini_parser,
-    n_total_iterations=13,
+    n_total_iterations=9,
     callback=plot_imaging,
     # constants=[p for p in likelihood_fixpointing.domain.tree if "nifty_mf" in p],
     # point_estimates=[p for p in likelihood_fixpointing.domain.tree if "nifty_mf" in p],
