### 2025-04-17 2833f62
jwst: no shifting center, prior spectral index flex, plotting diff
