+++ demos/configs/spt0418.yaml
@@ -152,7 +152,7 @@ sky:
               loglogslope: [-4.0, 0.1]  # normal
             spectral_index:
               mean: [-1.5, 0.5]         # normal
-              fluctuations: [0.05, 0.005]  # lognormal 
+              fluctuations: [0.5, 0.05]  # lognormal 
             spectral_deviations:
               process: 'wiener'
               sigma: [0.3, 0.1]  # lognormal
@@ -329,13 +329,13 @@ telescope:
       #     shift: ['delta', [-0.08923382,  0.16512098], 0]  # pixel units
       #     rotation: ['delta', -0.00151674, 0]
 
-      f444w:
-        0: 
-          shift: ['delta', 0, 0]  # pixel units
-          rotation: ['delta', 0, 0]
-        # 1: 
-        #   shift: ['delta', [-0.04078999, -0.01561007], 0]  # pixel units
-        #   rotation: ['delta', -0.00138025, 0]
+      # f444w:
+      #   0: 
+      #     shift: ['delta', 0, 0]  # pixel units
+      #     rotation: ['delta', 0, 0]
+      #   1: 
+      #     shift: ['delta', [-0.04078999, -0.01561007], 0]  # pixel units
+      #     rotation: ['delta', -0.00138025, 0]
 
       # f115w:
       #   0: 
@@ -346,7 +346,7 @@ telescope:
       #     rotation: ['delta', -1.95823434e-05, 0]
 
 files:
-  res_dir: './results/jwst_lens/cluster/07_fixpointing_tweakreag'
+  res_dir: './results/jwst_lens/cluster/08_fixpointing_no_shift_reference'
 
   filter:
 
@@ -452,7 +452,7 @@ minimization:
       lin_maxiter: [80, 250, 350, 350]
 
       # nonlin_xtol: [Null, 0.1, 0.01]
-      nonlin_maxiter: [7, 13, 18]
+      nonlin_maxiter: [7, 7, 13, 18]
       # nonlin_cg_atol: [0.1, 0.01]
       # nonlin_cg_maxiter: [20, 30]
 
