### 2025-04-18 85b3672
jwst: trying stop optimization
