+++ demos/configs/spt0418.yaml
@@ -346,7 +346,7 @@ telescope:
       #     rotation: ['delta', -1.95823434e-05, 0]
 
 files:
-  res_dir: './results/jwst_lens/cluster/08_fixpointing_no_shift_reference'
+  res_dir: './results/jwst_lens/cluster/09_fixpointing_stop_lens_optimization'
 
   filter:
 
@@ -426,7 +426,7 @@ files:
 
 
 minimization:
-    key: 20250412
+    key: 20250426
     resume: False
     n_total_iterations: 20
 
