### 2025-04-18 e5eb6ef
jwst: tmp constants bugfix
