### 2025-04-18 aa04e88
jwst: bugfix in residuals plot
