### 2025-04-20 084ec4b
jwst: plotting adjust
