+++ src/instruments/jwst/plotting/plotting.py
@@ -31,7 +31,6 @@ def get_plot(
     parametric_lens: bool,
     parametric_source: bool = False,
     max_residuals: int = 4,
-    residual_ylen_offset: int = 0,
 ):
     from charm_lensing.lens_system import LensSystem
 
@@ -47,11 +46,16 @@ def get_plot(
         reference=FieldPlottingConfig(vmax=5, norm=None),
         alpha=FieldPlottingConfig(norm=None),
     )
-
     plotting_config_lens_system = LensSystemPlottingConfig(
         source=source_plotting_config,
         lens_light=lens_light_plotting_config,
         share_source_vmin_vmax=False,
+    )
+    residual_plotting_config = ResidualPlottingConfig(
+        sky=lens_light_plotting_config.combined,
+        data=FieldPlottingConfig(norm="log", vmin=1e-3),
+        display_pointing=False,
+        xmax_residuals=max_residuals,
     )
 
     plot_lens = build_plot_lens_system(
@@ -70,13 +74,6 @@ def get_plot(
         grid=grid,
     )
 
-    residual_plotting_config = ResidualPlottingConfig(
-        sky=lens_light_plotting_config.combined,
-        data=FieldPlottingConfig(norm="log", vmin=1e-3),
-        display_pointing=False,
-        xmax_residuals=max_residuals,
-        ylen_offset=residual_ylen_offset,
-    )
     plot_residual = build_plot_sky_residuals(
         results_directory=results_directory,
         filter_projector=filter_projector,
@@ -96,27 +93,25 @@ def plot_prior(
     plot_source: Union[bool, callable],
     plot_lens: Union[bool, callable],
     data_dict: dict,
-    parametric_flag: bool,
     sky_key: str = "sky",
-    residual_ylen_offset: int = 0,
 ):
     import yaml
     from charm_lensing.lens_system import build_lens_system
-
     from jubik0.instruments.jwst.config_handler import insert_spaces_in_lensing_new
+    from charm_lensing.lens_system import LensSystem
 
     test_key, _ = random.split(random.PRNGKey(42), 2)
 
     cfg = yaml.load(open(config_path, "r"), Loader=yaml.SafeLoader)
     insert_spaces_in_lensing_new(cfg["sky"])
 
-    lens_system = build_lens_system(cfg["sky"])
+    lens_system: LensSystem = build_lens_system(cfg["sky"])
     if cfg["nonparametric_lens"]:
         sky_model = lens_system.get_forward_model_full()
-        parametric_flag = False
+        parametric_lens = False
     else:
         sky_model = lens_system.get_forward_model_parametric()
-        parametric_flag = True
+        parametric_lens = True
     sky_model = jft.Model(jft.wrap_left(sky_model, sky_key), domain=sky_model.domain)
     sky_model_with_keys = jft.Model(
         lambda x: filter_projector(sky_model(x)), init=sky_model.init
@@ -124,48 +119,52 @@ def plot_prior(
 
     grid = Grid.from_grid_model(GridModel.from_yaml_dict(cfg["sky"]["grid"]))
     results_directory = cfg["files"]["res_dir"]
-    ll_alpha, ll_nonpar, sl_alpha, sl_nonpar = get_alpha_nonpar(lens_system)
+    # ll_alpha, ll_nonpar, sl_alpha, sl_nonpar = get_alpha_nonpar(lens_system)
+
+    source_plotting_config = MultiFrequencyPlottingConfig(
+        combined=FieldPlottingConfig(vmin=1e-4, norm="log"),
+        reference=FieldPlottingConfig(vmax=5, norm=None),
+        alpha=FieldPlottingConfig(norm=None),
+    )
[…]
