### 2025-04-20 3530467
jwst: tmp new priors for all data
