### 2025-04-21 a447959
jwst: tmp some testing of priors
