+++ demos/configs/spt0418.yaml
@@ -305,7 +305,7 @@ telescope:
 
 
 files:
-  res_dir: './results/jwst_lens/cluster/11_01_priors_fewerdata'
+  res_dir: './results/jwst_lens/cluster/11_02_priors_fewerdata'
 
   filter:
 
@@ -385,7 +385,7 @@ files:
 
 
 minimization:
-    key: 20250426
+    key: 20250421
     resume: False
     n_total_iterations: 20
 
