### 2025-04-21 9d2a031
jwst: tmp easter prior with true ndof
