### 2025-04-22 a75fbc1
jwst: tmp covariance adjustable sources
