### 2025-04-22 52dc0bd
jwst: tmp; multiple covariances etc., questionable mf-priors
