### 2025-04-22 4afcb10
jwst: tmp; testing priors furhter;
