### 2025-04-22 b19611b
jwst: try full data
