### 2025-04-22 5a75f03
jwst: activate spectral deviations in lens light
