### 2025-04-22 42bfa43
jwst: plotting fix
