### 2025-04-22 3b17446
jwst: bugfix in minimization from samples
