### 2025-04-22 2f62837
jwst: new full data from pickle
