### 2025-04-22 653db26
jwst: reconstruction from starting pos for shifts
