### 2025-04-22 e672ea9
jwst: some more handling of state etc
