+++ src/minimization/minimization_from_samples.py
@@ -3,6 +3,7 @@ from ..minimization_parser import MinimizationParser
 import nifty8.re as jft
 from jax import random
 
+import os
 import pickle
 from dataclasses import dataclass
 from typing import Callable, Optional
@@ -27,26 +28,26 @@ class KLSettings:
 def _initial_position(
     init_key: random.PRNGKey,
     domain,
-    position_rescaling,
+    kl_settings: KLSettings,
     starting_samples: Optional[jft.Samples] = None,
     not_take_starting_pos_keys: tuple[str] = (),
-    resume_from_pickle_path: str | None = None,
 ):
+    position_rescaling = kl_settings.sample_multiply
     initial_position = jft.random_like(init_key, domain) * position_rescaling
     while isinstance(initial_position, jft.Vector):
         initial_position = initial_position.tree
 
-    if resume_from_pickle_path is not None:
+    if kl_settings.resume_from_pickle_path is not None:
         if starting_samples is not None:
             jft.logger.warning(
                 "Warning: Overwriting starting position from path: "
-                f"{resume_from_pickle_path}."
+                f"{kl_settings.resume_from_pickle_path}."
             )
         else:
-            jft.logger.info(f"Loading result: {resume_from_pickle_path}")
+            jft.logger.info(f"Loading result: {kl_settings.resume_from_pickle_path}")
 
-        with open(resume_from_pickle_path, "rb") as f:
-            starting_samples, opt_vi_st = pickle.load(f)
+        with open(kl_settings.resume_from_pickle_path, "rb") as f:
+            starting_samples, _ = pickle.load(f)
 
     if starting_samples is not None:
         starting_pos = starting_samples.pos
@@ -58,7 +59,21 @@ def _initial_position(
                 jft.logger.info(f"Taking {key}: starting_pos[key]")
                 initial_position[key] = starting_pos[key]
 
-    return jft.Vector(initial_position)
+    initial_position, opt_vi_state = jft.Vector(initial_position), None
+
+    LAST_FILENAME = "last.pkl"
+    resume = False
+    if kl_settings.resume:
+        if isinstance(kl_settings.resume, str):
+            last_pkl = kl_settings.resume
+        else:
+            last_pkl = os.path.join(kl_settings.outputdir, LAST_FILENAME)
+
+        if os.path.isfile(last_pkl):
+            with open(last_pkl, "rb") as f:
+                initial_position, opt_vi_state = pickle.load(f)
+
+    return initial_position, opt_vi_state
 
 
 def minimization_from_initial_samples(
@@ -91,7 +106,7 @@ def minimization_from_initial_samples(
 
     init_key, mini_key = random.split(kl_settings.random_key, 2)
 
-    initial_position = _initial_position(
+    initial_position, opt_vi_state = _initial_position(
         init_key,
         likelihood.domain,
         position_rescaling=kl_settings.sample_multiply,
@@ -116,8 +131,9 @@ def minimization_from_initial_samples(
         kl_kwargs=minimization.kl_kwargs,
         constants=kl_settings.constants,
         point_estimates=kl_settings.point_estimates,
-        resume=kl_settings.resume,
+        resume=False,
         kl_jit=kl_settings.kl_jit,
         residual_jit=kl_settings.residual_jit,
+        _optimize_vi_state=opt_vi_state,
     )
     return samples, state
