### 2025-04-22 8921a0a
jwst: minimization bugfix
