+++ src/instruments/jwst/jwst_data.py
@@ -145,7 +145,12 @@ class JwstData:
         data : ArrayLike
             Data values inside the extrema.
         """
-        minx, maxx, miny, maxy = self.wcs.index_from_world_extrema(extrema, self.shape)
+        # NOTE : Carefull replacing bounding_box_indices_from_world_extrema since this
+        # is coupled inside self.load_cutout_data_mask_std_by_world_corners.
+
+        minx, maxx, miny, maxy = self.wcs.bounding_box_indices_from_world_extrema(
+            extrema, self.shape
+        )
 
         # NOTE : The data needs matrix indexing, hence y is on the first axis.
         return self.dm.data[miny : maxy + 1, minx : maxx + 1]
@@ -164,7 +169,12 @@ class JwstData:
         data : ArrayLike
             Data values inside the extrema.
         """
-        minx, maxx, miny, maxy = self.wcs.index_from_world_extrema(extrema, self.shape)
+        # NOTE : Carefull replacing bounding_box_indices_from_world_extrema since this
+        # is coupled inside self.load_cutout_data_mask_std_by_world_corners.
+
+        minx, maxx, miny, maxy = self.wcs.bounding_box_indices_from_world_extrema(
+            extrema, self.shape
+        )
         # NOTE : The data needs matrix indexing, hence y is on the first axis.
         return self.dm.err[miny : maxy + 1, minx : maxx + 1]
 
@@ -183,7 +193,12 @@ class JwstData:
         nan-mask : ArrayLike
             Mask corresponding to the nan values inside the extrema.
         """
-        minx, maxx, miny, maxy = self.wcs.index_from_world_extrema(extrema, self.shape)
+        # NOTE : Carefull replacing bounding_box_indices_from_world_extrema since this
+        # is coupled inside self.load_cutout_data_mask_std_by_world_corners.
+
+        minx, maxx, miny, maxy = self.wcs.bounding_box_indices_from_world_extrema(
+            extrema, self.shape
+        )
         # NOTE : The data needs matrix indexing, hence y is on the first axis.
         return (~np.isnan(self.dm.data[miny : maxy + 1, minx : maxx + 1])) * (
             ~np.isnan(self.dm.err[miny : maxy + 1, minx : maxx + 1])
@@ -211,54 +226,79 @@ class JwstData:
         pivot, bw, effective_response, blue, red = JWST_FILTERS[self.filter]
         return effective_response
 
+    def bounding_data_mask_std_by_world_corners(
+        self,
+        reconstruction_grid: Grid,
+        world_corners: list[SkyCoord],
+        additional_masks_corners: list[CornerMaskConfig],
+    ) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
+        """Return a subpart of the data, mask, and std inside a bounding box surrounding
+        the `world_corners`.
 
-def load_jwst_data_mask_std(
-    filepath: str,
-    grid: Grid,
-    world_corners: list[SkyCoord],
-    mask_corners: list[CornerMaskConfig],
-) -> [JwstData, ArrayLike, ArrayLike, ArrayLike]:
-    """Load the data from filepath and return the data, mask, and std cutouts
-    according to the reconstruction grid and `world_corners`.
+        Parameters
+        ----------
+        world_corners: list[SkyCoord]
+            The world corners of the cutout. I.e. the absolute positions in the world.
+        additional_masks_corners: list[CornerMaskConfig]
+            Holds the egde points of additional masks for the data.
 
-    Parameters
-    ----------
-    filepath: str
-        Path to the jwst data.
-    grid: Grid
-        The physical grid underlying the reconstruction
-    world_corners: list[SkyCoord]
[…]
