+++ demos/configs/spt0418.yaml
@@ -40,62 +40,62 @@ sky:
     energy_unit: eV
     energy_bin:
       e_min:
-        - 0.054 # F2100W
+        # - 0.054 # F2100W
 
         - 0.068 # F1800W  # mistake here actually 0.064
 
-        - 0.075 # F1500W
-        - 0.093 # F1280W  # mistake here actually 0.088
-        - 0.114 # F1000W
-        - 0.143 # F770W
-        - 0.201 # F560W
+        # - 0.075 # F1500W
+        # - 0.093 # F1280W  # mistake here actually 0.088
+        # - 0.114 # F1000W
+        # - 0.143 # F770W
+        # - 0.201 # F560W
 
         - 0.249 # F444W
 
-        - 0.319 # F356W  # mistake here actually 0.311
-        - 0.396 # F277W
-        - 0.557 # F200W
-        - 0.743 # F150W
+        # - 0.319 # F356W  # mistake here actually 0.311
+        # - 0.396 # F277W
+        # - 0.557 # F200W
+        # - 0.743 # F150W
 
         - 0.967 # F115W
       e_max:
-        - 0.067 # F2100W
+        # - 0.067 # F2100W
 
         - 0.075 # F1800W
 
-        - 0.092 # F1500W
-        - 0.107 # F1280W
-        - 0.137 # F1000W
-        - 0.188 # F770W
-        - 0.245 # F560W
+        # - 0.092 # F1500W
+        # - 0.107 # F1280W
+        # - 0.137 # F1000W
+        # - 0.188 # F770W
+        # - 0.245 # F560W
 
         - 0.319 # F444W
 
-        - 0.395 # F356W
-        - 0.512 # F277W
-        - 0.707 # F200W
-        - 0.932 # F150W
+        # - 0.395 # F356W
+        # - 0.512 # F277W
+        # - 0.707 # F200W
+        # - 0.932 # F150W
 
         - 1.224 # F115W
-      reference_bin: 7 # F444W
+      reference_bin: 1 # F444W
 
   model_fixing_pointing:
     source:
       light:
         mean:
           gaussian_01:
-            # i0: ['log_normal', 1.0 , 20.0, [3, 1, 1]]
-            i0: ['log_normal', 1.0 , 20.0, [13, 1, 1]]
+            i0: ['log_normal', 1.0 , 20.0, [3, 1, 1]]
+            # i0: ['log_normal', 1.0 , 20.0, [13, 1, 1]]
             center: ['normal', 0, 0.1]
-            # covariance: ['log_normal', 0.08, 0.02]
-            covariance: ['log_normal', 0.08, 0.02, [2, 13, 1, 1]]
+            covariance: ['log_normal', 0.08, 0.02]
+            # covariance: ['log_normal', 0.08, 0.02, [2, 13, 1, 1]]
             off_diagonal: ['normal', 0.0, 10.0]
           gaussian_02: # i0*Gauss
-            # i0: ['log_normal', 1.0 , 20.0, [3, 1, 1]]
-            i0: ['log_normal', 1.0 , 20.0, [13, 1, 1]]
+            i0: ['log_normal', 1.0 , 20.0, [3, 1, 1]]
+            # i0: ['log_normal', 1.0 , 20.0, [13, 1, 1]]
             center: ['normal', [-0.5, -0.5], 0.1]
-            # covariance: ['log_normal', 0.08, 0.02]
-            covariance: ['log_normal', 0.08, 0.02, [2, 13, 1, 1]]
+            covariance: ['log_normal', 0.08, 0.02]
+            # covariance: ['log_normal', 0.08, 0.02, [2, 13, 1, 1]]
[…]
