### 2025-04-25 38b0175
jwst: Deleting Sparse Interpolation
