### 2025-04-25 5a0b77f
jwst: delete unused shift model
