### 2025-04-25 935c13d
jwst: simplifying RotationAndShift
