+++ src/instruments/jwst/rotation_and_shift/__init__.py
@@ -7,5 +7,4 @@
 
 from .linear_rotation_and_shift import build_linear_rotation_and_shift
 from .nufft_rotation_and_shift import build_nufft_rotation_and_shift
-from .rotation_and_shift import (
-    build_rotation_and_shift_model, RotationAndShiftModel)
+from .rotation_and_shift import build_rotation_and_shift, RotationAndShift
