### 2025-04-25 e7d10d0
jwst: restructure likelihood building
