+++ src/instruments/jwst/jwst_response.py
@@ -5,18 +5,25 @@
 
 # %
 
-from typing import Callable
+from typing import Callable, Union
 
 import nifty8.re as jft
+import numpy as np
 from astropy import units as u
 from numpy.typing import ArrayLike
 
+from .parse.parametric_model.parametric_prior import ProbabilityConfig
+from .parse.rotation_and_shift.rotation_and_shift import LinearConfig, NufftConfig
+from .parse.jwst_response import SkyMetaInformation
+from .jwst_data import DataMetaInformation
+
 from ...wcs import subsample_grid_centers_in_index_grid
+from ...wcs.wcs_jwst_data import WcsJwstData
+from ...wcs.wcs_astropy import WcsAstropy
 from .integration.unit_conversion import build_unit_conversion
 from .integration.integration import integration_factory
 from .jwst_psf import build_psf_operator
 from .masking.build_mask import build_mask
-from .parse.parametric_model.parametric_prior import ProbabilityConfig
 from .rotation_and_shift import RotationAndShift, build_rotation_and_shift
 from .rotation_and_shift.coordinates_correction import (
     build_coordinates_corrected_from_grid,
@@ -24,6 +31,7 @@ from .rotation_and_shift.coordinates_correction import (
 )
 from .zero_flux_model import build_zero_flux_model
 from ...grid import Grid
+from .jwst_data import JwstData
 
 
 class JwstResponse(jft.Model):
@@ -112,16 +120,16 @@ class JwstResponse(jft.Model):
 
 def build_jwst_response(
     sky_domain: dict,
-    data_identifier: str,
-    data_subsample: int,
-    rotation_and_shift_kwargs: dict | None,
+    data_wcs: WcsJwstData,
+    data_meta: DataMetaInformation,
+    sky_wcs: WcsAstropy,
+    sky_meta: SkyMetaInformation,
+    rotation_and_shift_algorithm: Union[LinearConfig, NufftConfig],
     shift_and_rotation_correction: ShiftAndRotationCorrection | None,
-    psf_kernel: ArrayLike | None,
+    psf_kernel: np.ndarray | None,
     transmission: float,
     zero_flux_prior_config: ProbabilityConfig | None,
     data_mask: ArrayLike | None,
-    sky_unit: u.Unit,
-    data_unit: u.Unit,
 ) -> JwstResponse:
     """
     Builds the data model for a Jwst observation.
@@ -133,23 +141,14 @@ def build_jwst_response(
     ----------
     sky_domain: dict
         Containing the sky_key and the shape_dtype of the reconstruction sky.
-    data_subsample: int
-        The subsample factor for the data grid.
-    rotation_and_shift_kwargs: dict
-        reconstruction_grid: Grid
-        data_dvol: Unit, the volume of a data pixel
-        data_wcs: WcsBase,
-        data_model_type: str,
-        world_corners: Tuple[SkyCoord]
+    data: JwstData
+        The data with meta information.
+    rotation_and_shift_algorithm: Union[LinearConfig, NufftConfig],
+        The information for the rotation_and_shift algorithm.
     shift_and_rotation_correction: CoordiantesCorrectionPriorConfig | None
         The prior for the shift and rotation coordinates correction.
-    psf_kernel_model:
-        camera: str, NIRCam or MIRI
-        filter: str
-        center_pix: tuple, pixel according to which to evaluate the psf model
-        webbpsf_path: str
-        fov_pixels: int | None, how many pixles considered for the psf
-        fov_arcsec: float | None, the arcseconds for the psf evaluation
+    psf_kernel: np.ndarray
+        The
     data_mask: ArrayLike
         The mask on the data
[…]
