+++ src/instruments/jwst/parse/jwst_response.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+
+from astropy import units as u
+
+
+@dataclass
+class SkyMetaInformation:
+    grid_extension: tuple[int, int]
+    unit: u.Unit
