+++ src/wcs/wcs_jwst_data.py
@@ -6,9 +6,6 @@
 # %%
 
 import numpy as np
-from astropy.coordinates import SkyCoord
-from numpy.typing import ArrayLike
-
 from .wcs_base import WcsMixin
 
 
