+++ demos/configs/spt0418.yaml
@@ -306,7 +306,7 @@ telescope:
 
 
 files:
-  res_dir: './results/jwst_lens/cluster/13_02_testing_alignement'
+  res_dir: './results/jwst_lens/cluster/13_03_testing_alignement'
 
   filter:
 
@@ -388,7 +388,7 @@ files:
 minimization:
     key: 20250425
     resume: True
-    n_total_iterations: 20
+    n_total_iterations: 5
 
     fixpointing:
       resume: False

[…]
