### 2025-04-25 fc53bfc
jwst: rename tests for pytest to run
