### 2025-04-25 a54e3d3
jwst: simplifying bounding_data_mask_std loadin
