### 2025-04-25 96ae416
jwst: structuring plotting information
