### 2025-04-25 ad5a254
jwst: bugfix in dataloading
