### 2025-04-26 a5c672f
jwst: tmp restructuring data loading, into arrays
