+++ src/instruments/jwst/jwst_likelihoods.py
@@ -1,40 +1,60 @@
 from ...likelihood import build_gaussian_likelihood
-from .config_handler import (
-    get_grid_extension_from_config,
-)
-
-from .jwst_response import build_jwst_response
-from .jwst_data import JwstData
 from ...grid import Grid
+from ...wcs.wcs_jwst_data import WcsJwstData
+from .jwst_response import build_jwst_response
+from .jwst_data import JwstData, DataMetaInformation
+from .jwst_psf import load_psf_kernel
 from .filter_projector import FilterProjector, build_filter_projector
 from .rotation_and_shift.coordinates_correction import (
     ShiftAndRotationCorrection,
 )
-from .jwst_psf import load_psf_kernel
+from .plotting.residuals import ResidualPlottingInformation
+from .config_handler import (
+    get_grid_extension_from_config,
+)
 
 # Parsing
 from .parse.jwst_psf import yaml_to_psf_kernel_config
 from .parse.zero_flux_model import yaml_to_zero_flux_prior_config
 from .parse.rotation_and_shift.coordinates_correction import (
     yaml_to_coordinates_correction_config,
+    CoordinatesCorrectionConfig,
 )
 from .parse.rotation_and_shift.rotation_and_shift import (
     rotation_and_shift_algorithm_config_factory,
 )
 from .parse.jwst_response import SkyMetaInformation
 from .parse.masking.data_mask import yaml_to_corner_mask_configs
-from .plotting.residuals import ResidualPlottingInformation
 
 
 # Libraries
 import jax.numpy as jnp
 import nifty8.re as jft
 from nifty8.logger import logger
+import numpy as np
 
 # std
 from functools import reduce
 from astropy import units as u
+from astropy.coordinates import SkyCoord
 from typing import Union
+from dataclasses import dataclass, field
+
+
+@dataclass
+class FilterData:
+    data: list[np.ndarray] = field(default_factory=list)
+    mask: list[np.ndarray] = field(default_factory=list)
+    std: list[np.ndarray] = field(default_factory=list)
+    psf_kernel: list[np.ndarray] = field(default_factory=list)
+    boresight: list[SkyCoord] = field(default_factory=list)
+    transmission: list[float] = field(default_factory=list)
+    meta: list[DataMetaInformation] = field(default_factory=list)
+    data_wcs: list[WcsJwstData] = field(default_factory=list)
+    correction_prior: list[CoordinatesCorrectionConfig] = field(default_factory=list)
+
+    def __len__(self):
+        return len(self.data)
 
 
 def build_jwst_likelihoods(
@@ -72,6 +92,8 @@ def build_jwst_likelihoods(
     target_plotting = {}
     likelihoods = []
     for fltname, flt in cfg[files_key]["filter"].items():
+        filter_data = FilterData()
+
         for ii, filepath in enumerate(flt):
             logger.info(f"Loading: {fltname} {ii} {filepath}")
 
@@ -79,12 +101,17 @@ def build_jwst_likelihoods(
             jwst_data = JwstData(
                 filepath, identifier=f"{fltname}_{ii}", subsample=data_subsample
             )
+            energy_name = filter_projector.get_key(jwst_data.pivot_wavelength)
+            if ii == 0:
+                previous_energy_name = energy_name
+            else:
+                assert energy_name == previous_energy_name
+
             data, mask, std = jwst_data.bounding_data_mask_std_by_world_corners(
                 grid.spatial,
                 grid.spatial.world_corners(extension_value=sky_meta.grid_extension),
                 yaml_to_corner_mask_configs(cfg[telescope_key]),
             )
-            energy_name = filter_projector.get_key(jwst_data.pivot_wavelength)
 
[…]
