### 2025-04-26 8252d97
jwst: simplify boresight jwst getter
