### 2025-04-26 3b08e7c
jwst: restructure pixel center evaluaution
