### 2025-04-27 5c3ab86
jwst: laoding data directly from the xmin_xmax_ymin_ymax indices
