### 2025-04-27 ca8f15a
jwst: masking, replacing internal in sky mask building
