+++ src/instruments/jwst/masking/data_mask.py
@@ -5,7 +5,11 @@ from astropy.coordinates import SkyCoord
 
 from ....wcs.wcs_astropy import WcsAstropy
 from ....wcs.wcs_jwst_data import WcsJwstData
-from ....wcs.wcs_subsample_centers import subsample_grid_centers_in_index_grid
+from ....wcs.wcs_subsample_centers import (
+    subsample_grid_centers_in_index_grid,
+    subsample_pixel_centers,
+    world_coordinates_to_index_grid,
+)
 
 
 def get_mask_from_index_centers_within_rgrid(
@@ -23,12 +27,13 @@ def get_mask_from_index_centers_within_rgrid(
         The shape of rgrid.
     """
 
-    data_pix_centers_in_index_grid: np.ndarray = subsample_grid_centers_in_index_grid(
-        world_corners=world_corners,
-        to_be_subsampled_grid_wcs=data_wcs,
-        index_grid_wcs=index_grid_wcs,
+    data_pix_centers_world_coordinates = subsample_pixel_centers(
+        data_wcs.bounding_box_indices_from_world_extrema(world_corners),
+        data_wcs,
         subsample=1,
-        indexing="xy",
+    )
+    data_pix_centers_in_index_grid = world_coordinates_to_index_grid(
+        data_pix_centers_world_coordinates, index_grid_wcs=index_grid_wcs, indexing="xy"
     )
 
     return (
