### 2025-04-27 f441a79
jwst: masking, replacing internal in extra mask building
