### 2025-04-28 9f31b67
gaia: create the possibility to join gaia tables
