### 2025-04-28 39a2e1c
jwst: return the data bounds as right handed points
