### 2025-04-28 4fd267e
jwst: possiblity to load data from bounding indices
