+++ src/wcs/wcs_subsample_centers.py
@@ -16,7 +16,7 @@ from .wcs_astropy import WcsAstropy
 
 
 def subsample_pixel_centers(
-    index_bounds_minx_maxx_miny_maxy: tuple[int, int, int, int],
+    bounding_box_xmin_xmax_ymin_ymax: tuple[int, int, int, int],
     to_be_subsampled_grid_wcs: Union[WcsAstropy, WcsJwstData],
     subsample: int,
     as_pixel_values: bool = False,
@@ -44,7 +44,7 @@ def subsample_pixel_centers(
 
     # NOTE : GWCS.wcs expects `xy` indexing. Other arrays are not tested.
     tbsg_pixcenter_indices = to_be_subsampled_grid_wcs.index_grid_from_bounding_indices(
-        *index_bounds_minx_maxx_miny_maxy, indexing="xy"
+        *bounding_box_xmin_xmax_ymin_ymax, indexing="xy"
     )
 
     ps = np.arange(0.5 / subsample, 1, 1 / subsample) - 0.5
