### 2025-04-28 87cd4ba
jwst: delete depricated subsample_grid_centers_in_index_grid
