+++ src/instruments/jwst/masking/data_mask.py
@@ -6,7 +6,6 @@ from astropy.coordinates import SkyCoord
 from ....wcs.wcs_astropy import WcsAstropy
 from ....wcs.wcs_jwst_data import WcsJwstData
 from ....wcs.wcs_subsample_centers import (
-    subsample_grid_centers_in_index_grid,
     subsample_pixel_centers,
     world_coordinates_to_index_grid,
 )
