+++ README.md
@@ -19,7 +19,7 @@ Next to many useful generic tools and building blocks, JUBIK comes with a series
 ## Instrument specific Requirements 
 - [Chandra](#chandra)
 - [eROSITA](#erosita)
-- [James Webb Space Telescope](#jwst)
+- [James Webb Space Telescope](#james-webb-space-telescope)
 
 ## Installation
 This package can be installed via pip. 
@@ -34,7 +34,7 @@ for a regular installation. For editable installation add the `-e` flag.
 ## Additional Files
 Additional calibration files might be needed for instrument-specific pipelines.
 
-### Chandra{#chandra}
+### Chandra
 J-UBIK allows to process observations from Chandra x-ray observatory.
 
 #### Requirements
@@ -46,7 +46,7 @@ NOTE: in case you install ciao via conda, make sure that all environmental are s
 conda env config vars set MARX_ROOT /soft/marx/marx-5.2.0 (or where your marx is installed)
 conda env config vars set MARX_DATA_DIR ${MARX_ROOT}/share/marx/data
 
-### eROSITA{#erosita}
+### eROSITA
 J-UBIK allows to process and image event files from the eROSITA x-ray observatory.
 
 #### Requirements
@@ -61,6 +61,7 @@ Either the caldb from data release 1 (DR1) or from the early data release (EDR) 
 inside the `data/` directory. 
 This folder can be downloaded at [caldb download](https://erosita.mpe.mpg.de/dr1/eSASS4DR1/eSASS4DR1_installation/caldb4DR1.tgz).
 - Download the data if you want to work with public eROSITA data, see [edr](https://erosita.mpe.mpg.de/edr/index.php) and [dr1](https://erosita.mpe.mpg.de/dr1/index.html).  
+
 #### Demo
 In the `demo/` repository, `erosita_inference.py` allows to run a generic 
 image reconstruction with real and synthetic (mock) eROSITA data.
@@ -70,7 +71,7 @@ in order to build realistic exposure maps.
 A good example is [LMC_dataset](https://erosita.mpe.mpg.de/edr/eROSITAObservations/CalPvObs/LMC_SN1987A.tar.gz).
 For more information on how to run `erosita_demo.py` see the corresponding docstring.
 
-### James Webb Space Telescope{#jwst}
+### James Webb Space Telescope
 J-UBIK allows to process and image event files from the James Webb Space Telescope.
 
 #### Requirements
