### 2025-04-28 aaf326c
remove ciao legacy note from readme
recomment conda env installation instead
