### 2025-04-28 3e0f580
jwst: implement star alignment
