### 2025-04-29 9b9c98d
jwst: delete unneeded function
