### 2025-04-29 9de9af0
jwst: aligned data check
