### 2025-04-29 82fb691
jwst: correct alignment of data
