### 2025-04-29 f979e22
jwst: after stage 2 one doesn't need to account for transmission
