### 2025-04-29 bd92b26
jwst: build_shift_and_rotation_correction input
