### 2025-04-29 4798555
jwst: what is wrong with the coordinates calclulation
