### 2025-04-29 0efafbb
check for the Type not Boolean
