### 2025-04-30 32f851a
jwst: distances as u.Quantity, new ShiftCorrection, multiple pixel_coords
