### 2025-04-30 150db28
jwst: implement dynamic loading
