### 2025-04-30 380d7bb
jwst: rename to prior_config_factory
