### 2025-05-02 NIFTy-PPL/J-UBIK#11 @Eberle-Vincent
@hamogu thank you for finding this one. It is fixed in 24633b9
