### 2025-05-02 de2cefe
jwst: cleaner DataBoundsPreloading
