+++ src/instruments/jwst/data/data_loading.py
@@ -3,7 +3,7 @@ import numpy as np
 
 
 @dataclass
-class Preloading:
+class DataBoundsPreloading:
     shapes: tuple[int, int] | list[tuple[int, int]] | list[np.ndarray] = field(
         default_factory=list
     )
@@ -11,16 +11,10 @@ class Preloading:
         tuple[int, int, int, int] | list[tuple[int, int, int, int]] | list[np.ndarray]
     ) = field(default_factory=list)
 
-    def append_shape(self, shape: tuple[int, int]):
-        self.shapes.append(shape)
-
-    def append_bounding_indices(self, bounding_indices: tuple[int, int, int, int]):
-        self.bounding_indices.append(bounding_indices)
-
     def __getitem__(self, index: int):
-        return Preloading(self.shapes[index], self.bounding_indices[index])
+        return DataBoundsPreloading(self.shapes[index], self.bounding_indices[index])
 
-    def align_data(self):
+    def align_shapes(self):
         shapes = np.array(self.shapes)
         shapes_new = np.full_like(shapes, np.max(shapes, axis=0))
 
@@ -28,4 +22,4 @@ class Preloading:
         for ii in range(len(bounding_indices_new)):
             bounding_indices_new[ii, [3, 1]] += shapes_new[ii] - shapes[ii]
 
-        return Preloading(list(shapes_new), list(bounding_indices_new))
+        return DataBoundsPreloading(list(shapes_new), list(bounding_indices_new))
