### 2025-05-02 3fb2c97
jwst: restructure Star preloading
