### 2025-05-02 3675d94
jwst: delete old parametric prior builder
