+++ src/instruments/jwst/parametric_model/parametric_prior.py
@@ -12,12 +12,12 @@ import jax.numpy as jnp
 import nifty8.re as jft
 
 DISTRIBUTION_MAPPING = {
-    'normal': (jft.normal_prior, ['mean', 'sigma']),
-    'log_normal': (jft.lognormal_prior, ['mean', 'sigma']),
-    'lognormal': (jft.lognormal_prior, ['mean', 'sigma']),
-    'uniform': (jft.uniform_prior, ['min', 'max']),
-    'delta': (lambda x: lambda _: x, ['mean']),
-    None: (lambda x: lambda _: x, ['mean'])
+    "normal": (jft.normal_prior, ["mean", "sigma"]),
+    "log_normal": (jft.lognormal_prior, ["mean", "sigma"]),
+    "lognormal": (jft.lognormal_prior, ["mean", "sigma"]),
+    "uniform": (jft.uniform_prior, ["min", "max"]),
+    "delta": (lambda x: lambda _: x, ["mean"]),
+    None: (lambda x: lambda _: x, ["mean"]),
 }
 
 
@@ -34,15 +34,15 @@ def _infer_shape(params: dict, shape: tuple):
     if shape != ():
         return shape
 
-    match params['distribution']:
-        case 'delta' | None:
-            return jnp.shape(params['mean'])
+    match params["distribution"]:
+        case "delta" | None:
+            return jnp.shape(params["mean"])
 
-        case 'uniform':
-            shp1, shp2 = map(jnp.shape, (params['min'], params['max']))
+        case "uniform":
+            shp1, shp2 = map(jnp.shape, (params["min"], params["max"]))
 
         case _:
-            shp1, shp2 = map(jnp.shape, (params['mean'], params['sigma']))
+            shp1, shp2 = map(jnp.shape, (params["mean"], params["sigma"]))
 
     # TODO: do some checks on compatibility of the two shapes
     return shp1
@@ -56,117 +56,16 @@ def _transform_setting(parameters: Union[dict, tuple]):
     distribution = parameters[0].lower()
 
     match distribution:
-        case 'uniform':
-            return dict(
-                distribution=distribution,
-                min=parameters[1],
-                max=parameters[2])
+        case "uniform":
+            return dict(distribution=distribution, min=parameters[1], max=parameters[2])
 
-        case 'delta' | None:
-            return dict(
-                distribution=distribution,
-                mean=parameters[1]
-            )
+        case "delta" | None:
+            return dict(distribution=distribution, mean=parameters[1])
 
         case _:
             return dict(
-                distribution=distribution,
-                mean=parameters[1],
-                sigma=parameters[2]
+                distribution=distribution, mean=parameters[1], sigma=parameters[2]
             )
-
-
-def build_parametric_prior(
-    domain_key: str,
-    parameters: Union[dict, tuple],
-    shape: Tuple[int] = ()
-) -> Callable:
-    """
-    Builds a parametric prior based on the specified distribution and
-    transformation.
-
-    This function constructs a prior distribution for a given model domain by
-    interpreting the provided `parameters` and selecting the appropriate
-    prior function based on the distribution specified therein.
-    The prior can be optionally transformed if a transformation
-    function is specified.
-
-    Parameters
-    ----------
-    domain_key : str
-        A string key identifying the domain of the model to which this prior
-        applies.
-    parameters : dict or tuple
-        A dictionary or tuple containing the prior configuration.
-        The dictionary should have a 'distribution' key, which specifies the
-        distribution type, and additional keys
-        required for the distribution (e.g., 'mean', 'std' for a Gaussian).
-        The required keys depend on the distribution.
-    shape : tuple of int, optional
-        A tuple representing the shape of the parameters.
-        This shape is applied to adjust the values of the prior parameters to
[…]
