### 2025-05-02 284c336
jwst: implemnt star model
