### 2025-05-03 b4d3cd9
jwst: making target jwst response mapped over data axis.
