+++ src/instruments/jwst/jwst_likelihoods.py
@@ -6,10 +6,7 @@ from .data.jwst_data import JwstData, DataMetaInformation
 from .data.data_loading import DataBoundsPreloading
 from .jwst_psf import load_psf_kernel
 from .filter_projector import FilterProjector, build_filter_projector
-from .rotation_and_shift.coordinates_correction import (
-    ShiftAndRotationCorrection,
-    # build_shift_and_rotation_correction,
-)
+from .rotation_and_shift.coordinates_correction import ShiftAndRotationCorrection
 from .plotting.residuals import ResidualPlottingInformation
 from .config_handler import (
     get_grid_extension_from_config,
@@ -17,6 +14,9 @@ from .config_handler import (
 from ..gaia.star_finder import load_gaia_stars_in_fov, join_tables
 from .plotting.plotting_sky import plot_sky_coords, plot_jwst_panels
 from .alignment.star_alignment import FilterAlignment
+from .alignment.star_fields import build_stars
+
+from .zero_flux_model import build_zero_flux_model
 
 # Parsing
 from .parse.jwst_psf import yaml_to_psf_kernel_config
@@ -44,13 +44,60 @@ from typing import Union, Iterator
 from dataclasses import dataclass, field
 
 
+def plot_test(filter_alignment, ii, jwst_data):
+    import astropy
+    from functools import partial
+    import matplotlib.pyplot as plt
+
+    stars = filter_alignment.get_stars(ii)
+    plot_position_stars = partial(
+        plot_sky_coords,
+        sky_coords=[star.position for star in stars],
+        labels=[star.id for star in stars],
+        behavior_index=lambda index, sky_coords: (
+            sky_coords if index == 0 else [sky_coords[index - 1]]
+        ),
+    )
+    jwst_wcs = astropy.wcs.WCS(jwst_data.wcs.to_header())
+
+    shape = (
+        int(
+            (
+                filter_alignment.alignment_meta.fov
+                / jwst_data.meta.pixel_distance.to(
+                    filter_alignment.alignment_meta.fov.unit
+                )
+            ).value
+        ),
+    ) * 2
+
+    data = [jwst_data.dm.data]
+    wcs = [jwst_wcs]
+    for star in stars:
+        minx, maxx, miny, maxy = star.bounding_indices(jwst_data, shape)
+        print(star.id, minx, maxx, miny, maxy)
+        wcs.append(jwst_wcs[miny : maxy + 1, minx : maxx + 1])
+        data.append(jwst_data.data_inside_extrema((minx, maxx, miny, maxy)))
+
+    mean = np.nanmean(jwst_data.dm.data)
+    fig, axs = plot_jwst_panels(
+        data,
+        wcs,
+        nrows=1,
+        ncols=len(data),
+        vmin=0.9 * mean,
+        vmax=1.1 * mean,
+        coords_plotter=plot_position_stars,
+    )
+
+
 @dataclass
 class FilterData:
+    meta: DataMetaInformation | None = None
     data: np.ndarray | list[np.ndarray] = field(default_factory=list)
     mask: np.ndarray | list[np.ndarray] = field(default_factory=list)
     std: np.ndarray | list[np.ndarray] = field(default_factory=list)
     psf: np.ndarray | list[np.ndarray] = field(default_factory=list)
-    meta: DataMetaInformation | list[DataMetaInformation] = field(default_factory=list)
     subsample_centers: SkyCoord | list[SkyCoord] = field(default_factory=list)
 
     def __len__(self):
@@ -92,6 +139,12 @@ class FilterData:
         for i in range(len(self)):
             yield self[i]
 
+    def add_and_check_meta_data(self, filter_data_meta: DataMetaInformation):
[…]
