+++ demos/configs/spt0418.yaml
@@ -313,7 +313,7 @@ telescope:
 
 
 files:
-  res_dir: './results/jwst_lens/cluster/13_06_restructuring_data_loading_08'
+  res_dir: './results/jwst_lens/cluster/14_01_single_target'
 
   step_type: 'cleanflickernoisestep'  # tweakregstep, cleanflickernoisestep, cal
 

[…]
