### 2025-05-03 ab75a5f
jwst: unload filter_alignmen
