### 2025-05-03 f5218aa
jwst: bugfix laoding data with float64
