### 2025-05-03 5ee3c0f
jwst: vmap psf kernel
