### 2025-05-03 037fdbf
jwst: bugfix vmap psf_kernel; testing
