### 2025-05-04 521e09c
jwst: tmp, loading & saving star data
