+++ src/instruments/jwst/jwst_likelihoods.py
@@ -139,11 +139,27 @@ class FilterData:
         for i in range(len(self)):
             yield self[i]
 
-    def add_and_check_meta_data(self, filter_data_meta: DataMetaInformation):
+    def add_or_check_meta_data(self, filter_data_meta: DataMetaInformation):
         if self.meta is None:
             self.meta = filter_data_meta
         else:
             assert self.meta == filter_data_meta
+
+    def append_observation(
+        self,
+        meta: DataMetaInformation,
+        data: np.ndarray,
+        mask: np.ndarray,
+        std: np.ndarray,
+        subsample_centers: SkyCoord,
+        psf: np.ndarray,
+    ):
+        self.add_or_check_meta_data(meta)
+        self.data.append(data)
+        self.mask.append(mask)
+        self.std.append(std)
+        self.subsample_centers.append(subsample_centers)
+        self.psf.append(psf)
 
 
 def build_jwst_likelihoods(
@@ -184,8 +200,6 @@ def build_jwst_likelihoods(
     target_plotting = ResidualPlottingInformation()
     likelihoods = []
     for filter_and_files in data_paths.filters:
-        filter_data = FilterData()
-
         target_preloading = DataBoundsPreloading()
 
         filter_alignment = FilterAlignment(
@@ -196,16 +210,22 @@ def build_jwst_likelihoods(
             number_of_observations=len(filter_and_files.filepaths),
         )
 
+        # NOTE : Preloading Data
         for ii, filepath in enumerate(filter_and_files.filepaths):
             print(ii, filepath)
             jwst_data = JwstData(filepath, subsample=data_subsample)
+            energy_name = filter_projector.get_key(jwst_data.pivot_wavelength)
+            if ii == 0:
+                previous_energy_name = energy_name
+            else:
+                assert energy_name == previous_energy_name
 
-            # filter_alignment.star_tables.append(
-            #     load_gaia_stars_in_fov(
-            #         jwst_data.wcs.world_corners(),
-            #         filter_alignment.alignment_meta.library_path,
-            #     )
-            # )
+            filter_alignment.star_tables.append(
+                load_gaia_stars_in_fov(
+                    jwst_data.wcs.world_corners(),
+                    filter_alignment.alignment_meta.library_path,
+                )
+            )
 
             target_preloading.append_shapes_and_bounds(
                 jwst_data,
@@ -231,16 +251,17 @@ def build_jwst_likelihoods(
         # stars(stars.init(key))
         # list(stars._target_ids)
 
+        target_data = FilterData()
+
+        stars = filter_alignment.get_stars()
+        stars_data = {star.id: FilterData() for star in stars}
+
         for ii, filepath in enumerate(filter_and_files.filepaths):
             logger.info(f"Loading: {filter_and_files.name} {ii} {filepath}")
 
             # Loading data, std, and mask.
             jwst_data = JwstData(filepath, subsample=data_subsample)
-            energy_name = filter_projector.get_key(jwst_data.pivot_wavelength)
-            if ii == 0:
-                previous_energy_name = energy_name
-            else:
-                assert energy_name == previous_energy_name
+            filter_alignment.boresight.append(jwst_data.get_boresight_world_coords())
 
             data, mask, std, data_subsampled_centers = (
                 jwst_data.bounding_data_mask_std_subpixel_by_bounding_indices(
@@ -249,21 +270,59 @@ def build_jwst_likelihoods(
                     yaml_to_corner_mask_configs(cfg[telescope_key]),
                 )
             )
-            print(data.shape, data_subsampled_centers.shape)
-
             psf = load_psf_kernel(
                 jwst_data=jwst_data,
                 target_center=grid.spatial.center,
                 config_parameters=psf_kernel_configs,
             )
+            target_data.append_observation(
+                meta=jwst_data.meta,
[…]
