### 2025-05-06 4521f64
jwst: star model from square point sources
