### 2025-05-06 967a2aa
jwst: tmp; build stars in jwst data
