### 2025-05-06 c4b846c
jwst: tmp testing ground for star algorihm
