### 2025-05-07 7842de4
jwst: fixing bilinear point source
