+++ src/instruments/jwst/alignment/star_model.py
@@ -1,8 +1,8 @@
-import jax
 import jax.numpy as jnp
 import nifty8.re as jft
 import numpy as np
-from jax import lax, vmap
+from jax import lax, vmap, Array
+
 
 from ..parametric_model.parametric_prior import build_parametric_prior_from_prior_config
 from ..parse.jwst_likelihoods import StarData
@@ -15,11 +15,50 @@ from ..rotation_and_shift.coordinates_correction import (
 )
 
 
-def square_point_source_in_sky_from_location(sky, location, flux):
-    loc_in_pixel, pixel_pos = jnp.modf(location[::-1] - jnp.array([0.5, 0.5]))
-    pixel_pos = pixel_pos.astype(int)
+def bilinear_point_source_evaluation(
+    sky: np.ndarray | Array,
+    location_xy: np.ndarray | Array,
+    flux: float | np.ndarray | Array,
+):
+    """Calculate the point source in sky using bilinear interpolation weights.
 
-    y, x = loc_in_pixel
+    Parameters
+    ----------
+    sky: np.ndarray | jax.Array
+        The sky to which the point source is added.
+    location_xy: np.ndarray
+        The location of the point source. Note:
+        1) Assuming the astropy convention that the center of the reference pixel has
+           the index (0, 0). Hence the upper-left corner (in `ij` indexing) of the
+           reference pixel has the index (-0.5, -0.5).
+        2) Assuming `xy` indexing of the location, hence x-position of the location is
+           in the first index of the location array, and the y-position in the second
+           index of the location, i.e.: (x, y).
+           Note: The sky array is `ij` indexed, hence the first index tells the
+           y-position. Hence, we have to revert the pixel position, see `pixel_pos`.
+           Furthermore, the bilinear weights are evaluated accordingly:
+           w00           w01
+              +----|-x--+
+              |    |    |
+              |    |    |
+              +---------+
+              y    | .  |
+              |    |    |
+              +----|----+
+           w10           w11
+    flux: float | np.ndarray | Array
+        The flux value of the point source.
+
+    Warning
+    -------
+    The implementation breaks down for locations outside the grid spanned by the center
+    of the sky pixels. Hence, all negative locations, e.g. (-0.5, -0.5), and all
+    locations beyond (shape-1, shape-1), lead to erroneous evaluations of the flux.
+    """
+    loc_in_pixel, pixel_pos = jnp.modf(location_xy)
+    pixel_pos = pixel_pos.astype(int)[::-1]
+
+    x, y = loc_in_pixel
 
     # bilinear weights
     w00 = (1.0 - x) * (1.0 - y)
@@ -30,9 +69,9 @@ def square_point_source_in_sky_from_location(sky, location, flux):
     point_source = jnp.array(((w00, w01), (w10, w11)))
     point_source = flux * point_source
 
+    # Indexing into array
     sky_slice = lax.dynamic_slice(sky, pixel_pos, point_source.shape)
-    sky = lax.dynamic_update_slice(sky, sky_slice + point_source, pixel_pos)
-    return sky
+    return lax.dynamic_update_slice(sky, sky_slice + point_source, pixel_pos)
 
 
 class StarInData(jft.Model):
@@ -45,7 +84,7 @@ class StarInData(jft.Model):
         self._skies = skies
         self.brightness = brightness
         self.location = location
-        self._call = vmap(square_point_source_in_sky_from_location, (0, 0, None))
[…]
