### 2025-05-07 8b3e586
jwst: re-implementing sub-pixel position of the star.
