### 2025-05-07 fc7419f
jwst: deleting star_model testing
