+++ src/instruments/jwst/alignment/star_model.py
@@ -132,55 +132,3 @@ def build_star_in_data(
     )
 
     return StarInData(skies, brightness, location_of_star_in_data_subpixels)
-
-
-if __name__ == "__main__":
-    import numpy as np
-    import astropy.units as u
-    from astropy.coordinates import SkyCoord
-    from ..plotting.plotting_sky import plot_sky_coords, plot_jwst_panels
-    from functools import partial
-    from astropy.wcs import WCS
-    import matplotlib.pyplot as plt
-
-    position = np.array((0.5, 1.0))  # xy indexing
-
-    pix_scale = 1.0 * u.arcsec  # desired pixel size
-    deg_per_pix = pix_scale.to(u.deg).value  # ≃ 2.777…×10⁻⁴ deg
-
-    # -- data already given -------------------------------------------------
-    pix_zero = SkyCoord(ra=11.0 * u.deg, dec=41.0 * u.deg, frame="icrs")
-
-    # -- 4×4 WCS with 1″ pixels --------------------------------------------
-    N = 4
-    wcs = WCS(naxis=2)
-    wcs.wcs.crpix = [1, 1]  # centre pixel (0, 0)
-    wcs.wcs.cdelt = [-deg_per_pix, deg_per_pix]  # RA decreases to the right
-    wcs.wcs.crval = [pix_zero.ra.deg, pix_zero.dec.deg]
-    wcs.wcs.ctype = ["RA---TAN", "DEC--TAN"]  # gnomonic projection
-
-    # convert that pixel to RA/Dec with the same WCS
-    star_sc = wcs.pixel_to_world(*position)
-
-    x_star, y_star = wcs.world_to_pixel(pix_zero)
-    print(x_star, y_star)
-
-    # pixel centres
-    yy, xx = np.mgrid[:N, :N] + 0.5
-    coords_small = wcs.pixel_to_world(xx, yy)  # 4×4 SkyCoord
-
-    data = np.arange(N * N).reshape(N, N) / 16
-    sky = np.zeros((N, N), dtype=float)
-    star = bilinear_point_source_evaluation(jnp.array(sky), position, 1.0)
-
-    star_position_plot = partial(plot_sky_coords, sky_coords=[star_sc])
-    fig, axes = plot_jwst_panels(
-        [data, star],
-        [wcs, wcs],
-        vmax=data.max(),
-        nrows=1,
-        ncols=2,
-        coords_plotter=star_position_plot,
-    )
-    axes[0].scatter(*position)
-    plt.show()
