### 2025-05-07 0cd7da5
jwst: re-implementing data shapes, adjusting for cleaner slicin
