### 2025-05-08 10240c6
jwst: changing data loading index to numpy shape slicing (+1)
