### 2025-05-08 84bd6d3
jwst: bugfix condition for star outside of data is `or` on both axes
