### 2025-05-08 a258e6b
jwst: implementation of the star likelihoods done
