+++ src/instruments/jwst/jwst_likelihoods.py
@@ -1,20 +1,14 @@
 from ...likelihood import build_gaussian_likelihood
 from ...grid import Grid
-from ...wcs.wcs_jwst_data import WcsJwstData
-from .jwst_response import build_jwst_response, build_jwst_response_stars
+from ..gaia.star_finder import load_gaia_stars_in_fov
+from .jwst_response import build_jwst_response, build_sky_to_subsampled_data
 from .data.jwst_data import JwstData
 from .data.data_loading import DataBoundsPreloading
 from .jwst_psf import load_psf_kernel
 from .filter_projector import FilterProjector, build_filter_projector
-from .rotation_and_shift.coordinates_correction import (
-    ShiftAndRotationCorrection,
-)
+from .rotation_and_shift.coordinates_correction import ShiftAndRotationCorrection
 from .plotting.residuals import ResidualPlottingInformation
-from .config_handler import (
-    get_grid_extension_from_config,
-)
-from ..gaia.star_finder import load_gaia_stars_in_fov, join_tables
-from .plotting.plotting_sky import plot_sky_coords, plot_jwst_panels
+from .config_handler import get_grid_extension_from_config
 from .alignment.star_alignment import FilterAlignment
 from .alignment.star_model import build_star_in_data
 from .zero_flux_model import build_zero_flux_model
@@ -54,7 +48,7 @@ def build_jwst_likelihoods(
     telescope_key: str = "telescope",
     sky_unit: u.Unit | None = None,
 ) -> Union[jft.Likelihood, FilterProjector, dict]:
-    """Build the jwst likelihood according to the config and grid."""
+    """Build the jwst likelihood_target according to the config and grid."""
 
     filter_projector = build_filter_projector(
         sky_model, grid, cfg[files_key]["filter"].keys()
@@ -72,6 +66,7 @@ def build_jwst_likelihoods(
     sky_meta = SkyMetaInformation(
         grid_extension=get_grid_extension_from_config(cfg[telescope_key], grid),
         unit=sky_unit,
+        dvol=grid.spatial.dvol,
     )
 
     gaia_alignment_meta_data = FilterAlignmentMeta.from_yaml_dict(
@@ -81,7 +76,8 @@ def build_jwst_likelihoods(
     data_paths = DataFilePaths.from_yaml_dict(cfg[files_key])
 
     target_plotting = ResidualPlottingInformation()
-    likelihoods = []
+    target_filter_likelihoods = []
+    stars_alignment_likelihoods = {}
     for filter_and_files in data_paths.filters:
         target_preloading = DataBoundsPreloading()
 
@@ -110,11 +106,12 @@ def build_jwst_likelihoods(
                 )
             )
 
+            bounding_indices = jwst_data.wcs.bounding_indices_from_world_extrema(
+                grid.spatial.world_corners(extension_value=sky_meta.grid_extension)
+            )
             target_preloading.append_shapes_and_bounds(
-                jwst_data,
-                sky_corners=grid.spatial.world_corners(
-                    extension_value=sky_meta.grid_extension
-                ),
+                jwst_data=jwst_data,
+                bounding_indices=bounding_indices,
             )
 
         target_preloading = target_preloading.align_shapes()
@@ -203,55 +200,28 @@ def build_jwst_likelihoods(
             rotation_center=SkyCoord(filter_alignment.boresight),
         )
 
+        target_in_subsampled_data = build_sky_to_subsampled_data(
+            sky_domain={energy_name: filter_projector.target[energy_name]},
+            data_subsampled_centers=target_data.subsample_centers,
+            sky_wcs=grid.spatial,
+            rotation_and_shift_algorithm=rotation_and_shift_algorithm,
[…]
