+++ src/instruments/jwst/jwst_response.py
@@ -18,8 +18,7 @@ from .parse.rotation_and_shift.rotation_and_shift import LinearConfig, NufftConf
 from .parse.jwst_response import SkyMetaInformation
 from .data.jwst_data import DataMetaInformation
 
-from .alignment.star_alignment import Star
-from ...wcs.wcs_jwst_data import WcsJwstData
+from .alignment.star_model import StarInData
 from ...wcs.wcs_astropy import WcsAstropy
 from .integration.unit_conversion import build_unit_conversion
 from .integration.integration import integration_factory
@@ -28,44 +27,35 @@ from .masking.build_mask import build_mask
 from .rotation_and_shift import RotationAndShift, build_rotation_and_shift
 from .rotation_and_shift.coordinates_correction import (
     build_coordinates_corrected_for_field,
-    build_coordinates_corrected_for_stars,
     ShiftAndRotationCorrection,
 )
-from .parse.jwst_likelihoods import StarData
 
 
 class JwstResponse(jft.Model):
     """
-    A that connects observational data to the corresponding sky and
-    instrument models.
+    Linear response model that connects jwst observational data to the corresponding a
+    sky model.
 
-    This class models a data pipeline that includes rotation, shifting,
-    PSF application, integration, and masking, with an optional zero-flux model.
+    Schematic pipeline:
+    psf | unit_conversion | integrate | zero flux | mask
     """
 
     def __init__(
         self,
-        sky_domain: dict,
-        rotation_and_shift: RotationAndShift | Callable[ArrayLike, ArrayLike],
+        sky_model: jft.Model | RotationAndShift | StarInData,
         psf: Callable[ArrayLike, ArrayLike],
         unit_conversion: Callable[ArrayLike, ArrayLike],
         integrate: Callable[ArrayLike, ArrayLike],
         zero_flux_model: jft.Model | None,
         mask: Callable[ArrayLike, ArrayLike],
     ):
-        """
-        Initialize the DataModel with components for various data
+        """Initialize the Jwst response with different components of linear
         transformations.
 
         Parameters
         ----------
-        sky_domain : dict
-            A dictionary defining the sky domain, with a single key
-            corresponding to the internal target of the sky model.
-            This defines the input space of the data.
-        rotation_and_shift : RotationAndShift | None
-            A model that applies rotation and shift transformations
-            to the input data.
+        sky_model : jft.Model | RotationAndShift | StarInData,
+            A model has as output the sky in the frame of the data.
         psf : callable
             A function that applies a point spread function (PSF) to the
             input data.
@@ -78,33 +68,22 @@ class JwstResponse(jft.Model):
             If provided, its output is added to the domain model's output.
         mask : callable
             A function that applies a mask to the final output.
-
-        Raises
-        ------
-        AssertionError
-            If `sky_domain` is not a dictionary or if it contains
-            more than one key.
         """
-        need_sky_key = "Need to provide an internal key to the target of the sky model"
-        assert isinstance(sky_domain, dict), need_sky_key
-        assert len(sky_domain.keys()) == 1, need_sky_key
 
-        self.rotation_and_shift = rotation_and_shift
+        self.sky_model = sky_model
         self.psf = psf
         self.unit_conversion = unit_conversion
         self.integrate = integrate
         self.zero_flux_model = zero_flux_model
         self.mask = mask
 
-        domain = sky_domain
-        if rotation_and_shift is not None:
[…]
