### 2025-05-08 ef06acc
likelihood: adjust gaussian likelihood to amend model directly
