### 2025-05-09 2ca3786
jwst: subsample differnet for alignment and target
