+++ src/instruments/jwst/data/jwst_data.py
@@ -30,7 +30,6 @@ except ImportError:
 
 @dataclass
 class DataMetaInformation:
-    subsample: int
     unit: units.Unit
     dvol: units.Quantity
     pixel_distance: units.Quantity
@@ -39,7 +38,7 @@ class DataMetaInformation:
 class JwstData:
     """Class to contain JWST data metadata."""
 
-    def __init__(self, filepath: str, subsample: int = 1):
+    def __init__(self, filepath: str):
         """
         Initializes the JwstData class.
 
@@ -55,7 +54,6 @@ class JwstData:
         self.camera = self.dm.meta.instrument.name.upper()
 
         self.meta: DataMetaInformation = DataMetaInformation(
-            subsample=subsample,
             unit=units.Unit(self.dm.meta.bunit_data),
             dvol=get_dvol(self.filter),
             pixel_distance=get_pixel_distance(self.filter),
