### 2025-05-09 c4576db
jwst: tmp psf as model
