+++ src/instruments/jwst/jwst_response.py
@@ -22,7 +22,7 @@ from .alignment.star_model import StarInData
 from ...wcs.wcs_astropy import WcsAstropy
 from .integration.unit_conversion import build_unit_conversion
 from .integration.integration import integration_factory
-from .jwst_psf import build_psf_operator
+from .jwst_psf import build_psf_operator, PsfModel
 from .masking.build_mask import build_mask
 from .rotation_and_shift import RotationAndShift, build_rotation_and_shift
 from .rotation_and_shift.coordinates_correction import (
@@ -43,7 +43,7 @@ class JwstResponse(jft.Model):
     def __init__(
         self,
         sky_model: jft.Model | RotationAndShift | StarInData,
-        psf: Callable[ArrayLike, ArrayLike],
+        psf: Callable[ArrayLike, ArrayLike] | PsfModel,
         unit_conversion: Callable[ArrayLike, ArrayLike],
         integrate: Callable[ArrayLike, ArrayLike],
         zero_flux_model: jft.Model | None,
@@ -78,13 +78,18 @@ class JwstResponse(jft.Model):
         self.mask = mask
 
         domain = sky_model.domain
+        if isinstance(psf, PsfModel):
+            domain = domain | psf.domain[1]
         if zero_flux_model is not None:
             domain = domain | zero_flux_model.domain
         super().__init__(domain=domain)
 
     def __call__(self, x):
         out = self.sky_model(x)
-        out = self.psf(out)
+        if isinstance(self.psf, PsfModel):
+            out = self.psf((out, x))
+        else:
+            out = self.psf(out)
         out = self.unit_conversion(out)
         out = self.integrate(out)
         if self.zero_flux_model is not None:
@@ -145,7 +150,7 @@ def build_jwst_response(
     data_meta: DataMetaInformation,
     data_subsample: int,
     sky_meta: SkyMetaInformation,
-    psf: np.ndarray | None,
+    psf: np.ndarray | jft.Model | None,
     zero_flux_model: jft.Model | None,
     data_mask: ArrayLike | None,
 ) -> JwstResponse:
@@ -177,7 +182,7 @@ def build_jwst_response(
         The mask on the data
     """
 
-    psf = build_psf_operator(psf)
+    psf = build_psf_operator(psf, sky_in_subsampled_data.target)
 
     unit_conversion = build_unit_conversion(
         sky_unit=sky_meta.unit,
