### 2025-05-09 f87eef3
jwst: delete coordinate outside of data.
