### 2025-05-09 fc82940
jwst: some tests with different psf modifications of the stars.
