+++ src/instruments/jwst/jwst_psf.py
@@ -209,26 +209,66 @@ def load_psf_kernel(
     return psf
 
 
-def multiple_psf_operator(
-    vmap_fftconvolve: Callable[ArrayLike, ArrayLike],
-    psf_kernel: ArrayLike,
-):
-    return lambda x: vmap_fftconvolve(x, psf_kernel)
+class PsfStatic(jft.Model):
+    """Implements the convolution by a static psf kernel"""
+
+    def __init__(
+        self, psf_kernel: np.ndarray | None, sky_shape_with_dtype: jft.ShapeWithDtype
+    ):
+        """
+        Parameters
+        ----------
+        psf_kernel: np.ndarray | None
+            If None, the apply will just return the input field.
+            Else, the input field will by convolved by the `psf_kernel`.
+        sky_shape_with_dtype: jft.ShapeWithDtype
+            The `ShapeWithDtype` of the sky.
+        """
+
+        self._kernel = psf_kernel
+
+        if len(psf_kernel.shape) == 2:
+            self._convolve = partial(fftconvolve, mode="same")
+        elif len(psf_kernel.shape) == 3:
+            self._convolve = vmap(Partial(fftconvolve, mode="same"), in_axes=(0, 0))
+        else:
+            raise ValueError("Unknown psf_kernel shape")
+
+        super().__init__(domain=(sky_shape_with_dtype, {}))
+
+    def __call__(self, x):
+        field, _ = x
+        if self._kernel is None:
+            return field
+        return self._convolve(field, self._kernel)
 
 
-class PsfModel(jft.Model):
-    def __init__(self, psf_model: jft.Model, sky_shape_and_dtype: jft.ShapeWithDtype):
+class PsfDynamic(jft.Model):
+    """Implements the convolution by a dynamic psf kernel, i.e. a psf kernel that is
+    learned."""
+
+    def __init__(self, psf_model: jft.Model, sky_shape_with_dtype: jft.ShapeWithDtype):
+        """
+        Parameters
+        ----------
+        psf_kernel: np.ndarray | None
+            If None, the apply will just return the input field.
+            Else, the input field will by convolved by the `psf_kernel`.
+        sky_shape_with_dtype: jft.ShapeWithDtype
+            The `ShapeWithDtype` of the sky.
+        """
         assert isinstance(psf_model, jft.Model)
         self.model = psf_model
 
         if len(psf_model.target.shape) == 2:
             self._convolve = partial(fftconvolve, mode="same")
-
         elif len(psf_model.target.shape) == 3:
             self._convolve = vmap(Partial(fftconvolve, mode="same"), in_axes=(0, 0))
+        else:
+            raise ValueError("Unknown psf_kernel shape")
 
         super().__init__(
-            domain=(sky_shape_and_dtype, psf_model.domain),
+            domain=(sky_shape_with_dtype, psf_model.domain),
             white_init=True,
         )
 
@@ -238,55 +278,55 @@ class PsfModel(jft.Model):
         return self._convolve(field, psf_kernel / psf_kernel.sum())
 
 
-def build_psf_operator(
-    psf_kernel: np.ndarray | jft.Model | None,
-    sky_shape_and_dtype: jft.ShapeWithDtype | None = None,
-):
-    if isinstance(psf_kernel, np.ndarray) or psf_kernel is None:
-        return build_psf_operator_from_static_kernel(psf_kernel)
+class PsfModel(jft.Model):
+    def __init__(self, psf_kernel: np.ndarray, modification_model: jft.Model):
+        self.static = psf_kernel
+        self.modification = modification_model
 
-    assert sky_shape_and_dtype is not None
-    return PsfModel(psf_kernel, sky_shape_and_dtype)
[…]
