+++ src/instruments/jwst/jwst_likelihoods.py
@@ -4,7 +4,7 @@ from ..gaia.star_finder import load_gaia_stars_in_fov
 from .jwst_response import build_jwst_response, build_sky_to_subsampled_data
 from .data.jwst_data import JwstData
 from .data.data_loading import DataBoundsPreloading
-from .jwst_psf import load_psf_kernel, build_psf_model
+from .jwst_psf import load_psf_kernel, build_psf_modification_model_strategy, PsfModel
 from .filter_projector import FilterProjector, build_filter_projector
 from .rotation_and_shift.coordinates_correction import ShiftAndRotationCorrection
 from .plotting.residuals import ResidualPlottingInformation
@@ -134,24 +134,22 @@ def build_jwst_likelihoods(
                 grid.spatial,
                 yaml_to_corner_mask_configs(cfg[telescope_key]),
             )
-            data_subsampled_centers = jwst_data.data_subpixel_centers(
-                target_preloading.bounding_indices[observation_id],
-                subsample=target_subsample,
-            )
-            psf = load_psf_kernel(
-                jwst_data=jwst_data,
-                subsample=target_subsample,
-                target_center=grid.spatial.center,
-                config_parameters=psf_kernel_configs,
-            )
             target_data.append_observation(
                 meta=jwst_data.meta,
                 subsample=target_subsample,
                 data=data,
                 mask=mask,
                 std=std,
-                subsample_centers=data_subsampled_centers,
-                psf=psf,
+                subsample_centers=jwst_data.data_subpixel_centers(
+                    target_preloading.bounding_indices[observation_id],
+                    subsample=target_subsample,
+                ),
+                psf=load_psf_kernel(
+                    jwst_data=jwst_data,
+                    subsample=target_subsample,
+                    target_center=grid.spatial.center,
+                    config_parameters=psf_kernel_configs,
+                ),
             )
 
             for star in filter_alignment.get_stars(observation_id):
@@ -159,7 +157,9 @@ def build_jwst_likelihoods(
                     filter_alignment.alignment_meta.fov.to(u.arcsec)
                     / jwst_data.meta.pixel_distance.to(u.arcsec)
                 ).value
-                fov_pixel = (int(np.round(fov_pixel)),) * 2
+                fov_pixel = np.array((int(np.round(fov_pixel)),) * 2)
+                if (fov_pixel % 2).sum() == 0:
+                    fov_pixel += 1
 
                 bounding_indices = star.bounding_indices(jwst_data, fov_pixel)
                 data, mask, std = jwst_data.bounding_data_mask_std_by_bounding_indices(
@@ -168,18 +168,6 @@ def build_jwst_likelihoods(
                         cfg[telescope_key]
                     ),
                 )
-                psf = load_psf_kernel(
-                    jwst_data=jwst_data,
-                    subsample=alignment_subsample,
-                    target_center=star.position,
-                    config_parameters=psf_kernel_configs,
-                )
-                star_in_subsampled_pixels = star.pixel_position_in_subsampled_data(
-                    jwst_data.wcs,
-                    min_row=bounding_indices[0],
-                    min_column=bounding_indices[2],
-                    subsample_factor=alignment_subsample,
-                )
 
                 stars_data[star.id].append_observation(
                     meta=jwst_data.meta,
@@ -187,9 +175,19 @@ def build_jwst_likelihoods(
                     data=data,
                     mask=mask,
                     std=std,
+                    psf=load_psf_kernel(
+                        jwst_data=jwst_data,
+                        subsample=alignment_subsample,
+                        target_center=star.position,
+                        config_parameters=psf_kernel_configs,
+                    ),
[…]
