### 2025-05-12 e711a14
jwst: tmp source reconstruction testing
