+++ src/instruments/jwst/jwst_likelihoods.py
@@ -102,12 +102,13 @@ def build_jwst_likelihoods(
             else:
                 assert energy_name == previous_energy_name
 
-            filter_alignment.star_tables.append(
-                load_gaia_stars_in_fov(
-                    jwst_data.wcs.world_corners(),
-                    filter_alignment.alignment_meta.library_path,
+            if False:
+                filter_alignment.star_tables.append(
+                    load_gaia_stars_in_fov(
+                        jwst_data.wcs.world_corners(),
+                        filter_alignment.alignment_meta.library_path,
+                    )
                 )
-            )
 
             bounding_indices = jwst_data.wcs.bounding_indices_from_world_extrema(
                 grid.spatial.world_corners(extension_value=sky_meta.grid_extension)
@@ -118,10 +119,11 @@ def build_jwst_likelihoods(
             )
 
         target_preloading = target_preloading.align_shapes()
-
         target_data = TargetData()
-        stars = filter_alignment.get_stars()
-        stars_data = {star.id: StarData() for star in stars}
+
+        if False:
+            stars = filter_alignment.get_stars()
+            stars_data = {star.id: StarData() for star in stars}
 
         for observation_id, filepath in enumerate(filter_and_files.filepaths):
             logger.info(f"Loading: {filter_and_files.name} {observation_id} {filepath}")
@@ -153,44 +155,49 @@ def build_jwst_likelihoods(
                 ),
             )
 
-            for star in filter_alignment.get_stars(observation_id):
-                fov_pixel = (
-                    filter_alignment.alignment_meta.fov.to(u.arcsec)
-                    / jwst_data.meta.pixel_distance.to(u.arcsec)
-                ).value
-                fov_pixel = np.array((int(np.round(fov_pixel)),) * 2)
-                if (fov_pixel % 2).sum() == 0:
-                    fov_pixel += 1
+            if False:
+                for star in filter_alignment.get_stars(observation_id):
+                    fov_pixel = (
+                        filter_alignment.alignment_meta.fov.to(u.arcsec)
+                        / jwst_data.meta.pixel_distance.to(u.arcsec)
+                    ).value
+                    fov_pixel = np.array((int(np.round(fov_pixel)),) * 2)
+                    if (fov_pixel % 2).sum() == 0:
+                        fov_pixel += 1
 
-                bounding_indices = star.bounding_indices(jwst_data, fov_pixel)
-                data, mask, std = jwst_data.bounding_data_mask_std_by_bounding_indices(
-                    row_minmax_column_minmax=bounding_indices,
-                    additional_masks_corners=yaml_to_corner_mask_configs(
-                        cfg[telescope_key]
-                    ),
-                )
+                    bounding_indices = star.bounding_indices(jwst_data, fov_pixel)
+                    data, mask, std = (
+                        jwst_data.bounding_data_mask_std_by_bounding_indices(
+                            row_minmax_column_minmax=bounding_indices,
+                            additional_masks_corners=yaml_to_corner_mask_configs(
+                                cfg[telescope_key]
+                            ),
+                        )
+                    )
 
-                stars_data[star.id].append_observation(
-                    meta=jwst_data.meta,
-                    subsample=alignment_subsample,
-                    data=data,
-                    mask=mask,
-                    std=std,
-                    psf=load_psf_kernel(
-                        jwst_data=jwst_data,
+                    stars_data[star.id].append_observation(
+                        meta=jwst_data.meta,
                         subsample=alignment_subsample,
-                        target_center=star.position,
-                        config_parameters=psf_kernel_configs,
-                    ),
-                    sky_array=np.zeros([s * alignment_subsample for s in data.shape]),
[…]

[…]
