### 2025-05-14 6bea79e
jwst: seperate out data preloading
