+++ src/instruments/jwst/alignment/filter_alignment.py
@@ -1,0 +1,53 @@
+from dataclasses import dataclass, field
+
+import numpy as np
+from astropy import units as u
+from astropy.coordinates import SkyCoord
+from astropy.table import Table
+
+from .star_alignment import StarAlignment
+from ..parse.alignment.star_alignment import StarAlignmentMeta
+from ..parse.parametric_model.parametric_prior import (
+    prior_config_factory,
+)
+from ..parse.rotation_and_shift.coordinates_correction import (
+    ROTATION_KEY,
+    ROTATION_UNIT_KEY,
+    SHIFT_KEY,
+    SHIFT_UNIT_KEY,
+    CoordinatesCorrectionPriorConfig,
+)
+
+
+DEFAULT_KEY = "default"
+
+
+@dataclass
+class FilterAlignment:
+    filter_name: str
+    correction_prior: CoordinatesCorrectionPriorConfig | None = None
+    boresight: list[SkyCoord] = field(default_factory=list)
+    star_alignment: StarAlignment | None = None
+
+    def load_correction_prior(self, raw: dict, number_of_observations: int):
+        if self.filter_name in raw:
+            config = raw[self.filter_name]
+        else:
+            config = raw[DEFAULT_KEY]
+
+        self.correction_prior = CoordinatesCorrectionPriorConfig(
+            shift=prior_config_factory(
+                config[SHIFT_KEY], shape=(number_of_observations, 2)
+            ),
+            rotation=prior_config_factory(
+                config[ROTATION_KEY], shape=(number_of_observations, 1)
+            ),
+            shift_unit=getattr(u, raw[SHIFT_UNIT_KEY]),
+            rotation_unit=getattr(u, raw[ROTATION_UNIT_KEY]),
+        )
+
+    def load_star_alignment(self, star_alignment_meta: StarAlignmentMeta | None):
+        if star_alignment_meta is None:
+            return None
+
+        self.star_alignment = StarAlignment(alignment_meta=star_alignment_meta)
