### 2025-05-15 f08dcee
jwst: replace yaml_to_psf_kernel_config, with classmethod
