### 2025-05-15 cf64f6a
jwst: replace yaml_to_zero_flux_prior_config with classmethod
