### 2025-05-15 64a30aa
jwst: restructure, creating separate loading data process
