+++ src/instruments/jwst/jwst_likelihoods.py
@@ -1,19 +1,17 @@
 from ...likelihood import build_gaussian_likelihood
 from ...grid import Grid
 from .jwst_response import build_jwst_response, build_sky_to_subsampled_data
-from .data.jwst_data import JwstData
-from .data.preloading.target import DataBounds, target_preloading
-from .psf.jwst_kernel import load_psf_kernel
-from .psf.psf_learning import build_psf_modification_model_strategy, LearnablePsf
 from .filter_projector import FilterProjector, build_filter_projector
 from .rotation_and_shift.coordinates_correction import ShiftAndRotationCorrection
 from .plotting.residuals import ResidualPlottingInformation
 from .plotting.alignment import FilterAlignmentPlottingInformation
 from .config_handler import get_grid_extension_from_config
 from .alignment.filter_alignment import FilterAlignment
+from .alignment.star_alignment import StarAlignment
 from .alignment.star_model import build_star_in_data
 from .zero_flux_model import build_zero_flux_model
 
+from .data.loading.data_loading import data_loading
 from .data.preloading.preloading import data_preloading
 
 # Parsing
@@ -22,17 +20,16 @@ from .parse.rotation_and_shift.rotation_and_shift import (
     rotation_and_shift_algorithm_config_factory,
 )
 from .parse.jwst_response import SkyMetaInformation
-from .parse.masking.data_mask import yaml_to_corner_mask_configs
 from .parse.alignment.star_alignment import StarAlignmentMeta
 from .parse.data.data_loading import DataFilePaths
-from .parse.jwst_likelihoods import TargetData, StarData
 from .parse.jwst_psf import JwstPsfKernelConfig
+from .parse.masking.data_mask import ExtraMasks
 
+from .psf.psf_learning import build_psf_modification_model_strategy, LearnablePsf
 
 # Libraries
 import jax.numpy as jnp
 import nifty8.re as jft
-from nifty8.logger import logger
 import numpy as np
 
 # std
@@ -63,7 +60,6 @@ def build_jwst_likelihoods(
     rotation_and_shift_algorithm = rotation_and_shift_algorithm_config_factory(
         cfg[telescope_key]["rotation_and_shift"]
     )
-    target_subsample = cfg[telescope_key]["rotation_and_shift"]["subsample"]
     sky_meta = SkyMetaInformation(
         grid_extension=get_grid_extension_from_config(cfg[telescope_key], grid),
         unit=sky_unit,
@@ -73,6 +69,7 @@ def build_jwst_likelihoods(
     psf_kernel_configs = JwstPsfKernelConfig.from_yaml_dict(
         cfg[telescope_key].get("psf")
     )
+    extra_masks = ExtraMasks.from_yaml_dict(cfg[telescope_key])
 
     gaia_alignment_meta_data = StarAlignmentMeta.from_yaml_dict(
         cfg[telescope_key].get("gaia_alignment")
@@ -97,91 +94,15 @@ def build_jwst_likelihoods(
             filter_and_filepaths, grid, sky_meta, filter_alignment
         )
 
-        target_data = TargetData()
-
-        if False:
-            stars = filter_alignment.get_stars()
-            stars_data = {star.id: StarData() for star in stars}
-
-        for observation_id, filepath in enumerate(filter_and_filepaths.filepaths):
-            logger.info(
-                f"Loading: {filter_and_filepaths.filter} {observation_id} {filepath}"
-            )
-
-            # Loading data, std, and mask.
-            jwst_data = JwstData(filepath)
-            filter_alignment.boresight.append(jwst_data.get_boresight_world_coords())
-
-            data, mask, std = jwst_data.bounding_data_mask_std_by_bounding_indices(
-                target_bounds.bounds[observation_id],
-                grid.spatial,
[…]
