+++ src/instruments/jwst/data/preloading/star_alignment.py
@@ -1,13 +1,12 @@
 from ....gaia.star_finder import load_gaia_stars_in_fov
-from ...alignment.filter_alignment import FilterAlignment
+from ...alignment.star_alignment import StarAlignment
 from ..jwst_data import JwstData
 
 
-def star_alignment_preloading(filter_alignment: FilterAlignment, jwst_data: JwstData):
-    filter_alignment.star_alignment.star_tables.append(
+def star_alignment_preloading(star_alignment: StarAlignment, jwst_data: JwstData):
+    star_alignment.star_tables.append(
         load_gaia_stars_in_fov(
             jwst_data.wcs.world_corners(),
-            filter_alignment.alignment_meta.library_path,
+            star_alignment.alignment_meta.library_path,
         )
     )
-    return filter_alignment
