### 2025-05-16 72ce3c3
jwst: tmp, starting to implement different model controll
