+++ demos/configs/spt0418.yaml
@@ -40,44 +40,44 @@ sky:
     energy_unit: eV
     energy_bin:
       e_min:
-        # - 0.054 # F2100W
+        - 0.054 # F2100W
 
         - 0.068 # F1800W  # mistake here actually 0.064
 
-        # - 0.075 # F1500W
-        # - 0.093 # F1280W  # mistake here actually 0.088
-        # - 0.114 # F1000W
-        # - 0.143 # F770W
-        # - 0.201 # F560W
+        - 0.075 # F1500W
+        - 0.093 # F1280W  # mistake here actually 0.088
+        - 0.114 # F1000W
+        - 0.143 # F770W
+        - 0.201 # F560W
 
         - 0.249 # F444W
 
-        # - 0.319 # F356W  # mistake here actually 0.311
-        # - 0.396 # F277W
-        # - 0.557 # F200W
-        # - 0.743 # F150W
+        - 0.319 # F356W  # mistake here actually 0.311
+        - 0.396 # F277W
+        - 0.557 # F200W
+        - 0.743 # F150W
 
         - 0.967 # F115W
       e_max:
-        # - 0.067 # F2100W
+        - 0.067 # F2100W
 
         - 0.075 # F1800W
 
-        # - 0.092 # F1500W
-        # - 0.107 # F1280W
-        # - 0.137 # F1000W
-        # - 0.188 # F770W
-        # - 0.245 # F560W
+        - 0.092 # F1500W
+        - 0.107 # F1280W
+        - 0.137 # F1000W
+        - 0.188 # F770W
+        - 0.245 # F560W
 
         - 0.319 # F444W
 
-        # - 0.395 # F356W
-        # - 0.512 # F277W
-        # - 0.707 # F200W
-        # - 0.932 # F150W
+        - 0.395 # F356W
+        - 0.512 # F277W
+        - 0.707 # F200W
+        - 0.932 # F150W
 
         - 1.224 # F115W
-      reference_bin: 1 # F444W
+      reference_bin: 7 # F444W
 
   model_fixing_pointing:
     source:
@@ -286,13 +286,16 @@ telescope:
     f1800w: ['lognormal', 92.0, 1.5]
     f2100w: ['lognormal', 226.0, 1.0]
 
-  # gaia_alignment:
-  #   fov: 3.5arcsec
-  #   shape: [129, 129]
-  #   star_light: ['lognormal', 1000, 10000]
-  #   subsample: 1
-  #   library_path: './data/GAIA/library_stars'
-  #   exclude_source_id: [4788834899488736128, 4788858367190047104]
+  gaia_alignment:
+    fov: 3.5arcsec
+    shape: [129, 129]
+    star_light: ['lognormal', 1000, 10000]
+    subsample: 1
+    library_path: './data/GAIA/library_stars'
+    exclude_source_id: [4788834899488736128, 4788858367190047104]
+
+  target: 
+    subsample: 3
 
   rotation_and_shift:
     # nufft: 
@@ -302,9 +305,8 @@ telescope:
       mode: 'constant'
       order: 1
 
-    subsample: 3
-
     correction_priors:
+      model: 'shift' # shift, rshift
       rotation_unit: deg
       shift_unit: arcsec
       default:
@@ -313,9 +315,9 @@ telescope:
 
 
 files:
-  res_dir: './results/jwst_lens/cluster/15_03_restructuring'
+  res_dir: './results/jwst_lens/cluster/15_04_restructuring'
 
-  step_type: 'cleanflickernoisestep'  # tweakregstep, cleanflickernoisestep, cal
[…]
