+++ src/instruments/jwst/jwst_likelihoods.py
@@ -21,7 +21,7 @@ from .parse.rotation_and_shift.rotation_and_shift import (
 )
 from .parse.jwst_response import SkyMetaInformation
 from .parse.alignment.star_alignment import StarAlignmentMeta
-from .parse.data.data_loading import DataFilePaths
+from .parse.data.data_loading import DataLoadingConfig
 from .parse.jwst_psf import JwstPsfKernelConfig
 from .parse.masking.data_mask import ExtraMasks
 
@@ -75,37 +75,41 @@ def build_jwst_likelihoods(
         cfg[telescope_key].get("gaia_alignment")
     )
 
-    data_paths = DataFilePaths.from_yaml_dict(cfg[files_key])
+    load_config = DataLoadingConfig.from_yaml_dict(cfg[files_key])
 
     target_plotting = ResidualPlottingInformation()
     alignment_plotting = []
     target_filter_likelihoods = []
     stars_alignment_likelihoods = []
 
-    for filter_and_filepaths in data_paths.filters:
-        filter_alignment = FilterAlignment(filter_name=filter_and_filepaths.filter)
+    for filter, filepaths in load_config.paths.items():
+        filter_alignment = FilterAlignment(filter_name=filter)
         filter_alignment.load_star_alignment(gaia_alignment_meta_data)
         filter_alignment.load_correction_prior(
             cfg[telescope_key]["rotation_and_shift"]["correction_priors"],
-            number_of_observations=len(filter_and_filepaths.filepaths),
+            number_of_observations=len(filepaths),
         )
 
         color, target_bounds, filter_alignment = data_preloading(
-            filter_and_filepaths, grid, sky_meta, filter_alignment
+            filepaths,
+            grid.spatial.world_corners(extension_value=sky_meta.grid_extension),
+            filter_alignment,
+            loading_mode=load_config.loading_mode,
+            workers=load_config.workers,
         )
 
         target_data, stars, stars_data = data_loading(
             telescope_cfg=cfg[telescope_key],
-            filter_and_filepaths=filter_and_filepaths,
+            filepaths=filepaths,
             target_grid=grid,
-            filter_alignment=filter_alignment,
+            star_alignment=filter_alignment.star_alignment,
             target_bounds=target_bounds,
             psf_kernel_configs=psf_kernel_configs,
             extra_masks=extra_masks,
         )
 
         shift_and_rotation_correction = ShiftAndRotationCorrection(
-            domain_key=filter_and_filepaths.filter,
+            domain_key=filter,
             correction_prior=filter_alignment.correction_prior,
             rotation_center=SkyCoord(filter_alignment.boresight),
         )
@@ -124,10 +128,8 @@ def build_jwst_likelihoods(
             sky_meta=sky_meta,
             psf=np.array(target_data.psf),
             zero_flux_model=build_zero_flux_model(
-                f"{filter_and_filepaths.filter}_target",
-                zero_flux_prior_configs.get_name_setting_or_default(
-                    filter_and_filepaths.filter
-                ),
+                f"{filter}_target",
+                zero_flux_prior_configs.get_name_setting_or_default(filter),
                 shape=(len(target_data), 1, 1),
             ),
             data_mask=np.array(target_data.mask),
@@ -144,7 +146,7 @@ def build_jwst_likelihoods(
         )
 
         target_plotting.append_information(
-            filter=filter_and_filepaths.filter,
+            filter=filter,
             data=np.array(target_data.data),
             std=np.array(target_data.std),
             mask=np.array(target_data.mask),
@@ -154,10 +156,11 @@ def build_jwst_likelihoods(
 
         star_alignment: StarAlignment | None = filter_alignment.star_alignment
         if star_alignment:
[…]

[…]
