### 2025-05-19 2851438
jwst: restructure load_gaia_stars
