### 2025-05-19 d2f5dcf
jwst: data preloading rework
