+++ src/instruments/jwst/data/preloading/preloading.py
@@ -1,53 +1,69 @@
 import time
-from pathlib import Path
 from dataclasses import dataclass
-from typing import Iterable, Callable, Literal
+from typing import Iterable
 
 import numpy as np
 from astropy.coordinates import SkyCoord
+from astropy.table import Table
 from nifty8.re import logger
 
 from .....color import Color
 from ....gaia.star_finder import load_gaia_stars_in_fov
 from ...alignment.filter_alignment import FilterAlignment
+from ...alignment.star_alignment import StarTables
+from ...parse.alignment.star_alignment import StarAlignmentConfig
+from ...parse.data.data_loading import IndexAndPath, LoadingModeConfig
 from ..concurrent_loading import load_bundles
-from ..jwst_data import JwstData
-from .checks import PreloadingChecks
+from ..jwst_data import DataMetaInformation, JwstData
+from .checks import FilterConsistency
 from .data_bounds import DataBounds
-from ...parse.data.data_loading import LoadingMode
 
 
 @dataclass(slots=True)
 class PreloadBundle:
     index: int
-    filepath: str
+    filepath: IndexAndPath
     pivot_color: Color
     bounding_indices: tuple[int, int, int, int]
     boresight: SkyCoord
-    jwst_world_corners: SkyCoord
+    meta: DataMetaInformation
+    star_table: Table | None
 
 
-def load_one_preload(index: int, filepath: str, grid_corners: SkyCoord):
-    jwst_data = JwstData(filepath)
+def load_one_preload(
+    filepath: IndexAndPath,
+    grid_corners: SkyCoord,
+    star_alignment_config: StarAlignmentConfig | None,
+):
+    jwst_data = JwstData(filepath.path)
+
+    if star_alignment_config:
+        jwst_world_corners = jwst_data.wcs.world_corners()
+        star_table = load_gaia_stars_in_fov(
+            fov_corners=jwst_world_corners,
+            library_path=star_alignment_config.library_path,
+            exclude_source_ids=star_alignment_config.exclude_source_id,
+        )
+    else:
+        star_table = None
 
     return PreloadBundle(
-        index=index,
+        index=filepath.index,
         filepath=filepath,
         pivot_color=jwst_data.pivot_wavelength,
         bounding_indices=jwst_data.wcs.bounding_indices_from_world_extrema(
             grid_corners
         ),
         boresight=jwst_data.get_boresight_world_coords(),
-        jwst_world_corners=jwst_data.wcs.world_corners(),
+        star_table=star_table,
+        meta=jwst_data.meta,
     )
 
 
-def _apply_preload_side_effects(
+def _preloading_data_products(
     bundles: Iterable[PreloadBundle],
-    target_bounds: DataBounds,
-    checks: PreloadingChecks,
-    filter_alignment: FilterAlignment,
-) -> None:
+    star_alignment_config: StarAlignmentConfig,
+) -> tuple[DataMetaInformation, DataBounds, StarTables | None, list[SkyCoord]]:
     """Apply side effects from the preloaded data bundles.
 
     This function processes each preload bundle to:
@@ -60,41 +76,50 @@ def _apply_preload_side_effects(
     ----------
     bundles: Iterable[PreloadBundle]
         The preloaded data bundles to process
-    target_bounds: DataBounds
-        The target bounds object to update with cutout indices
-    checks: PreloadingChecks
-        Object containing validation checks to perform on the data
-    filter_alignment: FilterAlignment
-        The alignment data to update with:
-            - boresight information
-            - star positions (if star_alignment is enabled)
+    star_alignment_config: StarAlignmentMeta
+        The Star alignment meta information, needed for instantiating the StarTables.
 
[…]
