### 2025-05-19 e2af44c
jwst: IndexAndPath
