### 2025-05-19 5dcb4b3
jwst: star alignment prelaodign rework
