### 2025-05-19 c6e2382
jwst: DataMetaInfo holds the color
