### 2025-05-19 545aa32
jwst: rework data loading
