+++ src/instruments/jwst/data/loading/data_loading.py
@@ -1,161 +1,86 @@
-from pathlib import Path
-from typing import Any
+from dataclasses import dataclass
+from typing import Optional
 
-import numpy as np
-from astropy import units as u
 from nifty8.re import logger
 
-# Parse
-from ...parse.jwst_likelihoods import StarData, TargetData
+from .....grid import Grid
+from ...alignment.star_alignment import StarTables
+from ...parse.alignment.star_alignment import StarAlignmentConfig
+from ...parse.data.data_loading import IndexAndPath, LoadingModeConfig, Subsample
 from ...parse.jwst_psf import JwstPsfKernelConfig
 from ...parse.masking.data_mask import ExtraMasks
-
-from .....grid import Grid
-from ...alignment.star_alignment import StarAlignment
-from ...psf.jwst_kernel import load_psf_kernel
+from ..concurrent_loading import load_bundles
 from ..jwst_data import JwstData
 from ..preloading.data_bounds import DataBounds
+from .stars_loading import (
+    StarData,
+    StarsBundle,
+    load_one_stars_bundle,
+    load_one_stars_bundle_from_filepath,
+)
+from .target_loading import (
+    TargetBundle,
+    TargetData,
+    load_one_target_bundle,
+    load_one_target_bundle_from_filepath,
+)
 
 
-def target_loading(
-    target_data: TargetData,
-    observation_id: int,
-    jwst_data: JwstData,
-    target_bounds: DataBounds,
-    target_grid: Grid,
-    extra_masks: ExtraMasks,
+@dataclass
+class DataLoaderTarget:
+    grid: Grid
+    data_bounds: DataBounds
+    subsample: Subsample | int
+
+
+@dataclass
+class DataLoaderStarAlignment:
+    config: StarAlignmentConfig
+    tables: StarTables
+
+    @classmethod
+    def from_optional(
+        cls, config: StarAlignmentConfig | None, tables: StarTables
+    ) -> Optional["DataLoaderStarAlignment"]:
+        return cls(config, tables) if config else None
+
+
+def load_data(
+    filepaths: tuple[IndexAndPath],
+    target: DataLoaderTarget,
+    star_alignment: DataLoaderStarAlignment | None,
     psf_kernel_configs: JwstPsfKernelConfig,
-):
-    data, mask, std = jwst_data.bounding_data_mask_std_by_bounding_indices(
-        target_bounds.bounds[observation_id],
-        target_grid.spatial,
-        extra_masks,
-    )
-    target_data.append_observation(
-        meta=jwst_data.meta,
-        data=data,
-        mask=mask,
-        std=std,
-        subsample_centers=jwst_data.data_subpixel_centers(
-            target_bounds.bounds[observation_id],
-            subsample=target_data.subsample,
+    extra_masks: ExtraMasks,
+    loading_mode_config: LoadingModeConfig,
+) -> tuple[TargetData, StarData]:
+    target_bundles: list[TargetBundle] = load_bundles(
+        filepaths=filepaths,
+        load_one=load_one_target_bundle_from_filepath,
+        extra_kw_args=dict(
+            subsample=target.subsample,
+            target_grid=target.grid,
+            target_data_bounds=target.data_bounds,
+            psf_kernel_configs=psf_kernel_configs,
+            extra_masks=extra_masks,
         ),
-        psf=load_psf_kernel(
-            jwst_data=jwst_data,
-            subsample=target_data.subsample,
-            target_center=target_grid.spatial.center,
-            config_parameters=psf_kernel_configs,
-        ),
+        mode=loading_mode_config.loading_mode,
+        workers=loading_mode_config.workers,
     )
 
[…]
