### 2025-05-19 6aa57b1
jwst: star data rework
