+++ src/instruments/jwst/jwst_likelihoods.py
@@ -7,12 +7,20 @@ from .plotting.residuals import ResidualPlottingInformation
 from .plotting.alignment import FilterAlignmentPlottingInformation
 from .config_handler import get_grid_extension_from_config
 from .alignment.filter_alignment import FilterAlignment
-from .alignment.star_alignment import StarAlignment
+from .alignment.star_alignment import StarTables
 from .alignment.star_model import build_star_in_data
 from .zero_flux_model import build_zero_flux_model
 
-from .data.loading.data_loading import data_loading
-from .data.preloading.preloading import data_preloading
+from .likelihood.target_likelihood import build_target_likelihood_and_response
+
+
+from .parse.data.data_loading import LoadingModeConfig, Subsample
+from .data.loading.data_loading import (
+    load_data,
+    DataLoaderTarget,
+    DataLoaderStarAlignment,
+)
+from .data.preloading.preloading import preload_data
 
 # Parsing
 from .parse.zero_flux_model import ZeroFluxPriorConfigs
@@ -20,7 +28,7 @@ from .parse.rotation_and_shift.rotation_and_shift import (
     rotation_and_shift_algorithm_config_factory,
 )
 from .parse.jwst_response import SkyMetaInformation
-from .parse.alignment.star_alignment import StarAlignmentMeta
+from .parse.alignment.star_alignment import StarAlignmentConfig
 from .parse.data.data_loading import DataLoadingConfig
 from .parse.jwst_psf import JwstPsfKernelConfig
 from .parse.masking.data_mask import ExtraMasks
@@ -71,41 +79,49 @@ def build_jwst_likelihoods(
     )
     extra_masks = ExtraMasks.from_yaml_dict(cfg[telescope_key])
 
-    gaia_alignment_meta_data = StarAlignmentMeta.from_yaml_dict(
-        cfg[telescope_key].get("gaia_alignment")
+    star_alignment_config: StarAlignmentConfig | None = (
+        StarAlignmentConfig.from_yaml_dict(cfg[telescope_key].get("gaia_alignment"))
     )
 
-    load_config = DataLoadingConfig.from_yaml_dict(cfg[files_key])
+    data_loader = DataLoadingConfig.from_yaml_dict(cfg[files_key])
 
     target_plotting = ResidualPlottingInformation()
     alignment_plotting = []
     target_filter_likelihoods = []
     stars_alignment_likelihoods = []
 
-    for filter, filepaths in load_config.paths.items():
+    allstars = []
+    for filter, filepaths in data_loader.paths.items():
         filter_alignment = FilterAlignment(filter_name=filter)
-        filter_alignment.load_star_alignment(gaia_alignment_meta_data)
         filter_alignment.load_correction_prior(
             cfg[telescope_key]["rotation_and_shift"]["correction_priors"],
             number_of_observations=len(filepaths),
         )
 
-        color, target_bounds, filter_alignment = data_preloading(
-            filepaths,
-            grid.spatial.world_corners(extension_value=sky_meta.grid_extension),
-            filter_alignment,
-            loading_mode=load_config.loading_mode,
-            workers=load_config.workers,
+        filter_meta, target_bounds, star_tables = preload_data(
+            filepaths=filepaths,
+            grid_corners=grid.spatial.world_corners(
+                extension_value=sky_meta.grid_extension
+            ),
+            filter_alignment=filter_alignment,
+            star_alignment_config=star_alignment_config,
+            loading_mode_config=data_loader.loading_mode_config,
         )
 
-        target_data, stars, stars_data = data_loading(
-            telescope_cfg=cfg[telescope_key],
+        target_data, stars_data = load_data(
             filepaths=filepaths,
-            target_grid=grid,
[…]
