### 2025-05-19 34daf5f
jwst: target likelihood function
