### 2025-05-19 c0e39ed
jwst: tmp testing setup
