+++ src/instruments/jwst/data/loader/target_loader.py
@@ -10,14 +10,14 @@ from numpy.typing import NDArray
 
 from .....grid import Grid
 from ...alignment.star_alignment import StarTables
-from ...parse.data.data_loading import IndexAndPath, Subsample
+from ...parse.data.data_loader import IndexAndPath, Subsample
 from ...parse.alignment.star_alignment import StarAlignmentConfig
 from ...parse.jwst_psf import JwstPsfKernelConfig
 from ...parse.masking.data_mask import ExtraMasks
 from ...alignment.star_alignment import Star
 from ...psf.jwst_kernel import load_psf_kernel
 from ..jwst_data import DataMetaInformation, JwstData
-from ..preloading.data_bounds import DataBounds
+from ..preloader.data_bounds import DataBounds
 from .cutout import DataCutout
 
 
