### 2025-05-19 79dd810
jwst: structured preload
