+++ src/instruments/jwst/alignment/star_model.py
@@ -7,7 +7,7 @@ from jax import lax, vmap, Array
 from ..parametric_model.parametric_prior import build_parametric_prior_from_prior_config
 
 from ..data.jwst_data import DataMetaInformation
-from ..data.loading.stars_loader import StarData
+from ..data.loader.stars_loader import StarData
 from ..parse.parametric_model.parametric_prior import ProbabilityConfig
 from ..rotation_and_shift.coordinates_correction import (
     Coordinates,
