+++ src/instruments/jwst/data/loader/data_loader.py
@@ -45,23 +45,78 @@ class DataLoaderStarAlignment:
         return cls(config, tables) if config else None
 
 
+@dataclass
+class DataLoaderEssentials:
+    """Essential input data of `load_data`."""
+
+    target: DataLoaderTarget
+    psf_kernel_configs: JwstPsfKernelConfig
+
+
+@dataclass
+class DataLoaderOptionals:
+    """Optional input data of `load_data`."""
+
+    star_alignment: DataLoaderStarAlignment | None
+    extra_masks: ExtraMasks | None
+
+
 def load_data(
     filepaths: tuple[IndexAndPath],
-    target: DataLoaderTarget,
-    star_alignment: DataLoaderStarAlignment | None,
-    psf_kernel_configs: JwstPsfKernelConfig,
-    extra_masks: ExtraMasks,
+    essential: DataLoaderEssentials,
+    optional: DataLoaderOptionals,
     loading_mode_config: LoadingModeConfig,
-) -> tuple[TargetData, StarData]:
+) -> tuple[TargetData, StarData | None]:
+    """Load the data
+    1. Target cutouts for filter observations.
+    2. Psf kernels for the cutouts for filter observations.
+
+    Optional:
+    1. Star cutouts.
+    2. Psf kernels for the cutouts.
+
+    Parameters
+    ----------
+    filepaths: tuple[IndexAndPath]
+        The filepaths of JWST data files to be preloaded
+    essential: DataLoaderEssentials
+        target: DataLoaderTarget
+            - grid
+            - data_bounds
+            - subsample
+        psf_kernel_config: JwstPsfKernelConfig
+            Config parameters for psf kernel loading see `load_psf_kernel`.
+
+    optional: DataLoaderOptionals,
+        star_alignment: DataLoaderStarAlignment, optional
+            - config: StarAlignmentConfig
+            - tables: StarTables
+        extra_masks: ExtraMasks, optional
+            some extra masks, either in the target or the star data cutouts.
+    loading_mode_config: LoadingModeConfig:
+        loading_mode: LoadingMode, algorithm for loading data:
+            - "serial": Sequential processing
+            - "threads": Multi-threaded for I/O-bound operations
+            - "processes": Multi-process for CPU-bound operations
+        workers: int | None
+            Number of threads/processes to use (None = executor default)
+
+    Returns
+    -------
+    Tuple[DataMetaInformation, DataBounds, StarTables | None]
+        - DataMetaInformation: Filter information, checked for consistency.
+        - DataBounds: Aligned shapes and bounds for the target data
+        - StarTables, optional: Star tables from the gaia catalog
+    """
     target_bundles: list[TargetBundle] = load_bundles(
         filepaths=filepaths,
         load_one=load_one_target_bundle_from_filepath,
         extra_kw_args=dict(
-            subsample=target.subsample,
-            target_grid=target.grid,
-            target_data_bounds=target.data_bounds,
-            psf_kernel_configs=psf_kernel_configs,
-            extra_masks=extra_masks,
+            subsample=essential.target.subsample,
+            target_grid=essential.target.grid,
+            target_data_bounds=essential.target.data_bounds,
+            psf_kernel_configs=essential.psf_kernel_configs,
+            extra_masks=optional.extra_masks,
         ),
         mode=loading_mode_config.loading_mode,
         workers=loading_mode_config.workers,
@@ -71,16 +126,16 @@ def load_data(
         filepaths=filepaths,
         load_one=load_one_stars_bundle_from_filepath,
         extra_kw_args=dict(
-            star_tables=star_alignment.tables,
-            star_alignment_config=star_alignment.config,
-            extra_masks=extra_masks,
-            psf_kernel_configs=psf_kernel_configs,
[…]
