### 2025-05-19 c055cda
jwst: data loader restructure, to make it consistent with preloading
