### 2025-05-19 aa1d8ff
jwst: more robust plotting
