### 2025-05-19 ab924e6
jwst: stars with motion
