### 2025-05-19 1db9ab6
jwst: tmp, star plotting
