### 2025-05-19 e482426
jwst: tmp, trying the new stars
