### 2025-05-20 10c8be7
jwst: build target likelihood
