+++ src/instruments/jwst/jwst_response.py
@@ -5,31 +5,34 @@
 
 # %
 
-from typing import Callable, Union
+from dataclasses import dataclass
+from typing import Callable
 
 import nifty8.re as jft
 import numpy as np
-from astropy import units as u
 from astropy.coordinates import SkyCoord
 from numpy.typing import ArrayLike
 
-from .parse.parametric_model.parametric_prior import ProbabilityConfig
-from .parse.rotation_and_shift.rotation_and_shift import LinearConfig, NufftConfig
-from .parse.jwst_response import SkyMetaInformation
-from .data.jwst_data import DataMetaInformation
-
-from .alignment.star_model import StarInData
+from ...grid import Grid
 from ...wcs.wcs_astropy import WcsAstropy
-from .integration.unit_conversion import build_unit_conversion
+from .alignment.star_model import StarInData
+from .data.jwst_data import DataMetaInformation
+from .data.loader.target_loader import TargetData
+from .filter_projector import FilterProjector
 from .integration.integration import integration_factory
-from .psf.psf_learning import LearnablePsf
-from .psf.psf_operator import build_psf_operator_strategy, PsfDynamic, PsfStatic
+from .integration.unit_conversion import build_unit_conversion
 from .masking.build_mask import build_mask
+from .parse.jwst_response import SkyMetaInformation
+from .parse.rotation_and_shift.rotation_and_shift import LinearConfig, NufftConfig
+from .parse.zero_flux_model import ZeroFluxPriorConfigs
+from .psf.psf_learning import LearnablePsf
+from .psf.psf_operator import PsfDynamic, PsfStatic, build_psf_operator_strategy
 from .rotation_and_shift import RotationAndShift, build_rotation_and_shift
 from .rotation_and_shift.coordinates_correction import (
-    build_coordinates_corrected_for_field,
     ShiftAndRotationCorrection,
+    build_coordinates_corrected_for_field,
 )
+from .zero_flux_model import build_zero_flux_model
 
 
 class JwstResponse(jft.Model):
@@ -94,52 +97,6 @@ class JwstResponse(jft.Model):
         return out
 
 
-def build_sky_to_subsampled_data(
-    sky_domain: dict[str, jft.ShapeWithDtype],
-    data_subsampled_centers: SkyCoord | list[SkyCoord],
-    sky_wcs: WcsAstropy,
-    rotation_and_shift_algorithm: LinearConfig | NufftConfig,
-    shift_and_rotation_correction: ShiftAndRotationCorrection | None,
-) -> RotationAndShift:
-    """Build the sky to subsampled data. The sky will be rotated and shifted by
-    interpolation to the subsampled data center, whose central world coordinate is given
-    by `data_subsampled_centers`. Additionally we can correct for shift and rotation
-    errors.
-
-    Parameters
-    ----------
-    sky_domain: dict
-        The sky domain.
-    data_subsampled_centers: SkyCoord
-        The world coordinates of the subsampled data pixel centers.
-    sky_wcs: WcsAstropy
-        The wcs system of the sky (model target).
-    rotation_and_shift_algorithm: LinearConfig | NufftConfig
-        The interpolation algorithm, handling shift and rotation of data.
-    shift_and_rotation_correction: ShiftAndRotationCorrection
-        The probability distribution for the shift and rotation correction.
-        If None, we assume that the data_subsampled_centers are correct.
-
-    Return
-    ------
-    An operator that takes the output of sky model (saved in sky_domain) and projects
-    the sky onto the - potentially subsampled - data pixel.
-    """
-    coordinates = build_coordinates_corrected_for_field(
-        shift_and_rotation_correction=shift_and_rotation_correction,
[…]
