+++ src/instruments/jwst/data/preloader/preloader.py
@@ -19,6 +19,123 @@ from .checks import FilterConsistency
 from .data_bounds import DataBounds
 
 
+# --------------------------------------------------------------------------------------
+# Preload API
+# --------------------------------------------------------------------------------------
+
+
+@dataclass(slots=True, frozen=True)
+class Preloader:
+    """Prelaoding input data.
+
+    Parameters
+    ----------
+    grid_corners: list[SkyCoord]
+        The spatial corners defining the sky grid
+    star_alignment_config: StarAlignmentConfig, optional
+        Triggers a star search in the gaia catalog. See `StarAlignmentConfig` for
+        field definition.
+    """
+
+    # Essential --------------------------------------------
+    grid_corners: list[SkyCoord]
+
+    # Optional ---------------------------------------------
+    star_alignment_config: StarAlignmentConfig | None = None
+
+
+@dataclass(slots=True)
+class PreloaderSideEffects:
+    """Data to be side-effected.
+
+    Parameters
+    ----------
+    filter_alignment: FilterAlignment, updated keys:
+        - boresight, SkyCoord of the jwst observation
+    """
+
+    filter_alignment: FilterAlignment
+
+
+PreloadResult = tuple[DataMetaInformation, DataBounds, StarTables | None]
+
+
+def preload_data(
+    filepaths: tuple[IndexAndPath],
+    *,
+    preloader: Preloader,
+    side_effects: PreloaderSideEffects,
+    loading_mode_config: LoadingModeConfig,
+) -> PreloadResult:
+    """Preload JWST data and perform validation checks.
+
+    This function processes filepaths to:
+    #### 1. Perform validation checks
+    - Verify energy consistency across files
+
+    #### 2. Build target data information
+    - Load and align shapes for target data
+    - Collect corresponding cutout indices
+
+    #### 3. Gather alignment information
+    - Record boresight coordinates
+    - Optionally load stars in the field of view for alignment
+
+    Parameters
+    ----------
+    filepaths: tuple[IndexAndPath]
+        The filepaths of JWST data files to be preloaded
+    preloader: Preloader
+        Input information. See `Preloader` for field definition.
+    side_effects: PreloaderSideEffects
+        Side-affected data. See `PreloaderSideEffects` for field definition.
+    loading_mode_config: LoadingModeConfig:
+        loading_mode: LoadingMode, algorithm for loading data:
+            - "serial": Sequential processing
+            - "threads": Multi-threaded for I/O-bound operations
+            - "processes": Multi-process for CPU-bound operations
+        workers: int | None
+            Number of threads/processes to use (None = executor default)
+
+    Returns
+    -------
+    PreloadResult:
+        - DataMetaInformation: Filter information, checked for consistency.
+        - DataBounds: Aligned shapes and bounds for the target data
+        - StarTables, optional: Star tables from the gaia catalog
+    """
+
+    t = time.perf_counter()
+    logger.info("Preload JWST data")
+
+    bundles: Iterable[PreloadBundle] = load_bundles(
+        filepaths,
+        _load_one_preload_bundle,
+        mode=loading_mode_config.loading_mode,
+        workers=loading_mode_config.workers,
+        extra_kw_args=dict(
+            grid_corners=preloader.grid_corners,
+            star_alignment_config=preloader.star_alignment_config,
+        ),
+    )
+
+    filter_meta, target_bounds, star_tables, boresights = _preload_data_products(
+        bundles,
+        preloader.star_alignment_config,
+    )
+
+    _preload_side_effects(side_effects.filter_alignment, boresights)
+
+    logger.info(f"{time.perf_counter() - t}")
+
+    return filter_meta, target_bounds, star_tables
+
+
[…]
