### 2025-05-20 af86c45
jwst: with plotting and gaia stars
