### 2025-05-20 0113734
jwst: tmp; testing and plotting some offsets
