+++ src/instruments/jwst/alignment/utils.py
@@ -1,0 +1,218 @@
+import numpy as np
+import astropy.units as u
+from astropy.coordinates import SkyCoord
+
+
+# ------------------------------------------------------------------
+#  Small helpers
+# ------------------------------------------------------------------
+def _flat_sky_sep(p1: SkyCoord, p2: SkyCoord) -> u.Quantity:
+    """
+    “flat” separation: treat the sky as a flat plane that is tangent
+    at p1.  Good only for small angles.
+
+    Returned in arc-sec.
+    """
+    p1, p2 = p1.icrs, p2.icrs
+    dra = (p2.ra - p1.ra).wrap_at(180 * u.deg)  #  –180 … +180
+    ddec = p2.dec - p1.dec
+
+    dx = (dra.to(u.rad) * np.cos(p1.dec)).to(u.rad)
+    dy = ddec.to(u.rad)
+
+    return np.hypot(dx, dy).to(u.arcsec)  # √(dx²+dy²)
+
+
+def _true_sep(p1: SkyCoord, p2: SkyCoord) -> u.Quantity:
+    """Exact great-circle separation (arc-sec)."""
+    return p1.icrs.separation(p2.icrs).to(u.arcsec)
+
+
+def _true_sep2(p1: SkyCoord, p2: SkyCoord) -> u.Quantity:
+    """
+    Exact great-circle separation between two sky positions
+    (returned as an astropy Quantity in arc-seconds),
+    but without using SkyCoord.separation() internally.
+
+    Uses the numerically stable haversine formula:
+
+        Δσ = 2 · arcsin( √( hav(Δδ) + cos δ1 · cos δ2 · hav(Δα) ) )
+
+    where hav(x) = sin²(x/2).
+    """
+
+    # 1. extract long/lat in radians
+    ra1 = p1.icrs.ra.to_value(u.rad)
+    dec1 = p1.icrs.dec.to_value(u.rad)
+    ra2 = p2.icrs.ra.to_value(u.rad)
+    dec2 = p2.icrs.dec.to_value(u.rad)
+
+    # 2. haversine of longitude/latitude differences
+    d_ra = ra2 - ra1
+    d_dec = dec2 - dec1
+
+    sin2_ddec = np.sin(d_dec / 2.0) ** 2
+    sin2_dra = np.sin(d_ra / 2.0) ** 2
+
+    # 3. haversine formula
+    haversine = sin2_ddec + np.cos(dec1) * np.cos(dec2) * sin2_dra
+    ang_rad = 2.0 * np.arcsin(np.sqrt(haversine))
+
+    # 4. return as arc-seconds
+    return (ang_rad * u.rad).to(u.arcsec)
+
+
+# ------------------------------------------------------------------
+#  PUBLIC ROUTINE
+# ------------------------------------------------------------------
+def sky_offset_diagnostics(
+    pos1: SkyCoord,
+    pos2: SkyCoord,
+    dra_shift: u.Quantity = 0 * u.mas,
+    ddec_shift: u.Quantity = 0 * u.mas,
+):
+    """
+    1)  Compute the great-circle (“true”) separation σ₀ between pos1 ↔ pos2.
+    2)  Compute the flat-sky (“flat”) separation σ₀^flat .
+        ε₀  = |σ₀ − σ₀^flat|   (small-angle error)
+    3)  Optionally shift pos1 by (dra_shift, ddec_shift) and repeat.
+        Returns σ₁, σ₁^flat, ε₁ and the deltas with respect to the baseline.
+
+    All angles are `astropy.units.Quantity`; defaults mean “no shift”.
+
+    Returns
+    -------
+    dict with (always)
+        'sigma0'     : true separation        (arc-sec)
[…]
