+++ demos/configs/spt0418.yaml
@@ -316,7 +316,7 @@ telescope:
 
 
 files:
-  loading_mode: 'threads' # processes, threads, serial
+  loading_mode: 'serial' # processes, threads, serial
   # workers: 4
 
   res_dir: './results/jwst_lens/cluster/16_01_newstar'
