### 2025-05-22 90e3513
resolve: take sky key from model
