### 2025-05-22 4eaec8f
jwst: implement parsing step
