### 2025-05-22 4a831c8
jwst: restructure parsing & docstring
