+++ src/instruments/jwst/likelihood/target_likelihood.py
@@ -1,28 +1,16 @@
 from dataclasses import dataclass
 
-import numpy as np
-
-from ..data.loader.target_loader import TargetData
 from ..jwst_response import (
     JwstResponse,
-    TargetResponseEssentials,
-    TargetResponseOptionals,
+    TargetResponseInput,
     build_target_response,
 )
 from ..plotting.residuals import ResidualPlottingInformation
-from .likelihood import build_likelihood
+from .likelihood import build_likelihood, GaussianLikelihoodInput
 
 # --------------------------------------------------------------------------------------
 # Jwst Likelihood
 # --------------------------------------------------------------------------------------
-
-
-@dataclass
-class TargetGaussianLikelihoodEssentials:
-    """Essential input data of `load_data`."""
-
-    response: JwstResponse
-    data: TargetData
 
 
 @dataclass
@@ -31,26 +19,22 @@ class TargetLikelihoodSideEffects:
 
 
 def build_target_likelihood(
-    essentials: TargetResponseEssentials,
-    optionals: TargetResponseOptionals,
+    response: TargetResponseInput,
     side_effect: TargetLikelihoodSideEffects,
 ):
-    jwst_target_response = build_target_response(
-        essentials=essentials,
-        optionals=optionals,
-    )
+    jwst_target_response: JwstResponse = build_target_response(input_config=response)
 
     likelihood_target = build_likelihood(
-        TargetGaussianLikelihoodEssentials(
-            response=jwst_target_response, data=essentials.target_data
+        GaussianLikelihoodInput(
+            response=jwst_target_response, data=response.target_data
         )
     )
 
     side_effect.plotting.append_information(
-        filter=essentials.filter,
-        data=essentials.target_data.data,
-        std=essentials.target_data.std,
-        mask=essentials.target_data.mask,
+        filter=response.filter,
+        data=response.target_data.data,
+        std=response.target_data.std,
+        mask=response.target_data.mask,
         model=jwst_target_response,
     )
 
