### 2025-05-23 fa933c1
jwst: units & shifts better handling
