+++ src/instruments/jwst/rotation_and_shift/coordinates_correction.py
@@ -24,6 +24,7 @@ from ....wcs import world_coordinates_to_index_grid
 
 from ..parse.rotation_and_shift.coordinates_correction import (
     CoordinatesCorrectionPriorConfig,
+    CorrectionModel,
 )
 
 
@@ -34,24 +35,24 @@ class ShiftAndRotationCorrection(jft.Model):
         correction_prior: CoordinatesCorrectionPriorConfig,
         rotation_center: SkyCoord,
     ):
-        self.model: str = correction_prior.model  # shift, rshift
-        self.shift = build_parametric_prior_from_prior_config(
+        self.model = correction_prior.model  # shift, rshift
+        self.shift: jft.Model = build_parametric_prior_from_prior_config(
             domain_key + "_shift",
             correction_prior.shift_in(correction_prior.shift_unit),
             correction_prior.shift.mean.shape,
             as_model=True,
         )
-        self.shift_unit = correction_prior.shift_unit
+        self.shift_unit: u.Unit = correction_prior.shift_unit
 
-        self.rotation_angle = build_parametric_prior_from_prior_config(
+        self.rotation_angle: jft.Model = build_parametric_prior_from_prior_config(
             domain_key + "_rotation",
             correction_prior.rotation_in(u.rad),
             correction_prior.rotation.mean.shape,
             as_model=True,
         )
-        self.rotation_unit = u.rad
+        self.rotation_unit: u.Unit = u.rad
 
-        self.rotation_center = rotation_center
+        self.rotation_center: SkyCoord = rotation_center
 
         super().__init__(domain=self.rotation_angle.domain | self.shift.domain)
 
@@ -203,7 +204,6 @@ def build_coordinates_corrected_for_stars(
     pixel_coordinates: list[np.ndarray],
     pixel_distance: u.Quantity,
     observation_ids: tuple[int],
-    shift_only: bool = True,
 ) -> Union[
     Coordinates, CoordinatesCorrectedShiftOnly, CoordinatesCorrectedShiftAndRotation
 ]:
@@ -250,7 +250,7 @@ def build_coordinates_corrected_for_stars(
         (pixel_distance.to(shift_unit), pixel_distance.to(shift_unit))
     )
 
-    if shift_only:
+    if shift_and_rotation_correction.model == CorrectionModel.SHIFT:
         assert len(observation_ids) == pixel_coordinates.shape[0]
 
         if pixel_coordinates.shape == shift_and_rotation_correction.shift.target.shape:
@@ -326,9 +326,9 @@ def build_coordinates_corrected_for_field(
         return Coordinates(pixel_coordinates)
 
     shift_unit = shift_and_rotation_correction.shift_unit
-    pixel_distance = reconstruction_grid_wcs.distances.to(shift_unit)
+    reconstruction_scale = reconstruction_grid_wcs.distances.to(shift_unit)
 
-    if shift_and_rotation_correction.model == "shift":
+    if shift_and_rotation_correction.model == CorrectionModel.SHIFT:
         assert (  # Check that we have the same amount of observations (length of first/0th axis)
             shift_and_rotation_correction.shift.target.shape
             == pixel_coordinates.shape[:2]
@@ -337,10 +337,13 @@ def build_coordinates_corrected_for_field(
 
         shape = (Ellipsis, None, None)
         return CoordinatesCorrectedShiftOnly(
-            shift_and_rotation_correction, pixel_coordinates, pixel_distance, shape
+            shift_and_rotation_correction,
[…]
