### 2025-05-26 f34c8cb
jwst: adjusting plotting for alma + jwst
