### 2025-05-26 557bda0
jwst: restructure alignment likelihoods
