+++ src/instruments/jwst/jwst_likelihoods.py
@@ -1,20 +1,24 @@
+from dataclasses import dataclass
+
+
 from ...likelihood import build_gaussian_likelihood
 from ...grid import Grid
-from .jwst_response import (
-    build_jwst_response,
-    TargetResponseInput,
-)
+from .jwst_response import build_jwst_response, TargetResponseInput
 from .filter_projector import FilterProjector, build_filter_projector
 from .rotation_and_shift.coordinates_correction import ShiftAndRotationCorrection
 from .plotting.residuals import ResidualPlottingInformation
-from .plotting.alignment import FilterAlignmentPlottingInformation
+from .plotting.alignment import MultiFilterAlignmentPlottingInformation
 from .alignment.filter_alignment import FilterAlignment
-from .alignment.star_model import build_star_in_data
-from .zero_flux_model import build_zero_flux_model
 
 from .likelihood.target_likelihood import (
     TargetLikelihoodSideEffects,
     build_target_likelihood,
+)
+from .likelihood.alignment_likelihood import (
+    build_star_alignment_likelihood,
+    StarAlignmentResponseInput,
+    AlignmentLikelihoodSideEffects,
+    MultiFilterAlignmentLikelihoods,
 )
 from .likelihood.likelihood import build_likelihood
 
@@ -50,6 +54,35 @@ from astropy.coordinates import SkyCoord
 from typing import Union
 
 
+@dataclass
+class TargetLikelihoodProducts:
+    likelihood: jft.Likelihood
+    plotting: ResidualPlottingInformation
+    filter_projector: FilterProjector | None = None
+
+
+@dataclass
+class AlignemntLikelihoodProducts:
+    likelihood: MultiFilterAlignmentLikelihoods
+    plotting: MultiFilterAlignmentPlottingInformation
+
+    @classmethod
+    def from_optional(
+        cls,
+        likelihood: MultiFilterAlignmentLikelihoods | None,
+        plotting: MultiFilterAlignmentPlottingInformation | None,
+    ):
+        if likelihood is None:
+            return None
+        return cls(likelihood=likelihood, plotting=plotting)
+
+
+@dataclass
+class JwstLikelihoodProducts:
+    target: TargetLikelihoodProducts
+    alignment: AlignemntLikelihoodProducts | None
+
+
 def build_jwst_likelihoods(
     cfg: dict,
     grid: Grid,
@@ -57,7 +90,7 @@ def build_jwst_likelihoods(
     files_key: str = "files",
     telescope_key: str = "telescope",
     sky_unit: u.Unit | None = None,
-) -> Union[jft.Likelihood, FilterProjector, dict]:
+) -> JwstLikelihoodProducts:
     """Build the jwst likelihood_target according to the config and grid."""
 
     filter_projector = build_filter_projector(
@@ -77,9 +110,17 @@ def build_jwst_likelihoods(
     target_plotting = ResidualPlottingInformation(
         y_offset=min(filter_projector.keys_and_index.values())
     )
-    alignment_plotting = []
     target_filter_likelihoods = []
-    stars_alignment_likelihoods = []
+    alignment_plotting = (
+        MultiFilterAlignmentPlottingInformation()
+        if cfg_parser.star_alignment_config is not None
+        else None
+    )
+    alignment_likelihoods = (
+        MultiFilterAlignmentLikelihoods()
+        if cfg_parser.star_alignment_config is not None
+        else None
+    )
 
     for filter, filepaths in cfg_parser.data_loader.paths.items():
         filter_alignment = FilterAlignment(filter_name=filter)
@@ -118,6 +159,7 @@ def build_jwst_likelihoods(
             loading_mode_config=cfg_parser.data_loader.loading_mode_config,
         )
 
+        # Constructing the Likelihood
[…]
