+++ src/instruments/jwst/likelihood/alignment_likelihood.py
@@ -1,0 +1,156 @@
+from dataclasses import dataclass, field
+from functools import reduce
+
+import numpy as np
+import nifty8.re as jft
+
+from ..jwst_response import (
+    JwstResponse,
+    build_jwst_response,
+)
+from ..plotting.alignment import (
+    FilterAlignmentPlottingInformation,
+    MultiFilterAlignmentPlottingInformation,
+)
+from .likelihood import build_likelihood, GaussianLikelihoodInput
+from ..alignment.star_alignment import StarTables
+from ..parse.alignment.star_alignment import StarAlignmentConfig
+from ..alignment.star_model import build_star_in_data
+from ..data.loader.stars_loader import StarData
+from ..zero_flux_model import build_zero_flux_model
+from ..parse.zero_flux_model import ZeroFluxPriorConfigs
+from ..rotation_and_shift.coordinates_correction import ShiftAndRotationCorrection
+from ..data.jwst_data import DataMetaInformation
+from ..parse.jwst_response import SkyMetaInformation
+
+# --------------------------------------------------------------------------------------
+# Alignment Likelihood API
+# --------------------------------------------------------------------------------------
+
+
+@dataclass
+class AlignmentLikelihoodSideEffects:
+    plotting: MultiFilterAlignmentPlottingInformation
+
+
+@dataclass
+class StarAlignmentResponseInput:
+    filter_name: str
+    filter_meta: DataMetaInformation
+    sky_meta: SkyMetaInformation
+    star_tables: StarTables
+    stars_data: StarData
+    star_light_prior: StarAlignmentConfig
+    shift_and_rotation_correction: ShiftAndRotationCorrection | None
+    zero_flux_prior_configs: ZeroFluxPriorConfigs
+
+
+@dataclass
+class AlignmentLikelihoodProducts:
+    likelihood: jft.Likelihood
+    likelihood_convolved: jft.Likelihood | None
+
+
+@dataclass
+class MultiFilterAlignmentLikelihoods:
+    likelihoods: list[AlignmentLikelihoodProducts] = field(default_factory=list)
+
+    @property
+    def likelihood(self):
+        return reduce(lambda x, y: x + y, (ll.likelihood for ll in self.likelihoods))
+
+    @property
+    def likelihood_convolved(self):
+        return reduce(
+            lambda x, y: x + y, (ll.likelihood_convolved for ll in self.likelihoods)
+        )
+
+
+def build_star_alignment_likelihood(
+    response: StarAlignmentResponseInput,
+    side_effect: AlignmentLikelihoodSideEffects,
+) -> AlignmentLikelihoodProducts:
+    filter_alignment_plotting = dict(
+        psf_convolved=FilterAlignmentPlottingInformation(response.filter_name),
+        psf=FilterAlignmentPlottingInformation(response.filter_name),
+    )
+    filter_alignment_likelihoods = dict(psf_convolved=[], psf=[])
+
+    # psf_shape = np.array(stars_data[stars[0].id].psf).shape
+    # for star in stars:
+    #     psf_shape_ii = np.array(stars_data[star.id].psf).shape
+    #     assert psf_shape[1:] == psf_shape_ii[1:]
+    # psf_model = build_psf_modification_model_strategy(
+    #     f"{filter_and_filepaths.response.filter_name}", psf_shape, strategy="single"
+    # )
+
+    for star in response.star_tables.get_stars():
+        psf = response.stars_data[star.id].psf
+        # psf = build_psf_model_strategy(f"{filter_and_filepaths.response.filter_name}_{star.id}", psf_shape, strategy='full')
+        # psf = LearnablePsf(psf, psf_model)
+
+        import scipy
+
+        psf_convolved = np.array(
+            [
[…]
