### 2025-05-26 e40cf17
jwst: bugfix in the star data loading
