### 2025-05-27 a5b91b3
jwst: delete model fixpointing parametric
