### 2025-05-27 dccd3bb
jwst: better output naming
