### 2025-05-27 763cb99
jwst: Complete Star data nans will be ignored
