### 2025-05-27 87f3c19
jwst: alignment minimization process outside the script
