### 2025-05-27 b435bf3
jwst: delete alignment model setup
