### 2025-05-27 f52f20a
Bugfix
copy the dict and do not change in place. Otherwise the filepaths get
longer if the dict is reused
