+++ src/instruments/chandra/chandra_observation.py
@@ -17,11 +17,7 @@ from ...messages import message_obs, message_binning, message_exposure
 
 
 class ChandraObservationInformation():
-
-    """
-    Base class to provide an interface with the CXC analysis and simulation tools.
-
-    """
+    """Base class to provide an interface with the CXC analysis and simulation tools."""
 
     def __init__(self, obsInfo, npix_s, npix_e, fov, elim, center=None, energy_ranges=None, chips_off=()):
         """
