+++ demos/configs/spt0418.yaml
@@ -272,7 +272,7 @@ files:
   loading_mode: 'serial' # processes, threads, serial
   # workers: 4
 
-  res_dir: './results/jwst_lens/cluster/17_02_f150w'
+  res_dir: './results/jwst_lens/cluster/17_03_f150w_with_alignment'
 
   step_type: 'cleanflickernoisestep'  # tweakregstep, cleanflickernoisestep, cal
 
