### 2025-05-28 1546792
jwst: starting to make target data optional
