### 2025-05-28 8b43df3
jwst: bugfix, making fixpointing run over convolved & unconvolved psf
