### 2025-05-28 68b35d6
88-new-mf-model: add demo.
