### 2025-05-28 13e00ba
88-new-mf-model: remove unnecessary _isscalar function.
