### 2025-05-29 4185357
jwst: plotting adjust for scatter stars into alignment
