### 2025-05-29 f5c80f2
jwst: bugfix (take xmax-residuals); no star alignment in final reco
