### 2025-06-30 3d28b14
97-github-issue-8: add comparison discussion with jolideco and pylira.
