### 2026-01-12 #7768 @hamogu
@ivastar Yes, I'm still available.
