### 2026-02-04 NIFTy-PPL/J-UBIK#16 @prashjet
webbpsf --> stpsf
To run `jwst_demo.py` I needed to tinker around to handle the renaming of `webbpsf` to `stpsf`.

The page linked as `webbpsf` on [the J-UBIK docs](https://ift.pages.mpcdf.de/j-ubik/user/jwst_demo.html) tells me that the package has been renamed to `stpsf`. I've installed `stpsf` as per [their instructions](https://stpsf.readthedocs.io/en/latest/installation.html) and, following their [transition guide, ](https://stpsf.readthedocs.io/en/latest/webbpsf_to_stpsf.html#import-stpsf-as-webbpsf) I have replaced `pip install stpsf` with `import stpsf as webbpsf` where it appears in J-UBIK i.e. line 106 of `instruments/jwst/jwst_psf.py`. I'm also not sure whether the instruction to manually download and unzip `webbpsf-data` is relevant anymore, since `stpsf` automatically downloads data when installed.

To run the demo, I also had to manually create the empty directory `'../data/JWST/psf_library` which should be made more explicit in the docs.

I suggest updating the docs and scripts to be compatible with `stpsf`.
