### 2026-02-04 NIFTy-PPL/J-UBIK#17 @prashjet
Community guidelines
Some community guidelines need to be added to the docs. As described on the JOSS checklist:

> Community guidelines : Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support
