### 2026-02-04 NIFTy-PPL/J-UBIK#18 @prashjet
Add documentation describing the point source model
Having run the `jwst_demo`, I more or less understand the parameters controlling the diffuse component thanks to the nice walk through on the _Multi-Frequency Spectral Sky Model Demo_ page, but I still have lots of questions about the point source part of the model e.g.

 - why are these distributed as an inverse gamma distribution?
 - does this mean two identical inverse gamma distributions for the two spatial dimensions?
 - what parameterisation of the inverse gamma distribution is this? (I can't  find a standard reference which uses alpha and q parameters)

Similar to how _Multi-Frequency Spectral Sky Model Demo_ describes the diffuse component, I suggest adding a page to the docs which describes the generative model for the point sources.
