### 2026-02-04 NIFTy-PPL/J-UBIK#17 @prashjet
Yeah I'd say drop that in as a page in the docs 👍🏽
