### 2026-02-05 NIFTy-PPL/J-UBIK#20 @prashjet
Restructure docs/demos
I think the content of the demos do a comprehensive job of demonstrating the constituent parts of J-UBIK however it's currently not so clear how everything strings together. I think some restructuring and additional explanation on the _What is J-UBIK?_  could be very helpful. In addition to the existing content, I would suggest adding something along the following:

---

# What is J-UBIK?
...
## Prior models
...

J-UBIK SkyModel consist of diffuse components and point source components. The prior models for these are demonstrated here:

- diffuse component [i.e. Multi-Frequency Spectral Sky Model Demo]
- point source component [as discussed in #18]

A summary of the options configuration for SkyModel is given here:

- Configuration options for SkyModel
   - move the list of options from JWST and eRosita pipeline demos here, so there's no need to then repeat them later... just refer back to this page
   - if possible, list all possible options for `diffuse: fluctuations: non_parametric_kind` (currently only `power` example is given)
   - keep naming consistent with the description of the prior models in the prior model pages. This is generally clear already, except for `offset`... is this related to `zero_mode_settings`...?

## Inference with J-UBIK
...

The following demo shows how we load instrumental and observational data for Chandra:

- Chandra demo

Next we demonstrate all the parts of J-UBIK together: we construct the instrument response function function, define the likelihood and run inference to recover posterior samples modelling a Chandra observation of supernova remnant G299.2-2.9:

- Imaging of SNR G299.2-2.9 with Chandra

All of these steps are handled automatically when J-UBIK as an inference pipeline. Below we show example pipelines for Chandra, eRosita and JWST.

## Inference pipelines
...

Example pipelines for Chandra, eRosita and JWST. These demos explain how to set up the configuration

- Pipeline Demo: Chandra
- Pipeline Demo: JWST
- Pipeline Demo: eRosita
   - make the data for `LMC_SN1987A` available if possible (and link to the relevant paper), otherwise mention that to run this demo you'll need some processed eRosita data which isn't directly available

---

That's just an example. If any of the specific text I've written isn't quite right, please adjust according - the main aim is to restructure things to make it clearer how it all fits together. In the same vein, I think it would be helpful if the navigation bar on the left sub-divides the demos in the same way that they are structured on the _What is J-UBIK?_ page.
