### 2026-02-05 #7768 @prashjet
My review is done. The documentation is much improved! A few more tweaks and I think it will be a great resource for anyone starting to use `J-UBIK`. Here are my remaining issues before I check off the remaining two points from my checklist:

Functionality documentation
- [Add documentation describing the point source model](https://github.com/NIFTy-PPL/J-UBIK/issues/18)
- [webbpsf --> stpsf](https://github.com/NIFTy-PPL/J-UBIK/issues/16)
- [Restructure docs/demos](https://github.com/NIFTy-PPL/J-UBIK/issues/20)
- [Documentation](https://github.com/NIFTy-PPL/J-UBIK/issues/14) (small fixes)

Community guidelines
- [Community guidelines](https://github.com/NIFTy-PPL/J-UBIK/issues/17)
