### 2026-02-05 NIFTy-PPL/J-UBIK#14 @prashjet
Ah - it's the Sky Model section of https://ift.pages.mpcdf.de/j-ubik/user/chandra_demo.html
