### 2026-02-07 NIFTy-PPL/J-UBIK#21 @hamogu
Improve test coverage
Tests do exist, but in most of them, it just tests that the return of a function ha the right type or shape, not that it actually does anything useful, e.g.
- https://github.com/NIFTy-PPL/J-UBIK/blob/main/test/instruments/chandra/test_chandra_observation.py would pass if the reader returns all Nans, or
- https://github.com/NIFTy-PPL/J-UBIK/blob/main/test/test_diagnostics.py is enough that the output is not Nan. Could be -123 or 1e38 and the test would still pass.

Writing good tests is work and I understand that a software like this does not aim for complete coverage of every situation like e.g. numpy does. However, tests could be strengthend a lot with comparatively effort. To give just one example, in  `test_calculate_nwr` the test is run with a fixed seed, so one could for example
- check that all elements in the `res` array are positive,
- that the average is between 0.4 and 1.3 (or whatever number makes sense - I don't actually know).
- Maybe one could even check a few specific elements exactly against a the value they have today (`res[0, :3] = pytest.approx([0.66138955, 0.82393329, 0.45938819])`. While not a guarantee that the number in the current version are correct, at least the tests would tell you if they accidentally change.

The same thinking applies to most of the other tests. Tests coverage isn't just about calling most functions once, it's also about making sure that the return values are correct and useful.
