### 2026-02-07 #7768 @hamogu
I've updated my review and I agree that it's evolving int the right direction. In particular, the installation now works as advertised.
- I'm still can't find a comparison to the state of the field in the paper. Or maybe I'm looking at an old version of the paper? If you've added a few sentences about that please let me know where the newest version lives.
- Tests do exist, and thus one can argue that they formally fulfill the JOSS criterion of "tests exists". However, the tests are really thin: https://github.com/NIFTy-PPL/J-UBIK/issues/21. JOSS asks for "Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.". So, I guess one can argue that "it's covered" because core functions are at least called in the tests and don't crash when called. However, it would be much better if the tests actually tested that the result is correct or at least reasonable.
- Is there documentation on the YAML files? I don't want to open an issue because maybe I'm just too stupid to find it, but I'm struggling to write YAML input files for anything that's not very, very close to what's in the examples.
