### 2026-02-10 #7768 @Eberle-Vincent
@hamogu Thanks for your review. We are working on your requests.
Considering the first point:
> I'm still can't find a comparison to the state of the field in the paper. Or maybe I'm looking at an old version of the paper? If you've added a few sentences about that please let me know where the newest version lives.

We inserted a comparison to the state of the field [here](https://github.com/NIFTy-PPL/J-UBIK/commit/3d28b14ff18624f99dd558a42dd3a281bfb2790e) which should be in the newest version of the [paper](https://github.com/NIFTy-PPL/J-UBIK/blob/801f5a8d9900e482917c12583e25254a1f7206e3/paper/paper.md?plain=1#L115-L120)
