### 2026-02-15 #7768 @Eberle-Vincent
@hamogu:
> Is there documentation on the YAML files? I don't want to open an issue because maybe I'm just too stupid to find it, but I'm struggling to write YAML input files for anything that's not very, very close to what's in the examples.

Thank you for that remark! The documentation of the yaml config files was supposed to be in the pipeline demos, since the keywords that need to be in the YAML files depend on the models/instruments which are used and in the inference scripts.

So the keyworkds were explained in the eROSITA / JWST pipeline demo, but indeed they were  missing in the Chandra pipeline demo, but I added this section now. Thanks for that catch!

In general, the pipeline demos are for demonstration purposes only and we do highly recommend new users to start building scripts part by part, such that one understands the meaning of all the parameters. (Prior distributions of all the processes involved etc.) Therefore we also added two demos for correlated / diffuse emission:

https://ift.pages.mpcdf.de/j-ubik/user/spectral_sky_demo.html

and point sources:

https://ift.pages.mpcdf.de/j-ubik/user/point_source_sky_demo.html

Also:
The YAML config files are basically just dictionaries which are unpacked and plugged into the functions/classes.
The usage of these basically not necessary but a "nice to have feature" to run the same script with different parameters on e.g. a cluster. Basically you can also run all of the scripts without any YAML file.
