### 2026-02-23 #7768 @ivastar
@prashjet looks like most (all?) or the issues you opened are closed or about to be. Are you satisfies with the changes or do you have further comments?
