### 2026-02-23 #7768 @ivastar
@hamogu seems like the test are still an open issue NIFTy-PPL/J-UBIK#21

What would be a minimum viable solution here? Can the authors set up a GitHub action that runs the tests and shows coverage?
