### 2026-02-23 #7768 @prashjet
> @prashjet looks like most (all?) or the issues you opened are closed or about to be. Are you satisfies with the changes or do you have further comments?

There's one outstanding issue from me (https://github.com/NIFTy-PPL/J-UBIK/issues/20). I'd like to see the result of that before ticking off the final box in my checklist.
