### 2026-02-23 #7768 @hamogu
I trust the authors that the tests are run somewhere, I don't need to see that on GitHub, running locally on their own instance is fine, too. My issue is more that the tests don't cover that much and most of the tests that exist don't check that the return values make sense, only e.g. that they are the right datatype.
Of course, the question when test coverage is "sufficient" has no general answer, but I suggest to add a few "assert" statement that check numerical return values for the existing tests as a minimal step for me to recommend acceptance.
