### 2026-03-03 NIFTy-PPL/J-UBIK#21 @hamogu
Test coverage is always a bit a grey area where it's hard to say when exactly it's "good enough". That's a decision that you as a developer should take, not me.
I think the steps that you just took are definitely going in the direction I was looking for, so I'll close this issue (and tick the box on the review form), but of course I'm not going to stop you from adding more tests ;-)
