### 2026-03-03 #7768 @hamogu
All my concerns are resolved and I now recommend to accept this submission. I updated the form above to tick all the boxes. Thanks to @Eberle-Vincent for writing and submitting this package. I'm looking forward to using it in my own research in the future if I do a project where it's useful!
