### 2026-03-03 #7768 @Eberle-Vincent
Thanks @hamogu! If you want to use it in the future, feel free to get in touch.
Credits also got to @wmarg, @matteani and the other co-authors.
