### 2026-03-03 NIFTy-PPL/J-UBIK#20 @Eberle-Vincent
Hi @prashjet,
I took some inspiration from your comments and refined the "What is J-UBIK" section and its structure.

https://ift.pages.mpcdf.de/j-ubik/user/jubik.html

I added some text, which makes things clearer, also I restructured and subdivided pages such that the navigation bar is more structured too.

We decided against putting the config explanation to the prior section, since configs are only necessary if you want to use the config reader. If one uses the python-interfaces it is possible to do everything without these yaml files, which I would recommend at the beginning (as written in the pipeline text now)

Let me know if you like it!
