### 2026-03-03 NIFTy-PPL/J-UBIK#20 @prashjet
Thanks @Eberle-Vincent, I think that is a much more helpful structure! Happy for you to close this issue.
