### 2026-03-03 #7768 @matteani
> My last issue is also now closed and I've ticked off the last point from the checklist. Happy to conclude this review here.
>
> Thanks to all of the authors for their contributions! I may also be back to give J-UBIK a try for some future projects.

Thank you very much and if you happen to use j-ubik in the future feel free to get in touch! :)
