### 2026-04-13 #7768 @matteani
> [@matteani](https://github.com/matteani) apologies for the wait! I have a handful of minor editorial changes below. Please implement these and then we can move to publication.
>
> Line 13: imaging --> do you mean recovery? or characterization? Line 38-39: the citations should be in round braces, rather than square ones. Selig and Steininger are not in the references. Line 51: "Here, J-UBIK comes into play which addresses this challenge... from two angles" --> "J-UBIK addresses the challenge from two directions." Line 55: maybe a new line on at "Second, " and then the new paragraph is about the instuments. Line 56: "several" --> specific? Line 57: "eROSITA pointings" - remove "pointings" Line 57-58: "with plans to expand the list as the user base grows" --> perhaps a better statement would be to invite users to contribute new instruments? Line 59-60: "Ultimately,..." --> "Ultimately, through Bayesian statistics, J-UBIK..." Line 60: measures --> measurements Line 61: "uncertainty of the signal for a several data sets" -- > maybe just "signal uncertainty". Line 64: delete "pointings" Line 64-65: "In the future..." There are several repeats of this statement, both above and later in the manuscript. Please edit the repeats and decide where this statement should be, also Line 107-108. Lines 95-97: "In the spectral dimension..." --> "In the spectral dimension, the model can take several forms: a power law, a Wiener process describing the correlation structure of the logarithmic flux along the spectral axis, or a combination of both." Figures 1 & 2 are very difficult to read - the range on the legend is much larger than the range in the figure, the details are not visible. Please correct.
>
> Please include in the acknowledgements citations to the major packages that J-UBIK depends on.

@ivastar no worries and thanks for the careful read and helpful suggestions! We've now addressed all of the points you raised in the latest revision. This includes the wording/formatting fixes, removal of repeated statements, improved readability of Figures 1 and 2, and added package citations in the acknowledgements. All changes are in the last seven commits on the repository (from d7d15). Cheers :)
