### 2026-04-14 #7768 @ivastar
OK! Final stretch! Please go through the list above. Let me know what the DOI and version are of the archive. Please keep in mind that the Zenodo archive does not by default match the license, version, title and author list of the paper - please do that by hand.
