### 2026-04-21 #7768 @matteani
@ivastar amazing, thanks.
We did all the checks and uploaded to Zenodo:

Version number is v0.3.
Zenodo DOI: 10.5281/zenodo.19678136
DOI url: https://doi.org/10.5281/zenodo.19678136

Let us know if there is anything else we should do :)
