### 2026-04-22 #7768 @ivastar
@matteani the Zenodo license is CC0 while the actual paper is BSD-2. Can you please make the change and then we'll be all set?
