### 2026-04-23 #7768 @Eberle-Vincent
> [@matteani](https://github.com/matteani) the Zenodo license is CC0 while the actual paper is BSD-2. Can you please make the change and then we'll be all set?

Thanks for spotting. We changed the license on Zenodo accordingly. Let us know if something else is missing.
