### 2026-04-24 #7768 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
