### 2026-04-27 #7768 @matteani
proof looks good from our side 👍🏼

> 👋 @openjournals/aass-eics, this paper is ready to be accepted and published.
>
> Check final proof 👉📄 [Download article](https://raw.githubusercontent.com/openjournals/joss-papers/joss.07768/joss.07768/10.21105.joss.07768.pdf)
>
> If the paper PDF and the deposit XML files look good in [openjournals/joss-papers#8050](https://github.com/openjournals/joss-papers/pull/8050), then the AEiC can move forward with accepting the submission by compiling again with the command `@editorialbot accept`
