### 2026-04-29 #7768 @warrickball
Congratulations @wmarg,  @Eberle-Vincent & @matteani, your paper has been published in JOSS!
