### 2026-04-30 #7768 @matteani
> Congratulations [@wmarg](https://github.com/wmarg), [@Eberle-Vincent](https://github.com/Eberle-Vincent) & [@matteani](https://github.com/matteani), your paper has been published in JOSS!
>
> Thanks once again to [@prashjet](https://github.com/prashjet) & [@hamogu](https://github.com/hamogu) for reviewing and [@ivastar](https://github.com/ivastar) for editing this submission! JOSS simply wouldn't be possible without its community of volunteers.

Thank you all very much! @warrickball @ivastar @prashjet @hamogu  :)
